[Platform] Code refactor
