--- docs/scripts/logo.js
@@ -1,2 +1,0 @@
-(()=>{var a=!1,t=0,e=0;function s(){let n=document.querySelector("#nav-bar__title");n.addEventListener("mouseenter",m),n.addEventListener("mouseleave",c),document.querySelectorAll(".nav-bar__tab-bar--links").forEach(o=>{o.addEventListener("click",()=>{!a&&t==0?(e=3.3,t=200,requestAnimationFrame(l)):(e=3.3,t=200)})})}document.readyState!=="loading"?s():document.addEventListener("DOMContentLoaded",s);function m(){!a&&t==0?(a=!0,requestAnimationFrame(l)):(a=!0,t=200)}function c(){a=!1}function l(){let n=document.querySelector("#nav-bar__title__logo"),i=getComputedStyle(n,null).getPropertyValue("transform").split("(")[1].split(")")[0].split(","),u=i[0],f=i[1],r=Math.atan2(f,u)*(180/Math.PI);a?e<1.8&&(e=e+.06):e>0&&(e=e-.03),e<0&&(e=0),r<0&&(r=r+360);let d=e+r;n.style.transform="rotate("+d+"deg)",a?(t=200,requestAnimationFrame(l)):(t--,e<=0&&(t=0),t>0&&requestAnimationFrame(l))}})();
-//# sourceMappingURL=logo.js.map
