+++ docs/scripts/main.js
@@ -1,0 +1,59 @@
+var H;function xe(e){return H=H||document.createElement("textarea"),H.innerHTML="&"+e+";",H.value}var Ne=Object.prototype.hasOwnProperty;function Oe(e,n){return e?Ne.call(e,n):!1}function ke(e){var n=[].slice.call(arguments,1);return n.forEach(function(r){if(r){if(typeof r!="object")throw new TypeError(r+"must be object");Object.keys(r).forEach(function(t){e[t]=r[t]})}}),e}var De=/\\([\\!"#$%&'()*+,.\/:;<=>?@[\]^_`{|}~-])/g;function B(e){return e.indexOf("\\")<0?e:e.replace(De,"$1")}function ye(e){return!(e>=55296&&e<=57343||e>=64976&&e<=65007||(e&65535)===65535||(e&65535)===65534||e>=0&&e<=8||e===11||e>=14&&e<=31||e>=127&&e<=159||e>1114111)}function re(e){if(e>65535){e-=65536;var n=55296+(e>>10),r=56320+(e&1023);return String.fromCharCode(n,r)}return String.fromCharCode(e)}var Re=/&([a-z#][a-z0-9]{1,31});/gi,He=/^#((?:x[a-f0-9]{1,8}|[0-9]{1,8}))/i;function Fe(e,n){var r=0,t=xe(n);return n!==t?t:n.charCodeAt(0)===35&&He.test(n)&&(r=n[1].toLowerCase()==="x"?parseInt(n.slice(2),16):parseInt(n.slice(1),10),ye(r))?re(r):e}function T(e){return e.indexOf("&")<0?e:e.replace(Re,Fe)}var Pe=/[&<>"]/,je=/[&<>"]/g,ze={"&":"&amp;","<":"&lt;",">":"&gt;",'"':"&quot;"};function $e(e){return ze[e]}function x(e){return Pe.test(e)?e.replace(je,$e):e}var m={};m.blockquote_open=function(){return`<blockquote>
+`};m.blockquote_close=function(e,n){return"</blockquote>"+S(e,n)};m.code=function(e,n){return e[n].block?"<pre><code>"+x(e[n].content)+"</code></pre>"+S(e,n):"<code>"+x(e[n].content)+"</code>"};m.fence=function(e,n,r,t,i){var o=e[n],l="",c=r.langPrefix,s="",a,u,p;if(o.params){if(a=o.params.split(/\s+/g),u=a.join(" "),Oe(i.rules.fence_custom,a[0]))return i.rules.fence_custom[a[0]](e,n,r,t,i);s=x(T(B(u))),l=' class="'+c+s+'"'}return r.highlight?p=r.highlight.apply(r.highlight,[o.content].concat(a))||x(o.content):p=x(o.content),"<pre><code"+l+">"+p+"</code></pre>"+S(e,n)};m.fence_custom={};m.heading_open=function(e,n){return"<h"+e[n].hLevel+">"};m.heading_close=function(e,n){return"</h"+e[n].hLevel+`>
+`};m.hr=function(e,n,r){return(r.xhtmlOut?"<hr />":"<hr>")+S(e,n)};m.bullet_list_open=function(){return`<ul>
+`};m.bullet_list_close=function(e,n){return"</ul>"+S(e,n)};m.list_item_open=function(){return"<li>"};m.list_item_close=function(){return`</li>
+`};m.ordered_list_open=function(e,n){var r=e[n],t=r.order>1?' start="'+r.order+'"':"";return"<ol"+t+`>
+`};m.ordered_list_close=function(e,n){return"</ol>"+S(e,n)};m.paragraph_open=function(e,n){return e[n].tight?"":"<p>"};m.paragraph_close=function(e,n){var r=!(e[n].tight&&n&&e[n-1].type==="inline"&&!e[n-1].content);return(e[n].tight?"":"</p>")+(r?S(e,n):"")};m.link_open=function(e,n,r){var t=e[n].title?' title="'+x(T(e[n].title))+'"':"",i=r.linkTarget?' target="'+r.linkTarget+'"':"";return'<a href="'+x(e[n].href)+'"'+t+i+">"};m.link_close=function(){return"</a>"};m.image=function(e,n,r){var t=' src="'+x(e[n].src)+'"',i=e[n].title?' title="'+x(T(e[n].title))+'"':"",o=' alt="'+(e[n].alt?x(T(B(e[n].alt))):"")+'"',l=r.xhtmlOut?" /":"";return"<img"+t+o+i+l+">"};m.table_open=function(){return`<table>
+`};m.table_close=function(){return`</table>
+`};m.thead_open=function(){return`<thead>
+`};m.thead_close=function(){return`</thead>
+`};m.tbody_open=function(){return`<tbody>
+`};m.tbody_close=function(){return`</tbody>
+`};m.tr_open=function(){return"<tr>"};m.tr_close=function(){return`</tr>
+`};m.th_open=function(e,n){var r=e[n];return"<th"+(r.align?' style="text-align:'+r.align+'"':"")+">"};m.th_close=function(){return"</th>"};m.td_open=function(e,n){var r=e[n];return"<td"+(r.align?' style="text-align:'+r.align+'"':"")+">"};m.td_close=function(){return"</td>"};m.strong_open=function(){return"<strong>"};m.strong_close=function(){return"</strong>"};m.em_open=function(){return"<em>"};m.em_close=function(){return"</em>"};m.del_open=function(){return"<del>"};m.del_close=function(){return"</del>"};m.ins_open=function(){return"<ins>"};m.ins_close=function(){return"</ins>"};m.mark_open=function(){return"<mark>"};m.mark_close=function(){return"</mark>"};m.sub=function(e,n){return"<sub>"+x(e[n].content)+"</sub>"};m.sup=function(e,n){return"<sup>"+x(e[n].content)+"</sup>"};m.hardbreak=function(e,n,r){return r.xhtmlOut?`<br />
+`:`<br>
+`};m.softbreak=function(e,n,r){return r.breaks?r.xhtmlOut?`<br />
+`:`<br>
+`:`
+`};m.text=function(e,n){return x(e[n].content)};m.htmlblock=function(e,n){return e[n].content};m.htmltag=function(e,n){return e[n].content};m.abbr_open=function(e,n){return'<abbr title="'+x(T(e[n].title))+'">'};m.abbr_close=function(){return"</abbr>"};m.footnote_ref=function(e,n){var r=Number(e[n].id+1).toString(),t="fnref"+r;return e[n].subId>0&&(t+=":"+e[n].subId),'<sup class="footnote-ref"><a href="#fn'+r+'" id="'+t+'">['+r+"]</a></sup>"};m.footnote_block_open=function(e,n,r){var t=r.xhtmlOut?`<hr class="footnotes-sep" />
+`:`<hr class="footnotes-sep">
+`;return t+`<section class="footnotes">
+<ol class="footnotes-list">
+`};m.footnote_block_close=function(){return`</ol>
+</section>
+`};m.footnote_open=function(e,n){var r=Number(e[n].id+1).toString();return'<li id="fn'+r+'"  class="footnote-item">'};m.footnote_close=function(){return`</li>
+`};m.footnote_anchor=function(e,n){var r=Number(e[n].id+1).toString(),t="fnref"+r;return e[n].subId>0&&(t+=":"+e[n].subId),' <a href="#'+t+'" class="footnote-backref">\u21A9</a>'};m.dl_open=function(){return`<dl>
+`};m.dt_open=function(){return"<dt>"};m.dd_open=function(){return"<dd>"};m.dl_close=function(){return`</dl>
+`};m.dt_close=function(){return`</dt>
+`};m.dd_close=function(){return`</dd>
+`};function Ce(e,n){return++n>=e.length-2?n:e[n].type==="paragraph_open"&&e[n].tight&&e[n+1].type==="inline"&&e[n+1].content.length===0&&e[n+2].type==="paragraph_close"&&e[n+2].tight?Ce(e,n+2):n}var S=m.getBreak=function(n,r){return r=Ce(n,r),r<n.length&&n[r].type==="list_item_close"?"":`
+`};function ie(){this.rules=ke({},m),this.getBreak=m.getBreak}ie.prototype.renderInline=function(e,n,r){for(var t=this.rules,i=e.length,o=0,l="";i--;)l+=t[e[o].type](e,o++,n,r,this);return l};ie.prototype.render=function(e,n,r){for(var t=this.rules,i=e.length,o=-1,l="";++o<i;)e[o].type==="inline"?l+=this.renderInline(e[o].children,n,r):l+=t[e[o].type](e,o,n,r,this);return l};function k(){this.__rules__=[],this.__cache__=null}k.prototype.__find__=function(e){for(var n=this.__rules__.length,r=-1;n--;)if(this.__rules__[++r].name===e)return r;return-1};k.prototype.__compile__=function(){var e=this,n=[""];e.__rules__.forEach(function(r){r.enabled&&r.alt.forEach(function(t){n.indexOf(t)<0&&n.push(t)})}),e.__cache__={},n.forEach(function(r){e.__cache__[r]=[],e.__rules__.forEach(function(t){t.enabled&&(r&&t.alt.indexOf(r)<0||e.__cache__[r].push(t.fn))})})};k.prototype.at=function(e,n,r){var t=this.__find__(e),i=r||{};if(t===-1)throw new Error("Parser rule not found: "+e);this.__rules__[t].fn=n,this.__rules__[t].alt=i.alt||[],this.__cache__=null};k.prototype.before=function(e,n,r,t){var i=this.__find__(e),o=t||{};if(i===-1)throw new Error("Parser rule not found: "+e);this.__rules__.splice(i,0,{name:n,enabled:!0,fn:r,alt:o.alt||[]}),this.__cache__=null};k.prototype.after=function(e,n,r,t){var i=this.__find__(e),o=t||{};if(i===-1)throw new Error("Parser rule not found: "+e);this.__rules__.splice(i+1,0,{name:n,enabled:!0,fn:r,alt:o.alt||[]}),this.__cache__=null};k.prototype.push=function(e,n,r){var t=r||{};this.__rules__.push({name:e,enabled:!0,fn:n,alt:t.alt||[]}),this.__cache__=null};k.prototype.enable=function(e,n){e=Array.isArray(e)?e:[e],n&&this.__rules__.forEach(function(r){r.enabled=!1}),e.forEach(function(r){var t=this.__find__(r);if(t<0)throw new Error("Rules manager: invalid rule name "+r);this.__rules__[t].enabled=!0},this),this.__cache__=null};k.prototype.disable=function(e){e=Array.isArray(e)?e:[e],e.forEach(function(n){var r=this.__find__(n);if(r<0)throw new Error("Rules manager: invalid rule name "+n);this.__rules__[r].enabled=!1},this),this.__cache__=null};k.prototype.getRules=function(e){return this.__cache__===null&&this.__compile__(),this.__cache__[e]||[]};function Ve(e){e.inlineMode?e.tokens.push({type:"inline",content:e.src.replace(/\n/g," ").trim(),level:0,lines:[0,1],children:[]}):e.block.parse(e.src,e.options,e.env,e.tokens)}function M(e,n,r,t,i){this.src=e,this.env=t,this.options=r,this.parser=n,this.tokens=i,this.pos=0,this.posMax=this.src.length,this.level=0,this.pending="",this.pendingLevel=0,this.cache=[],this.isInLabel=!1,this.linkLevel=0,this.linkContent="",this.labelUnmatchedScopes=0}M.prototype.pushPending=function(){this.tokens.push({type:"text",content:this.pending,level:this.pendingLevel}),this.pending=""};M.prototype.push=function(e){this.pending&&this.pushPending(),this.tokens.push(e),this.pendingLevel=this.level};M.prototype.cacheSet=function(e,n){for(var r=this.cache.length;r<=e;r++)this.cache.push(0);this.cache[e]=n};M.prototype.cacheGet=function(e){return e<this.cache.length?this.cache[e]:0};function N(e,n){var r,t,i,o=-1,l=e.posMax,c=e.pos,s=e.isInLabel;if(e.isInLabel)return-1;if(e.labelUnmatchedScopes)return e.labelUnmatchedScopes--,-1;for(e.pos=n+1,e.isInLabel=!0,r=1;e.pos<l;){if(i=e.src.charCodeAt(e.pos),i===91)r++;else if(i===93&&(r--,r===0)){t=!0;break}e.parser.skipToken(e)}return t?(o=e.pos,e.labelUnmatchedScopes=0):e.labelUnmatchedScopes=r-1,e.pos=c,e.isInLabel=s,o}function Ue(e,n,r,t){var i,o,l,c,s,a;if(e.charCodeAt(0)!==42||e.charCodeAt(1)!==91||e.indexOf("]:")===-1||(i=new M(e,n,r,t,[]),o=N(i,1),o<0||e.charCodeAt(o+1)!==58))return-1;for(c=i.posMax,l=o+2;l<c&&i.src.charCodeAt(l)!==10;l++);return s=e.slice(2,o),a=e.slice(o+2,l).trim(),a.length===0?-1:(t.abbreviations||(t.abbreviations={}),typeof t.abbreviations[":"+s]>"u"&&(t.abbreviations[":"+s]=a),l)}function Ye(e){var n=e.tokens,r,t,i,o;if(!e.inlineMode){for(r=1,t=n.length-1;r<t;r++)if(n[r-1].type==="paragraph_open"&&n[r].type==="inline"&&n[r+1].type==="paragraph_close"){for(i=n[r].content;i.length&&(o=Ue(i,e.inline,e.options,e.env),!(o<0));)i=i.slice(o).trim();n[r].content=i,i.length||(n[r-1].tight=!0,n[r+1].tight=!0)}}}function te(e){var n=T(e);try{n=decodeURI(n)}catch{}return encodeURI(n)}function Ee(e,n){var r,t,i,o=n,l=e.posMax;if(e.src.charCodeAt(n)===60){for(n++;n<l;){if(r=e.src.charCodeAt(n),r===10)return!1;if(r===62)return i=te(B(e.src.slice(o+1,n))),e.parser.validateLink(i)?(e.pos=n+1,e.linkContent=i,!0):!1;if(r===92&&n+1<l){n+=2;continue}n++}return!1}for(t=0;n<l&&(r=e.src.charCodeAt(n),!(r===32||r<32||r===127));){if(r===92&&n+1<l){n+=2;continue}if(r===40&&(t++,t>1)||r===41&&(t--,t<0))break;n++}return o===n||(i=B(e.src.slice(o,n)),!e.parser.validateLink(i))?!1:(e.linkContent=i,e.pos=n,!0)}function Ae(e,n){var r,t=n,i=e.posMax,o=e.src.charCodeAt(n);if(o!==34&&o!==39&&o!==40)return!1;for(n++,o===40&&(o=41);n<i;){if(r=e.src.charCodeAt(n),r===o)return e.pos=n+1,e.linkContent=B(e.src.slice(t+1,n)),!0;if(r===92&&n+1<i){n+=2;continue}n++}return!1}function Te(e){return e.trim().replace(/\s+/g," ").toUpperCase()}function Ze(e,n,r,t){var i,o,l,c,s,a,u,p,f;if(e.charCodeAt(0)!==91||e.indexOf("]:")===-1||(i=new M(e,n,r,t,[]),o=N(i,0),o<0||e.charCodeAt(o+1)!==58))return-1;for(c=i.posMax,l=o+2;l<c&&(s=i.src.charCodeAt(l),!(s!==32&&s!==10));l++);if(!Ee(i,l))return-1;for(u=i.linkContent,l=i.pos,a=l,l=l+1;l<c&&(s=i.src.charCodeAt(l),!(s!==32&&s!==10));l++);for(l<c&&a!==l&&Ae(i,l)?(p=i.linkContent,l=i.pos):(p="",l=a);l<c&&i.src.charCodeAt(l)===32;)l++;return l<c&&i.src.charCodeAt(l)!==10?-1:(f=Te(e.slice(1,o)),typeof t.references[f]>"u"&&(t.references[f]={title:p,href:u}),l)}function Xe(e){var n=e.tokens,r,t,i,o;if(e.env.references=e.env.references||{},!e.inlineMode){for(r=1,t=n.length-1;r<t;r++)if(n[r].type==="inline"&&n[r-1].type==="paragraph_open"&&n[r+1].type==="paragraph_close"){for(i=n[r].content;i.length&&(o=Ze(i,e.inline,e.options,e.env),!(o<0));)i=i.slice(o).trim();n[r].content=i,i.length||(n[r-1].tight=!0,n[r+1].tight=!0)}}}function Ge(e){var n=e.tokens,r,t,i;for(t=0,i=n.length;t<i;t++)r=n[t],r.type==="inline"&&e.inline.parse(r.content,e.options,e.env,r.children)}function We(e){var n,r,t,i,o,l,c,s,a,u=0,p=!1,f={};if(e.env.footnotes&&(e.tokens=e.tokens.filter(function(h){return h.type==="footnote_reference_open"?(p=!0,s=[],a=h.label,!1):h.type==="footnote_reference_close"?(p=!1,f[":"+a]=s,!1):(p&&s.push(h),!p)}),!!e.env.footnotes.list)){for(l=e.env.footnotes.list,e.tokens.push({type:"footnote_block_open",level:u++}),n=0,r=l.length;n<r;n++){for(e.tokens.push({type:"footnote_open",id:n,level:u++}),l[n].tokens?(c=[],c.push({type:"paragraph_open",tight:!1,level:u++}),c.push({type:"inline",content:"",level:u,children:l[n].tokens}),c.push({type:"paragraph_close",tight:!1,level:--u})):l[n].label&&(c=f[":"+l[n].label]),e.tokens=e.tokens.concat(c),e.tokens[e.tokens.length-1].type==="paragraph_close"?o=e.tokens.pop():o=null,i=l[n].count>0?l[n].count:1,t=0;t<i;t++)e.tokens.push({type:"footnote_anchor",id:n,subId:t,level:u});o&&e.tokens.push(o),e.tokens.push({type:"footnote_close",level:--u})}e.tokens.push({type:"footnote_block_close",level:--u})}}var he=` 
+()[]'".,!?-`;function J(e){return e.replace(/([-()\[\]{}+?*.$\^|,:#<!\\])/g,"\\$1")}function Je(e){var n,r,t,i,o,l,c,s,a,u,p,f,h=e.tokens;if(e.env.abbreviations){for(e.env.abbrRegExp||(f="(^|["+he.split("").map(J).join("")+"])("+Object.keys(e.env.abbreviations).map(function(g){return g.substr(1)}).sort(function(g,d){return d.length-g.length}).map(J).join("|")+")($|["+he.split("").map(J).join("")+"])",e.env.abbrRegExp=new RegExp(f,"g")),u=e.env.abbrRegExp,r=0,t=h.length;r<t;r++)if(h[r].type==="inline"){for(i=h[r].children,n=i.length-1;n>=0;n--)if(o=i[n],o.type==="text"){for(s=0,l=o.content,u.lastIndex=0,a=o.level,c=[];p=u.exec(l);)u.lastIndex>s&&c.push({type:"text",content:l.slice(s,p.index+p[1].length),level:a}),c.push({type:"abbr_open",title:e.env.abbreviations[":"+p[2]],level:a++}),c.push({type:"text",content:p[2],level:a}),c.push({type:"abbr_close",level:--a}),s=u.lastIndex-p[3].length;c.length&&(s<l.length&&c.push({type:"text",content:l.slice(s),level:a}),h[r].children=i=[].concat(i.slice(0,n),c,i.slice(n+1)))}}}}var Ke=/\+-|\.\.|\?\?\?\?|!!!!|,,|--/,Qe=/\((c|tm|r|p)\)/ig,en={c:"\xA9",r:"\xAE",p:"\xA7",tm:"\u2122"};function nn(e){return e.indexOf("(")<0?e:e.replace(Qe,function(n,r){return en[r.toLowerCase()]})}function rn(e){var n,r,t,i,o;if(e.options.typographer){for(o=e.tokens.length-1;o>=0;o--)if(e.tokens[o].type==="inline")for(i=e.tokens[o].children,n=i.length-1;n>=0;n--)r=i[n],r.type==="text"&&(t=r.content,t=nn(t),Ke.test(t)&&(t=t.replace(/\+-/g,"\xB1").replace(/\.{2,}/g,"\u2026").replace(/([?!])…/g,"$1..").replace(/([?!]){4,}/g,"$1$1$1").replace(/,{2,}/g,",").replace(/(^|[^-])---([^-]|$)/mg,"$1\u2014$2").replace(/(^|\s)--(\s|$)/mg,"$1\u2013$2").replace(/(^|[^-\s])--([^-\s]|$)/mg,"$1\u2013$2")),r.content=t)}}var tn=/['"]/,pe=/['"]/g,on=/[-\s()\[\]]/,de="\u2019";function me(e,n){return n<0||n>=e.length?!1:!on.test(e[n])}function w(e,n,r){return e.substr(0,n)+r+e.substr(n+1)}function ln(e){var n,r,t,i,o,l,c,s,a,u,p,f,h,g,d,v,b;if(e.options.typographer){for(b=[],d=e.tokens.length-1;d>=0;d--)if(e.tokens[d].type==="inline"){for(v=e.tokens[d].children,b.length=0,n=0;n<v.length;n++)if(r=v[n],!(r.type!=="text"||tn.test(r.text))){for(c=v[n].level,h=b.length-1;h>=0&&!(b[h].level<=c);h--);b.length=h+1,t=r.content,o=0,l=t.length;e:for(;o<l&&(pe.lastIndex=o,i=pe.exec(t),!!i);){if(s=!me(t,i.index-1),o=i.index+1,g=i[0]==="'",a=!me(t,o),!a&&!s){g&&(r.content=w(r.content,i.index,de));continue}if(p=!a,f=!s,f){for(h=b.length-1;h>=0&&(u=b[h],!(b[h].level<c));h--)if(u.single===g&&b[h].level===c){u=b[h],g?(v[u.token].content=w(v[u.token].content,u.pos,e.options.quotes[2]),r.content=w(r.content,i.index,e.options.quotes[3])):(v[u.token].content=w(v[u.token].content,u.pos,e.options.quotes[0]),r.content=w(r.content,i.index,e.options.quotes[1])),b.length=h;continue e}}p?b.push({token:n,pos:i.index,single:g,level:c}):f&&g&&(r.content=w(r.content,i.index,de))}}}}}var K=[["block",Ve],["abbr",Ye],["references",Xe],["inline",Ge],["footnote_tail",We],["abbr2",Je],["replacements",rn],["smartquotes",ln]];function Se(){this.options={},this.ruler=new k;for(var e=0;e<K.length;e++)this.ruler.push(K[e][0],K[e][1])}Se.prototype.process=function(e){var n,r,t;for(t=this.ruler.getRules(""),n=0,r=t.length;n<r;n++)t[n](e)};function I(e,n,r,t,i){var o,l,c,s,a,u,p;for(this.src=e,this.parser=n,this.options=r,this.env=t,this.tokens=i,this.bMarks=[],this.eMarks=[],this.tShift=[],this.blkIndent=0,this.line=0,this.lineMax=0,this.tight=!1,this.parentType="root",this.ddIndent=-1,this.level=0,this.result="",l=this.src,u=0,p=!1,c=s=u=0,a=l.length;s<a;s++){if(o=l.charCodeAt(s),!p)if(o===32){u++;continue}else p=!0;(o===10||s===a-1)&&(o!==10&&s++,this.bMarks.push(c),this.eMarks.push(s),this.tShift.push(u),p=!1,u=0,c=s+1)}this.bMarks.push(l.length),this.eMarks.push(l.length),this.tShift.push(0),this.lineMax=this.bMarks.length-1}I.prototype.isEmpty=function(n){return this.bMarks[n]+this.tShift[n]>=this.eMarks[n]};I.prototype.skipEmptyLines=function(n){for(var r=this.lineMax;n<r&&!(this.bMarks[n]+this.tShift[n]<this.eMarks[n]);n++);return n};I.prototype.skipSpaces=function(n){for(var r=this.src.length;n<r&&this.src.charCodeAt(n)===32;n++);return n};I.prototype.skipChars=function(n,r){for(var t=this.src.length;n<t&&this.src.charCodeAt(n)===r;n++);return n};I.prototype.skipCharsBack=function(n,r,t){if(n<=t)return n;for(;n>t;)if(r!==this.src.charCodeAt(--n))return n+1;return n};I.prototype.getLines=function(n,r,t,i){var o,l,c,s,a,u=n;if(n>=r)return"";if(u+1===r)return l=this.bMarks[u]+Math.min(this.tShift[u],t),c=i?this.eMarks[u]+1:this.eMarks[u],this.src.slice(l,c);for(s=new Array(r-n),o=0;u<r;u++,o++)a=this.tShift[u],a>t&&(a=t),a<0&&(a=0),l=this.bMarks[u]+a,u+1<r||i?c=this.eMarks[u]+1:c=this.eMarks[u],s[o]=this.src.slice(l,c);return s.join("")};function cn(e,n,r){var t,i;if(e.tShift[n]-e.blkIndent<4)return!1;for(i=t=n+1;t<r;){if(e.isEmpty(t)){t++;continue}if(e.tShift[t]-e.blkIndent>=4){t++,i=t;continue}break}return e.line=t,e.tokens.push({type:"code",content:e.getLines(n,i,4+e.blkIndent,!0),block:!0,lines:[n,e.line],level:e.level}),!0}function sn(e,n,r,t){var i,o,l,c,s,a=!1,u=e.bMarks[n]+e.tShift[n],p=e.eMarks[n];if(u+3>p||(i=e.src.charCodeAt(u),i!==126&&i!==96)||(s=u,u=e.skipChars(u,i),o=u-s,o<3)||(l=e.src.slice(u,p).trim(),l.indexOf("`")>=0))return!1;if(t)return!0;for(c=n;c++,!(c>=r||(u=s=e.bMarks[c]+e.tShift[c],p=e.eMarks[c],u<p&&e.tShift[c]<e.blkIndent));)if(e.src.charCodeAt(u)===i&&!(e.tShift[c]-e.blkIndent>=4)&&(u=e.skipChars(u,i),!(u-s<o)&&(u=e.skipSpaces(u),!(u<p)))){a=!0;break}return o=e.tShift[n],e.line=c+(a?1:0),e.tokens.push({type:"fence",params:l,content:e.getLines(n+1,c,o,!0),lines:[n,e.line],level:e.level}),!0}function an(e,n,r,t){var i,o,l,c,s,a,u,p,f,h,g,d=e.bMarks[n]+e.tShift[n],v=e.eMarks[n];if(d>v||e.src.charCodeAt(d++)!==62||e.level>=e.options.maxNesting)return!1;if(t)return!0;for(e.src.charCodeAt(d)===32&&d++,s=e.blkIndent,e.blkIndent=0,c=[e.bMarks[n]],e.bMarks[n]=d,d=d<v?e.skipSpaces(d):d,o=d>=v,l=[e.tShift[n]],e.tShift[n]=d-e.bMarks[n],p=e.parser.ruler.getRules("blockquote"),i=n+1;i<r&&(d=e.bMarks[i]+e.tShift[i],v=e.eMarks[i],!(d>=v));i++){if(e.src.charCodeAt(d++)===62){e.src.charCodeAt(d)===32&&d++,c.push(e.bMarks[i]),e.bMarks[i]=d,d=d<v?e.skipSpaces(d):d,o=d>=v,l.push(e.tShift[i]),e.tShift[i]=d-e.bMarks[i];continue}if(o)break;for(g=!1,f=0,h=p.length;f<h;f++)if(p[f](e,i,r,!0)){g=!0;break}if(g)break;c.push(e.bMarks[i]),l.push(e.tShift[i]),e.tShift[i]=-1337}for(a=e.parentType,e.parentType="blockquote",e.tokens.push({type:"blockquote_open",lines:u=[n,0],level:e.level++}),e.parser.tokenize(e,n,i),e.tokens.push({type:"blockquote_close",level:--e.level}),e.parentType=a,u[1]=e.line,f=0;f<l.length;f++)e.bMarks[f+n]=c[f],e.tShift[f+n]=l[f];return e.blkIndent=s,!0}function un(e,n,r,t){var i,o,l,c=e.bMarks[n],s=e.eMarks[n];if(c+=e.tShift[n],c>s||(i=e.src.charCodeAt(c++),i!==42&&i!==45&&i!==95))return!1;for(o=1;c<s;){if(l=e.src.charCodeAt(c++),l!==i&&l!==32)return!1;l===i&&o++}return o<3?!1:(t||(e.line=n+1,e.tokens.push({type:"hr",lines:[n,e.line],level:e.level})),!0)}function ge(e,n){var r,t,i;return t=e.bMarks[n]+e.tShift[n],i=e.eMarks[n],t>=i||(r=e.src.charCodeAt(t++),r!==42&&r!==45&&r!==43)||t<i&&e.src.charCodeAt(t)!==32?-1:t}function ve(e,n){var r,t=e.bMarks[n]+e.tShift[n],i=e.eMarks[n];if(t+1>=i||(r=e.src.charCodeAt(t++),r<48||r>57))return-1;for(;;){if(t>=i)return-1;if(r=e.src.charCodeAt(t++),!(r>=48&&r<=57)){if(r===41||r===46)break;return-1}}return t<i&&e.src.charCodeAt(t)!==32?-1:t}function fn(e,n){var r,t,i=e.level+2;for(r=n+2,t=e.tokens.length-2;r<t;r++)e.tokens[r].level===i&&e.tokens[r].type==="paragraph_open"&&(e.tokens[r+2].tight=!0,e.tokens[r].tight=!0,r+=2)}function hn(e,n,r,t){var i,o,l,c,s,a,u,p,f,h,g,d,v,b,A,E,_,ae,ue=!0,G,R,fe,W;if((p=ve(e,n))>=0)v=!0;else if((p=ge(e,n))>=0)v=!1;else return!1;if(e.level>=e.options.maxNesting)return!1;if(d=e.src.charCodeAt(p-1),t)return!0;for(A=e.tokens.length,v?(u=e.bMarks[n]+e.tShift[n],g=Number(e.src.substr(u,p-u-1)),e.tokens.push({type:"ordered_list_open",order:g,lines:_=[n,0],level:e.level++})):e.tokens.push({type:"bullet_list_open",lines:_=[n,0],level:e.level++}),i=n,E=!1,G=e.parser.ruler.getRules("list");i<r&&(b=e.skipSpaces(p),f=e.eMarks[i],b>=f?h=1:h=b-p,h>4&&(h=1),h<1&&(h=1),o=p-e.bMarks[i]+h,e.tokens.push({type:"list_item_open",lines:ae=[n,0],level:e.level++}),c=e.blkIndent,s=e.tight,l=e.tShift[n],a=e.parentType,e.tShift[n]=b-e.bMarks[n],e.blkIndent=o,e.tight=!0,e.parentType="list",e.parser.tokenize(e,n,r,!0),(!e.tight||E)&&(ue=!1),E=e.line-n>1&&e.isEmpty(e.line-1),e.blkIndent=c,e.tShift[n]=l,e.tight=s,e.parentType=a,e.tokens.push({type:"list_item_close",level:--e.level}),i=n=e.line,ae[1]=i,b=e.bMarks[n],!(i>=r||e.isEmpty(i)||e.tShift[i]<e.blkIndent));){for(W=!1,R=0,fe=G.length;R<fe;R++)if(G[R](e,i,r,!0)){W=!0;break}if(W)break;if(v){if(p=ve(e,i),p<0)break}else if(p=ge(e,i),p<0)break;if(d!==e.src.charCodeAt(p-1))break}return e.tokens.push({type:v?"ordered_list_close":"bullet_list_close",level:--e.level}),_[1]=i,e.line=i,ue&&fn(e,A),!0}function pn(e,n,r,t){var i,o,l,c,s,a=e.bMarks[n]+e.tShift[n],u=e.eMarks[n];if(a+4>u||e.src.charCodeAt(a)!==91||e.src.charCodeAt(a+1)!==94||e.level>=e.options.maxNesting)return!1;for(c=a+2;c<u;c++){if(e.src.charCodeAt(c)===32)return!1;if(e.src.charCodeAt(c)===93)break}return c===a+2||c+1>=u||e.src.charCodeAt(++c)!==58?!1:(t||(c++,e.env.footnotes||(e.env.footnotes={}),e.env.footnotes.refs||(e.env.footnotes.refs={}),s=e.src.slice(a+2,c-2),e.env.footnotes.refs[":"+s]=-1,e.tokens.push({type:"footnote_reference_open",label:s,level:e.level++}),i=e.bMarks[n],o=e.tShift[n],l=e.parentType,e.tShift[n]=e.skipSpaces(c)-c,e.bMarks[n]=c,e.blkIndent+=4,e.parentType="footnote",e.tShift[n]<e.blkIndent&&(e.tShift[n]+=e.blkIndent,e.bMarks[n]-=e.blkIndent),e.parser.tokenize(e,n,r,!0),e.parentType=l,e.blkIndent-=4,e.tShift[n]=o,e.bMarks[n]=i,e.tokens.push({type:"footnote_reference_close",level:--e.level})),!0)}function dn(e,n,r,t){var i,o,l,c=e.bMarks[n]+e.tShift[n],s=e.eMarks[n];if(c>=s||(i=e.src.charCodeAt(c),i!==35||c>=s))return!1;for(o=1,i=e.src.charCodeAt(++c);i===35&&c<s&&o<=6;)o++,i=e.src.charCodeAt(++c);return o>6||c<s&&i!==32?!1:(t||(s=e.skipCharsBack(s,32,c),l=e.skipCharsBack(s,35,c),l>c&&e.src.charCodeAt(l-1)===32&&(s=l),e.line=n+1,e.tokens.push({type:"heading_open",hLevel:o,lines:[n,e.line],level:e.level}),c<s&&e.tokens.push({type:"inline",content:e.src.slice(c,s).trim(),level:e.level+1,lines:[n,e.line],children:[]}),e.tokens.push({type:"heading_close",hLevel:o,level:e.level})),!0)}function mn(e,n,r){var t,i,o,l=n+1;return l>=r||e.tShift[l]<e.blkIndent||e.tShift[l]-e.blkIndent>3||(i=e.bMarks[l]+e.tShift[l],o=e.eMarks[l],i>=o)||(t=e.src.charCodeAt(i),t!==45&&t!==61)||(i=e.skipChars(i,t),i=e.skipSpaces(i),i<o)?!1:(i=e.bMarks[n]+e.tShift[n],e.line=l+1,e.tokens.push({type:"heading_open",hLevel:t===61?1:2,lines:[n,e.line],level:e.level}),e.tokens.push({type:"inline",content:e.src.slice(i,e.eMarks[n]).trim(),level:e.level+1,lines:[n,e.line-1],children:[]}),e.tokens.push({type:"heading_close",hLevel:t===61?1:2,level:e.level}),!0)}var Me={};["article","aside","button","blockquote","body","canvas","caption","col","colgroup","dd","div","dl","dt","embed","fieldset","figcaption","figure","footer","form","h1","h2","h3","h4","h5","h6","header","hgroup","hr","iframe","li","map","object","ol","output","p","pre","progress","script","section","style","table","tbody","td","textarea","tfoot","th","tr","thead","ul","video"].forEach(function(e){Me[e]=!0});var gn=/^<([a-zA-Z]{1,15})[\s\/>]/,vn=/^<\/([a-zA-Z]{1,15})[\s>]/;function bn(e){var n=e|32;return n>=97&&n<=122}function _n(e,n,r,t){var i,o,l,c=e.bMarks[n],s=e.eMarks[n],a=e.tShift[n];if(c+=a,!e.options.html||a>3||c+2>=s||e.src.charCodeAt(c)!==60)return!1;if(i=e.src.charCodeAt(c+1),i===33||i===63){if(t)return!0}else if(i===47||bn(i)){if(i===47){if(o=e.src.slice(c,s).match(vn),!o)return!1}else if(o=e.src.slice(c,s).match(gn),!o)return!1;if(Me[o[1].toLowerCase()]!==!0)return!1;if(t)return!0}else return!1;for(l=n+1;l<e.lineMax&&!e.isEmpty(l);)l++;return e.line=l,e.tokens.push({type:"htmlblock",level:e.level,lines:[n,e.line],content:e.getLines(n,l,0,!0)}),!0}function Q(e,n){var r=e.bMarks[n]+e.blkIndent,t=e.eMarks[n];return e.src.substr(r,t-r)}function xn(e,n,r,t){var i,o,l,c,s,a,u,p,f,h,g;if(n+2>r||(s=n+1,e.tShift[s]<e.blkIndent)||(l=e.bMarks[s]+e.tShift[s],l>=e.eMarks[s])||(i=e.src.charCodeAt(l),i!==124&&i!==45&&i!==58)||(o=Q(e,n+1),!/^[-:| ]+$/.test(o))||(a=o.split("|"),a<=2))return!1;for(p=[],c=0;c<a.length;c++){if(f=a[c].trim(),!f){if(c===0||c===a.length-1)continue;return!1}if(!/^:?-+:?$/.test(f))return!1;f.charCodeAt(f.length-1)===58?p.push(f.charCodeAt(0)===58?"center":"right"):f.charCodeAt(0)===58?p.push("left"):p.push("")}if(o=Q(e,n).trim(),o.indexOf("|")===-1||(a=o.replace(/^\||\|$/g,"").split("|"),p.length!==a.length))return!1;if(t)return!0;for(e.tokens.push({type:"table_open",lines:h=[n,0],level:e.level++}),e.tokens.push({type:"thead_open",lines:[n,n+1],level:e.level++}),e.tokens.push({type:"tr_open",lines:[n,n+1],level:e.level++}),c=0;c<a.length;c++)e.tokens.push({type:"th_open",align:p[c],lines:[n,n+1],level:e.level++}),e.tokens.push({type:"inline",content:a[c].trim(),lines:[n,n+1],level:e.level,children:[]}),e.tokens.push({type:"th_close",level:--e.level});for(e.tokens.push({type:"tr_close",level:--e.level}),e.tokens.push({type:"thead_close",level:--e.level}),e.tokens.push({type:"tbody_open",lines:g=[n+2,0],level:e.level++}),s=n+2;s<r&&!(e.tShift[s]<e.blkIndent||(o=Q(e,s).trim(),o.indexOf("|")===-1));s++){for(a=o.replace(/^\||\|$/g,"").split("|"),e.tokens.push({type:"tr_open",level:e.level++}),c=0;c<a.length;c++)e.tokens.push({type:"td_open",align:p[c],level:e.level++}),u=a[c].substring(a[c].charCodeAt(0)===124?1:0,a[c].charCodeAt(a[c].length-1)===124?a[c].length-1:a[c].length).trim(),e.tokens.push({type:"inline",content:u,level:e.level,children:[]}),e.tokens.push({type:"td_close",level:--e.level});e.tokens.push({type:"tr_close",level:--e.level})}return e.tokens.push({type:"tbody_close",level:--e.level}),e.tokens.push({type:"table_close",level:--e.level}),h[1]=g[1]=s,e.line=s,!0}function F(e,n){var r,t,i=e.bMarks[n]+e.tShift[n],o=e.eMarks[n];return i>=o||(t=e.src.charCodeAt(i++),t!==126&&t!==58)||(r=e.skipSpaces(i),i===r)||r>=o?-1:r}function kn(e,n){var r,t,i=e.level+2;for(r=n+2,t=e.tokens.length-2;r<t;r++)e.tokens[r].level===i&&e.tokens[r].type==="paragraph_open"&&(e.tokens[r+2].tight=!0,e.tokens[r].tight=!0,r+=2)}function yn(e,n,r,t){var i,o,l,c,s,a,u,p,f,h,g,d,v,b;if(t)return e.ddIndent<0?!1:F(e,n)>=0;if(u=n+1,e.isEmpty(u)&&++u>r||e.tShift[u]<e.blkIndent||(i=F(e,u),i<0)||e.level>=e.options.maxNesting)return!1;a=e.tokens.length,e.tokens.push({type:"dl_open",lines:s=[n,0],level:e.level++}),l=n,o=u;e:for(;;){for(b=!0,v=!1,e.tokens.push({type:"dt_open",lines:[l,l],level:e.level++}),e.tokens.push({type:"inline",content:e.getLines(l,l+1,e.blkIndent,!1).trim(),level:e.level+1,lines:[l,l],children:[]}),e.tokens.push({type:"dt_close",level:--e.level});;){if(e.tokens.push({type:"dd_open",lines:c=[u,0],level:e.level++}),d=e.tight,f=e.ddIndent,p=e.blkIndent,g=e.tShift[o],h=e.parentType,e.blkIndent=e.ddIndent=e.tShift[o]+2,e.tShift[o]=i-e.bMarks[o],e.tight=!0,e.parentType="deflist",e.parser.tokenize(e,o,r,!0),(!e.tight||v)&&(b=!1),v=e.line-o>1&&e.isEmpty(e.line-1),e.tShift[o]=g,e.tight=d,e.parentType=h,e.blkIndent=p,e.ddIndent=f,e.tokens.push({type:"dd_close",level:--e.level}),c[1]=u=e.line,u>=r||e.tShift[u]<e.blkIndent)break e;if(i=F(e,u),i<0)break;o=u}if(u>=r||(l=u,e.isEmpty(l))||e.tShift[l]<e.blkIndent||(o=l+1,o>=r)||(e.isEmpty(o)&&o++,o>=r)||e.tShift[o]<e.blkIndent||(i=F(e,o),i<0))break}return e.tokens.push({type:"dl_close",level:--e.level}),s[1]=u,e.line=u,b&&kn(e,a),!0}function Cn(e,n){var r,t,i,o,l,c=n+1,s;if(r=e.lineMax,c<r&&!e.isEmpty(c)){for(s=e.parser.ruler.getRules("paragraph");c<r&&!e.isEmpty(c);c++)if(!(e.tShift[c]-e.blkIndent>3)){for(i=!1,o=0,l=s.length;o<l;o++)if(s[o](e,c,r,!0)){i=!0;break}if(i)break}}return t=e.getLines(n,c,e.blkIndent,!1).trim(),e.line=c,t.length&&(e.tokens.push({type:"paragraph_open",tight:!1,lines:[n,e.line],level:e.level}),e.tokens.push({type:"inline",content:t,level:e.level+1,lines:[n,e.line],children:[]}),e.tokens.push({type:"paragraph_close",tight:!1,level:e.level})),!0}var P=[["code",cn],["fences",sn,["paragraph","blockquote","list"]],["blockquote",an,["paragraph","blockquote","list"]],["hr",un,["paragraph","blockquote","list"]],["list",hn,["paragraph","blockquote"]],["footnote",pn,["paragraph"]],["heading",dn,["paragraph","blockquote"]],["lheading",mn],["htmlblock",_n,["paragraph","blockquote"]],["table",xn,["paragraph"]],["deflist",yn,["paragraph"]],["paragraph",Cn]];function oe(){this.ruler=new k;for(var e=0;e<P.length;e++)this.ruler.push(P[e][0],P[e][1],{alt:(P[e][2]||[]).slice()})}oe.prototype.tokenize=function(e,n,r){for(var t=this.ruler.getRules(""),i=t.length,o=n,l=!1,c,s;o<r&&(e.line=o=e.skipEmptyLines(o),!(o>=r||e.tShift[o]<e.blkIndent));){for(s=0;s<i&&(c=t[s](e,o,r,!1),!c);s++);if(e.tight=!l,e.isEmpty(e.line-1)&&(l=!0),o=e.line,o<r&&e.isEmpty(o)){if(l=!0,o++,o<r&&e.parentType==="list"&&e.isEmpty(o))break;e.line=o}}};var En=/[\n\t]/g,An=/\r[\n\u0085]|[\u2424\u2028\u0085]/g,Tn=/\u00a0/g;oe.prototype.parse=function(e,n,r,t){var i,o=0,l=0;if(!e)return[];e=e.replace(Tn," "),e=e.replace(An,`
+`),e.indexOf("	")>=0&&(e=e.replace(En,function(c,s){var a;return e.charCodeAt(s)===10?(o=s+1,l=0,c):(a="    ".slice((s-o-l)%4),l=s-o+1,a)})),i=new I(e,this,n,r,t),this.tokenize(i,i.line,i.lineMax)};function Sn(e){switch(e){case 10:case 92:case 96:case 42:case 95:case 94:case 91:case 93:case 33:case 38:case 60:case 62:case 123:case 125:case 36:case 37:case 64:case 126:case 43:case 61:case 58:return!0;default:return!1}}function Mn(e,n){for(var r=e.pos;r<e.posMax&&!Sn(e.src.charCodeAt(r));)r++;return r===e.pos?!1:(n||(e.pending+=e.src.slice(e.pos,r)),e.pos=r,!0)}function In(e,n){var r,t,i=e.pos;if(e.src.charCodeAt(i)!==10)return!1;if(r=e.pending.length-1,t=e.posMax,!n)if(r>=0&&e.pending.charCodeAt(r)===32)if(r>=1&&e.pending.charCodeAt(r-1)===32){for(var o=r-2;o>=0;o--)if(e.pending.charCodeAt(o)!==32){e.pending=e.pending.substring(0,o+1);break}e.push({type:"hardbreak",level:e.level})}else e.pending=e.pending.slice(0,-1),e.push({type:"softbreak",level:e.level});else e.push({type:"softbreak",level:e.level});for(i++;i<t&&e.src.charCodeAt(i)===32;)i++;return e.pos=i,!0}var le=[];for(ee=0;ee<256;ee++)le.push(0);var ee;"\\!\"#$%&'()*+,./:;<=>?@[]^_`{|}~-".split("").forEach(function(e){le[e.charCodeAt(0)]=1});function wn(e,n){var r,t=e.pos,i=e.posMax;if(e.src.charCodeAt(t)!==92)return!1;if(t++,t<i){if(r=e.src.charCodeAt(t),r<256&&le[r]!==0)return n||(e.pending+=e.src[t]),e.pos+=2,!0;if(r===10){for(n||e.push({type:"hardbreak",level:e.level}),t++;t<i&&e.src.charCodeAt(t)===32;)t++;return e.pos=t,!0}}return n||(e.pending+="\\"),e.pos++,!0}function Ln(e,n){var r,t,i,o,l,c=e.pos,s=e.src.charCodeAt(c);if(s!==96)return!1;for(r=c,c++,t=e.posMax;c<t&&e.src.charCodeAt(c)===96;)c++;for(i=e.src.slice(r,c),o=l=c;(o=e.src.indexOf("`",l))!==-1;){for(l=o+1;l<t&&e.src.charCodeAt(l)===96;)l++;if(l-o===i.length)return n||e.push({type:"code",content:e.src.slice(c,o).replace(/[ \n]+/g," ").trim(),block:!1,level:e.level}),e.pos=l,!0}return n||(e.pending+=i),e.pos+=i.length,!0}function qn(e,n){var r,t,i,o=e.posMax,l=e.pos,c,s;if(e.src.charCodeAt(l)!==126||n||l+4>=o||e.src.charCodeAt(l+1)!==126||e.level>=e.options.maxNesting||(c=l>0?e.src.charCodeAt(l-1):-1,s=e.src.charCodeAt(l+2),c===126)||s===126||s===32||s===10)return!1;for(t=l+2;t<o&&e.src.charCodeAt(t)===126;)t++;if(t>l+3)return e.pos+=t-l,n||(e.pending+=e.src.slice(l,t)),!0;for(e.pos=l+2,i=1;e.pos+1<o;){if(e.src.charCodeAt(e.pos)===126&&e.src.charCodeAt(e.pos+1)===126&&(c=e.src.charCodeAt(e.pos-1),s=e.pos+2<o?e.src.charCodeAt(e.pos+2):-1,s!==126&&c!==126&&(c!==32&&c!==10?i--:s!==32&&s!==10&&i++,i<=0))){r=!0;break}e.parser.skipToken(e)}return r?(e.posMax=e.pos,e.pos=l+2,n||(e.push({type:"del_open",level:e.level++}),e.parser.tokenize(e),e.push({type:"del_close",level:--e.level})),e.pos=e.posMax+2,e.posMax=o,!0):(e.pos=l,!1)}function Bn(e,n){var r,t,i,o=e.posMax,l=e.pos,c,s;if(e.src.charCodeAt(l)!==43||n||l+4>=o||e.src.charCodeAt(l+1)!==43||e.level>=e.options.maxNesting||(c=l>0?e.src.charCodeAt(l-1):-1,s=e.src.charCodeAt(l+2),c===43)||s===43||s===32||s===10)return!1;for(t=l+2;t<o&&e.src.charCodeAt(t)===43;)t++;if(t!==l+2)return e.pos+=t-l,n||(e.pending+=e.src.slice(l,t)),!0;for(e.pos=l+2,i=1;e.pos+1<o;){if(e.src.charCodeAt(e.pos)===43&&e.src.charCodeAt(e.pos+1)===43&&(c=e.src.charCodeAt(e.pos-1),s=e.pos+2<o?e.src.charCodeAt(e.pos+2):-1,s!==43&&c!==43&&(c!==32&&c!==10?i--:s!==32&&s!==10&&i++,i<=0))){r=!0;break}e.parser.skipToken(e)}return r?(e.posMax=e.pos,e.pos=l+2,n||(e.push({type:"ins_open",level:e.level++}),e.parser.tokenize(e),e.push({type:"ins_close",level:--e.level})),e.pos=e.posMax+2,e.posMax=o,!0):(e.pos=l,!1)}function Nn(e,n){var r,t,i,o=e.posMax,l=e.pos,c,s;if(e.src.charCodeAt(l)!==61||n||l+4>=o||e.src.charCodeAt(l+1)!==61||e.level>=e.options.maxNesting||(c=l>0?e.src.charCodeAt(l-1):-1,s=e.src.charCodeAt(l+2),c===61)||s===61||s===32||s===10)return!1;for(t=l+2;t<o&&e.src.charCodeAt(t)===61;)t++;if(t!==l+2)return e.pos+=t-l,n||(e.pending+=e.src.slice(l,t)),!0;for(e.pos=l+2,i=1;e.pos+1<o;){if(e.src.charCodeAt(e.pos)===61&&e.src.charCodeAt(e.pos+1)===61&&(c=e.src.charCodeAt(e.pos-1),s=e.pos+2<o?e.src.charCodeAt(e.pos+2):-1,s!==61&&c!==61&&(c!==32&&c!==10?i--:s!==32&&s!==10&&i++,i<=0))){r=!0;break}e.parser.skipToken(e)}return r?(e.posMax=e.pos,e.pos=l+2,n||(e.push({type:"mark_open",level:e.level++}),e.parser.tokenize(e),e.push({type:"mark_close",level:--e.level})),e.pos=e.posMax+2,e.posMax=o,!0):(e.pos=l,!1)}function be(e){return e>=48&&e<=57||e>=65&&e<=90||e>=97&&e<=122}function _e(e,n){var r=n,t,i,o,l=!0,c=!0,s=e.posMax,a=e.src.charCodeAt(n);for(t=n>0?e.src.charCodeAt(n-1):-1;r<s&&e.src.charCodeAt(r)===a;)r++;return r>=s&&(l=!1),o=r-n,o>=4?l=c=!1:(i=r<s?e.src.charCodeAt(r):-1,(i===32||i===10)&&(l=!1),(t===32||t===10)&&(c=!1),a===95&&(be(t)&&(l=!1),be(i)&&(c=!1))),{can_open:l,can_close:c,delims:o}}function On(e,n){var r,t,i,o,l,c,s,a=e.posMax,u=e.pos,p=e.src.charCodeAt(u);if(p!==95&&p!==42||n)return!1;if(s=_e(e,u),r=s.delims,!s.can_open)return e.pos+=r,n||(e.pending+=e.src.slice(u,e.pos)),!0;if(e.level>=e.options.maxNesting)return!1;for(e.pos=u+r,c=[r];e.pos<a;){if(e.src.charCodeAt(e.pos)===p){if(s=_e(e,e.pos),t=s.delims,s.can_close){for(o=c.pop(),l=t;o!==l;){if(l<o){c.push(o-l);break}if(l-=o,c.length===0)break;e.pos+=o,o=c.pop()}if(c.length===0){r=o,i=!0;break}e.pos+=t;continue}s.can_open&&c.push(t),e.pos+=t;continue}e.parser.skipToken(e)}return i?(e.posMax=e.pos,e.pos=u+r,n||((r===2||r===3)&&e.push({type:"strong_open",level:e.level++}),(r===1||r===3)&&e.push({type:"em_open",level:e.level++}),e.parser.tokenize(e),(r===1||r===3)&&e.push({type:"em_close",level:--e.level}),(r===2||r===3)&&e.push({type:"strong_close",level:--e.level})),e.pos=e.posMax+r,e.posMax=a,!0):(e.pos=u,!1)}var Dn=/\\([ \\!"#$%&'()*+,.\/:;<=>?@[\]^_`{|}~-])/g;function Rn(e,n){var r,t,i=e.posMax,o=e.pos;if(e.src.charCodeAt(o)!==126||n||o+2>=i||e.level>=e.options.maxNesting)return!1;for(e.pos=o+1;e.pos<i;){if(e.src.charCodeAt(e.pos)===126){r=!0;break}e.parser.skipToken(e)}return!r||o+1===e.pos||(t=e.src.slice(o+1,e.pos),t.match(/(^|[^\\])(\\\\)*\s/))?(e.pos=o,!1):(e.posMax=e.pos,e.pos=o+1,n||e.push({type:"sub",level:e.level,content:t.replace(Dn,"$1")}),e.pos=e.posMax+1,e.posMax=i,!0)}var Hn=/\\([ \\!"#$%&'()*+,.\/:;<=>?@[\]^_`{|}~-])/g;function Fn(e,n){var r,t,i=e.posMax,o=e.pos;if(e.src.charCodeAt(o)!==94||n||o+2>=i||e.level>=e.options.maxNesting)return!1;for(e.pos=o+1;e.pos<i;){if(e.src.charCodeAt(e.pos)===94){r=!0;break}e.parser.skipToken(e)}return!r||o+1===e.pos||(t=e.src.slice(o+1,e.pos),t.match(/(^|[^\\])(\\\\)*\s/))?(e.pos=o,!1):(e.posMax=e.pos,e.pos=o+1,n||e.push({type:"sup",level:e.level,content:t.replace(Hn,"$1")}),e.pos=e.posMax+1,e.posMax=i,!0)}function Pn(e,n){var r,t,i,o,l,c,s,a,u=!1,p=e.pos,f=e.posMax,h=e.pos,g=e.src.charCodeAt(h);if(g===33&&(u=!0,g=e.src.charCodeAt(++h)),g!==91||e.level>=e.options.maxNesting||(r=h+1,t=N(e,h),t<0))return!1;if(c=t+1,c<f&&e.src.charCodeAt(c)===40){for(c++;c<f&&(a=e.src.charCodeAt(c),!(a!==32&&a!==10));c++);if(c>=f)return!1;for(h=c,Ee(e,c)?(o=e.linkContent,c=e.pos):o="",h=c;c<f&&(a=e.src.charCodeAt(c),!(a!==32&&a!==10));c++);if(c<f&&h!==c&&Ae(e,c))for(l=e.linkContent,c=e.pos;c<f&&(a=e.src.charCodeAt(c),!(a!==32&&a!==10));c++);else l="";if(c>=f||e.src.charCodeAt(c)!==41)return e.pos=p,!1;c++}else{if(e.linkLevel>0)return!1;for(;c<f&&(a=e.src.charCodeAt(c),!(a!==32&&a!==10));c++);if(c<f&&e.src.charCodeAt(c)===91&&(h=c+1,c=N(e,c),c>=0?i=e.src.slice(h,c++):c=h-1),i||(typeof i>"u"&&(c=t+1),i=e.src.slice(r,t)),s=e.env.references[Te(i)],!s)return e.pos=p,!1;o=s.href,l=s.title}return n||(e.pos=r,e.posMax=t,u?e.push({type:"image",src:o,title:l,alt:e.src.substr(r,t-r),level:e.level}):(e.push({type:"link_open",href:o,title:l,level:e.level++}),e.linkLevel++,e.parser.tokenize(e),e.linkLevel--,e.push({type:"link_close",level:--e.level}))),e.pos=c,e.posMax=f,!0}function jn(e,n){var r,t,i,o,l=e.posMax,c=e.pos;return c+2>=l||e.src.charCodeAt(c)!==94||e.src.charCodeAt(c+1)!==91||e.level>=e.options.maxNesting||(r=c+2,t=N(e,c+1),t<0)?!1:(n||(e.env.footnotes||(e.env.footnotes={}),e.env.footnotes.list||(e.env.footnotes.list=[]),i=e.env.footnotes.list.length,e.pos=r,e.posMax=t,e.push({type:"footnote_ref",id:i,level:e.level}),e.linkLevel++,o=e.tokens.length,e.parser.tokenize(e),e.env.footnotes.list[i]={tokens:e.tokens.splice(o)},e.linkLevel--),e.pos=t+1,e.posMax=l,!0)}function zn(e,n){var r,t,i,o,l=e.posMax,c=e.pos;if(c+3>l||!e.env.footnotes||!e.env.footnotes.refs||e.src.charCodeAt(c)!==91||e.src.charCodeAt(c+1)!==94||e.level>=e.options.maxNesting)return!1;for(t=c+2;t<l;t++){if(e.src.charCodeAt(t)===32||e.src.charCodeAt(t)===10)return!1;if(e.src.charCodeAt(t)===93)break}return t===c+2||t>=l||(t++,r=e.src.slice(c+2,t-1),typeof e.env.footnotes.refs[":"+r]>"u")?!1:(n||(e.env.footnotes.list||(e.env.footnotes.list=[]),e.env.footnotes.refs[":"+r]<0?(i=e.env.footnotes.list.length,e.env.footnotes.list[i]={label:r,count:0},e.env.footnotes.refs[":"+r]=i):i=e.env.footnotes.refs[":"+r],o=e.env.footnotes.list[i].count,e.env.footnotes.list[i].count++,e.push({type:"footnote_ref",id:i,subId:o,level:e.level})),e.pos=t,e.posMax=l,!0)}var $n=["coap","doi","javascript","aaa","aaas","about","acap","cap","cid","crid","data","dav","dict","dns","file","ftp","geo","go","gopher","h323","http","https","iax","icap","im","imap","info","ipp","iris","iris.beep","iris.xpc","iris.xpcs","iris.lwz","ldap","mailto","mid","msrp","msrps","mtqp","mupdate","news","nfs","ni","nih","nntp","opaquelocktoken","pop","pres","rtsp","service","session","shttp","sieve","sip","sips","sms","snmp","soap.beep","soap.beeps","tag","tel","telnet","tftp","thismessage","tn3270","tip","tv","urn","vemmi","ws","wss","xcon","xcon-userid","xmlrpc.beep","xmlrpc.beeps","xmpp","z39.50r","z39.50s","adiumxtra","afp","afs","aim","apt","attachment","aw","beshare","bitcoin","bolo","callto","chrome","chrome-extension","com-eventbrite-attendee","content","cvs","dlna-playsingle","dlna-playcontainer","dtn","dvb","ed2k","facetime","feed","finger","fish","gg","git","gizmoproject","gtalk","hcp","icon","ipn","irc","irc6","ircs","itms","jar","jms","keyparc","lastfm","ldaps","magnet","maps","market","message","mms","ms-help","msnim","mumble","mvn","notes","oid","palm","paparazzi","platform","proxy","psyc","query","res","resource","rmi","rsync","rtmp","secondlife","sftp","sgn","skype","smb","soldat","spotify","ssh","steam","svn","teamspeak","things","udp","unreal","ut2004","ventrilo","view-source","webcal","wtai","wyciwyg","xfire","xri","ymsgr"],Vn=/^<([a-zA-Z0-9.!#$%&'*+\/=?^_`{|}~-]+@[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?(?:\.[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?)*)>/,Un=/^<([a-zA-Z.\-]{1,25}):([^<>\x00-\x20]*)>/;function Yn(e,n){var r,t,i,o,l,c=e.pos;return e.src.charCodeAt(c)!==60||(r=e.src.slice(c),r.indexOf(">")<0)?!1:(t=r.match(Un),t?$n.indexOf(t[1].toLowerCase())<0||(o=t[0].slice(1,-1),l=te(o),!e.parser.validateLink(o))?!1:(n||(e.push({type:"link_open",href:l,level:e.level}),e.push({type:"text",content:o,level:e.level+1}),e.push({type:"link_close",level:e.level})),e.pos+=t[0].length,!0):(i=r.match(Vn),i?(o=i[0].slice(1,-1),l=te("mailto:"+o),e.parser.validateLink(l)?(n||(e.push({type:"link_open",href:l,level:e.level}),e.push({type:"text",content:o,level:e.level+1}),e.push({type:"link_close",level:e.level})),e.pos+=i[0].length,!0):!1):!1))}function j(e,n){return e=e.source,n=n||"",function r(t,i){return t?(i=i.source||i,e=e.replace(t,i),r):new RegExp(e,n)}}var Zn=/[a-zA-Z_:][a-zA-Z0-9:._-]*/,Xn=/[^"'=<>`\x00-\x20]+/,Gn=/'[^']*'/,Wn=/"[^"]*"/,Jn=j(/(?:unquoted|single_quoted|double_quoted)/)("unquoted",Xn)("single_quoted",Gn)("double_quoted",Wn)(),Kn=j(/(?:\s+attr_name(?:\s*=\s*attr_value)?)/)("attr_name",Zn)("attr_value",Jn)(),Qn=j(/<[A-Za-z][A-Za-z0-9]*attribute*\s*\/?>/)("attribute",Kn)(),er=/<\/[A-Za-z][A-Za-z0-9]*\s*>/,nr=/<!---->|<!--(?:-?[^>-])(?:-?[^-])*-->/,rr=/<[?].*?[?]>/,tr=/<![A-Z]+\s+[^>]*>/,ir=/<!\[CDATA\[[\s\S]*?\]\]>/,or=j(/^(?:open_tag|close_tag|comment|processing|declaration|cdata)/)("open_tag",Qn)("close_tag",er)("comment",nr)("processing",rr)("declaration",tr)("cdata",ir)();function lr(e){var n=e|32;return n>=97&&n<=122}function cr(e,n){var r,t,i,o=e.pos;return!e.options.html||(i=e.posMax,e.src.charCodeAt(o)!==60||o+2>=i)||(r=e.src.charCodeAt(o+1),r!==33&&r!==63&&r!==47&&!lr(r))||(t=e.src.slice(o).match(or),!t)?!1:(n||e.push({type:"htmltag",content:e.src.slice(o,o+t[0].length),level:e.level}),e.pos+=t[0].length,!0)}var sr=/^&#((?:x[a-f0-9]{1,8}|[0-9]{1,8}));/i,ar=/^&([a-z][a-z0-9]{1,31});/i;function ur(e,n){var r,t,i,o=e.pos,l=e.posMax;if(e.src.charCodeAt(o)!==38)return!1;if(o+1<l){if(r=e.src.charCodeAt(o+1),r===35){if(i=e.src.slice(o).match(sr),i)return n||(t=i[1][0].toLowerCase()==="x"?parseInt(i[1].slice(1),16):parseInt(i[1],10),e.pending+=ye(t)?re(t):re(65533)),e.pos+=i[0].length,!0}else if(i=e.src.slice(o).match(ar),i){var c=xe(i[1]);if(i[1]!==c)return n||(e.pending+=c),e.pos+=i[0].length,!0}}return n||(e.pending+="&"),e.pos++,!0}var ne=[["text",Mn],["newline",In],["escape",wn],["backticks",Ln],["del",qn],["ins",Bn],["mark",Nn],["emphasis",On],["sub",Rn],["sup",Fn],["links",Pn],["footnote_inline",jn],["footnote_ref",zn],["autolink",Yn],["htmltag",cr],["entity",ur]];function z(){this.ruler=new k;for(var e=0;e<ne.length;e++)this.ruler.push(ne[e][0],ne[e][1]);this.validateLink=fr}z.prototype.skipToken=function(e){var n=this.ruler.getRules(""),r=n.length,t=e.pos,i,o;if((o=e.cacheGet(t))>0){e.pos=o;return}for(i=0;i<r;i++)if(n[i](e,!0)){e.cacheSet(t,e.pos);return}e.pos++,e.cacheSet(t,e.pos)};z.prototype.tokenize=function(e){for(var n=this.ruler.getRules(""),r=n.length,t=e.posMax,i,o;e.pos<t;){for(o=0;o<r&&(i=n[o](e,!1),!i);o++);if(i){if(e.pos>=t)break;continue}e.pending+=e.src[e.pos++]}e.pending&&e.pushPending()};z.prototype.parse=function(e,n,r,t){var i=new M(e,this,n,r,t);this.tokenize(i)};function fr(e){var n=["vbscript","javascript","file","data"],r=e.trim().toLowerCase();return r=T(r),!(r.indexOf(":")!==-1&&n.indexOf(r.split(":")[0])!==-1)}var hr={options:{html:!1,xhtmlOut:!1,breaks:!1,langPrefix:"language-",linkTarget:"",typographer:!1,quotes:"\u201C\u201D\u2018\u2019",highlight:null,maxNesting:20},components:{core:{rules:["block","inline","references","replacements","smartquotes","references","abbr2","footnote_tail"]},block:{rules:["blockquote","code","fences","footnote","heading","hr","htmlblock","lheading","list","paragraph","table"]},inline:{rules:["autolink","backticks","del","emphasis","entity","escape","footnote_ref","htmltag","links","newline","text"]}}},pr={options:{html:!1,xhtmlOut:!1,breaks:!1,langPrefix:"language-",linkTarget:"",typographer:!1,quotes:"\u201C\u201D\u2018\u2019",highlight:null,maxNesting:20},components:{core:{},block:{},inline:{}}},dr={options:{html:!0,xhtmlOut:!0,breaks:!1,langPrefix:"language-",linkTarget:"",typographer:!1,quotes:"\u201C\u201D\u2018\u2019",highlight:null,maxNesting:20},components:{core:{rules:["block","inline","references","abbr2"]},block:{rules:["blockquote","code","fences","heading","hr","htmlblock","lheading","list","paragraph"]},inline:{rules:["autolink","backticks","emphasis","entity","escape","htmltag","links","newline","text"]}}},mr={default:hr,full:pr,commonmark:dr};function Ie(e,n,r){this.src=n,this.env=r,this.options=e.options,this.tokens=[],this.inlineMode=!1,this.inline=e.inline,this.block=e.block,this.renderer=e.renderer,this.typographer=e.typographer}function C(e,n){typeof e!="string"&&(n=e,e="default"),n&&n.linkify!=null&&console.warn(`linkify option is removed. Use linkify plugin instead:
+
+import Remarkable from 'remarkable';
+import linkify from 'remarkable/linkify';
+new Remarkable().use(linkify)
+`),this.inline=new z,this.block=new oe,this.core=new Se,this.renderer=new ie,this.ruler=new k,this.options={},this.configure(mr[e]),this.set(n||{})}C.prototype.set=function(e){ke(this.options,e)};C.prototype.configure=function(e){var n=this;if(!e)throw new Error("Wrong `remarkable` preset, check name/content");e.options&&n.set(e.options),e.components&&Object.keys(e.components).forEach(function(r){e.components[r].rules&&n[r].ruler.enable(e.components[r].rules,!0)})};C.prototype.use=function(e,n){return e(this,n),this};C.prototype.parse=function(e,n){var r=new Ie(this,e,n);return this.core.process(r),r.tokens};C.prototype.render=function(e,n){return n=n||{},this.renderer.render(this.parse(e,n),this.options,n)};C.prototype.parseInline=function(e,n){var r=new Ie(this,e,n);return r.inlineMode=!0,this.core.process(r),r.tokens};C.prototype.renderInline=function(e,n){return n=n||{},this.renderer.render(this.parseInline(e,n),this.options,n)};var $=class{constructor(){this.md=new C("full",{html:!0,typographer:!1}),this.openEmRenderer=this.md.renderer.rules.em_open,this.closeEmRenderer=this.md.renderer.rules.em_close,this.closeHeadingRenderer=this.md.renderer.rules.heading_close,this.injectedHeading=!1,this.injectedEm=!1,this.tagMap={r:"red",b:"blue",g:"green",y:"yellow",p:"pink",t:"teal","!":"spoiler"},this.updateRules()}updateRules(){this.md.renderer.rules.heading_open=function(n,r,t,i){if(n[r+1].content[0]==":"){if(this.injectedHeading=!0,n[r+1].content[1]=="{"){let o=this.parseJsonTag(n,r),l="<h"+n[r].hLevel+' id="'+n[r+1].content.replace(/\s+/g,"-").toLowerCase()+'">';return o+l}}else return"<h"+n[r].hLevel+' id="'+n[r+1].content.replace(/\s+/g,"-").toLowerCase()+'">'}.bind(this),this.md.renderer.rules.em_open=function(n,r,t,i){return this.parseTags(n,r,t,i,this.openEmRenderer)}.bind(this),this.md.renderer.rules.heading_close=function(n,r,t,i){return this.injectedHeading?(this.injectedHeading=!1,this.closeHeadingRenderer(n,r,t,i).concat("</span>").replace(/\n/g," ")+`
+`):this.closeHeadingRenderer(n,r,t,i)}.bind(this),this.md.renderer.rules.em_close=function(n,r,t,i){return this.injectedEm?(this.injectedEm=!1,"</span>"):this.closeEmRenderer(n,r,t,i)}.bind(this)}async loadFile(n){return new Promise(function(r,t){let i=new XMLHttpRequest;i.open("GET",n,!0),i.onload=function(){this.status>=200&&this.status<300?r(i.response):t({status:this.status,statusText:i.statusText})},i.onerror=function(){t({status:this.status,statusText:i.statusText})},i.send()})}async parseGFM(n){let r=await this.loadFile(n+".md");if(r.length<=1||r==null)return"";let t="";return t=this.md.render(r),t}parseTags(n,r,t,i,o){if(n[r+1].type=="text")if(n[r+1].content[0]==":"){if(this.injectedEm=!0,n[r+1].content[1]=="{")return this.parseJsonTag(n,r);{let l=n[r+1].content[1];return n[r+1].content=n[r+1].content.slice(2),'<span class="'+this.tagMap[l]+'">'}}else return o(n,r,t,i);else return o(n,r,t,i)}parseJsonTag(n,r){let t=n[r+1].content.indexOf("}"),i=n[r+1].content.substring(1,t+1);n[r+1].content=n[r+1].content.slice(t+2),n[r+1].type=="inline"&&(n[r+1].children[0].content=n[r+1].content);let o="tagging";return n[r].type=="em_open"&&(o="tagging-text"),'<span class="'+o+'" data-tags="'+i+'">'}};var V=class{constructor(n,r){this.navLock=!1,this.currentActiveSearch=null,this.techPages=n,this.addPageChangeEvent=r,this.searchBoxListener=this.searchBoxListener.bind(this)}initSearchIcons(){let n=document.querySelector("#nav-bar__search--icon");n.addEventListener("click",this.searchBoxListener),n.targetParam="search";let r=document.querySelector("#nav-bar__games--icon");r.addEventListener("click",this.searchBoxListener),r.targetParam="games",this.createBoxes(),this.fillSearch()}searchBoxListener(){if(this.navLock)return;let n=event.currentTarget.targetParam;if(this.currentActiveSearch=n,n=="search"?this.injectSearchBox("#nav-bar__search"):this.injectSearchBox("#nav-bar__games"),document.querySelector(".nav-bar__search--results")==null&&this.fillSearch(),n=="games"){let i=document.querySelector("#nav-bar__gamesbox").children;for(let o of i)o.style.display="inline-block";document.querySelector("#nav-bar__games--hr2D").style.display="none",document.querySelector("#nav-bar__games--hr3D").style.display="none"}}createBoxes(){let n=document.createElement("div"),r=document.createElement("div");n.id="nav-bar__searchbox",r.id="nav-bar__gamesbox",n.style.width="0px",r.style.width="0px",r.tabIndex="0";let t=document.createElement("input");t.id="nav-bar__search--input",t.setAttribute("aria-label","Search"),n.appendChild(t),document.querySelector("#nav-bar__search").appendChild(n),document.querySelector("#nav-bar__games").appendChild(r),t.addEventListener("keyup",this.filterFunction),t.addEventListener("focusout",l=>{let c=document.querySelector("#nav-bar__search");l.stopPropagation(),c.contains(l.relatedTarget)||this.clearFunction()}),r.addEventListener("focusout",l=>{let c=document.querySelector("#nav-bar__games");l.stopPropagation(),c.contains(l.relatedTarget)||this.clearFunction()})}injectSearchBox(n){if(this.navLock)return;document.querySelector(n).removeEventListener("click",this.searchBoxListener);var t=n+"box",i=n+"--input",o=n+"--icon",l=document.querySelector(t),c=document.querySelector(i);l=document.querySelector(t);var s=220;this.currentActiveSearch=="search"?document.body.clientWidth>=480&&(s=360):s=300,l.style.width=s+"px";let a=document.querySelector(o);a.className="material-symbols-rounded logo-fadeout",a.removeEventListener("click",this.searchBoxListener),a.addEventListener("click",this.clearFunction),this.currentActiveSearch=="search"?(c=document.querySelector(i),c.focus()):l.focus(),l.className="fadein"}clearFunction(){if(this.currentActiveSearch==null)return;var n,r,t=document.querySelector("#nav-bar__searchbox"),i=document.querySelector("#nav-bar__gamesbox");if(t==null||i==null)return;this.navLock=!0;var o=null;this.currentActiveSearch=="search"?o=t:o=i;let l="#nav-bar__"+this.currentActiveSearch;if(this.currentActiveSearch=="search"){for(n=o.getElementsByClassName("button__redirect"),r=0;r<n.length;r++)n[r].style.display="none";document.querySelector(l+"--hr").style.display="none",setTimeout(()=>{t.className="fadeout",this.navLock=!1},"300")}else{let s=o.getElementsByTagName("figure");for(r=0;r<s.length;r++)s[r].style.display="none";document.querySelector(l+"--hr2D").style.display="none",document.querySelector(l+"--hr3D").style.display="none",setTimeout(()=>{i.className="fadeout",this.navLock=!1},"300")}var c=document.querySelector(l+"--icon");c.className="material-symbols-rounded logo-fadein",c.removeEventListener("click",this.clearFunction),c.addEventListener("click",this.searchBoxListener),this.currentActiveSearch=null}filterFunction(){var n,r,t,i,o="#nav-bar__search";if(this.currentActiveSearch=="games")return;n=document.querySelector(o+"--input"),r=n.value.toUpperCase().trim();let l=r.split(" ");t=document.querySelector(o+"box").getElementsByClassName("button__redirect");let s=0;for(i=0;i<t.length;i++){let a=t[i].textContent||t[i].innerText;a=a+" "+t[i].dataset.section+" "+t[i].dataset.document;let u=a.toUpperCase(),p=!0;for(let f of l.values())u.includes(f)&&r.length!==0||(p=!1);p?(s<6||this.currentActiveSearch=="games"?t[i].style.display="block":t[i].style.display="none",s=s+1):t[i].style.display="none"}s>0?document.querySelector(o+"--hr").style.display="block":document.querySelector(o+"--hr").style.display="none"}fillSearch(){var n="",r="",t="",i="",o=document.querySelector("#nav-bar__searchbox"),l=document.querySelector("#nav-bar__gamesbox");o.insertAdjacentHTML("beforeend","<hr id='nav-bar__search--hr'  tabindex='0' style='display: none;'>"),l.insertAdjacentHTML("beforeend",'<span style="cursor: default;height: 50px; display: block; float: right; padding-left: 30px;"></span>'),l.insertAdjacentHTML("beforeend","<hr id='nav-bar__games--hr2D'  tabindex='0' style='display: none;'>"),l.insertAdjacentHTML("beforeend","<hr id='nav-bar__games--hr3D'  tabindex='0' style='display: none;'>");for(let[f,h]of this.techPages.entries()){if(f.dim==="N/A")continue;let g=new DOMParser,d="";d=d.concat('<figure class="nav-bar__games--wrapper" data-document="'),d=d.concat(f.document),d=d.concat('" data-dim="'),d=d.concat(f.dim),d=d.concat('" tabindex="0" data-section="'),d=d.concat(f.section),d=d.concat('"><figcaption>'),d=d.concat(f.document),d=d.concat('</figcaption><img class="nav-bar__search--results" src="media/'),d=d.concat(f.document.toLowerCase().trim()),d=d.concat("/"),d=d.concat(f.document.trim()),d=d.concat('ICONs3.webp" alt="'),d=d.concat(f.section),d=d.concat('"></img></figure>'),f.dim==="2D"?document.querySelector("#nav-bar__games--hr2D").insertAdjacentHTML("beforebegin",d):document.querySelector("#nav-bar__games--hr3D").insertAdjacentHTML("beforebegin",d),d="",document.querySelectorAll("figure.nav-bar__games--wrapper").forEach(_=>{this.addPageChangeEvent(_)});let v=g.parseFromString(h,"text/html");v.querySelectorAll("h1").forEach(_=>{n=n.concat('<div data-document="'),n=n.concat(f.document),n=n.concat('" class="nav-bar__search--results button__redirect" tabindex="0" data-section="'),n=n.concat(f.section),n=n.concat('" data-redirect="#'),n=n.concat(_.id),n=n.concat('"><b>'),n=n.concat(f.section),n=n.concat("</b> </div>")});let A=v.querySelectorAll("h1, h2, h3, h4"),E=null;A.forEach(_=>{_.tagName==="H1"?E=null:(_.tagName==="H2"&&(E=_.id),(E==="glitches"||E==="techniques")&&_.id!=="general-techniques"&&(r=r.concat('<div data-document="'),r=r.concat(f.document),r=r.concat('" class="nav-bar__search--results button__redirect" tabindex="0" data-section="'),r=r.concat(f.section),r=r.concat('" data-redirect="#'),r=r.concat(_.id),r=r.concat('"><b>'),r=r.concat(f.document),r=r.concat("</b> <br>&nbsp;&nbsp;&nbsp;"),r=r.concat(_.textContent),r=r.concat("</div>")))})}let p=n.concat(r);o.insertAdjacentHTML("beforeend",p),p="",document.querySelectorAll("div.nav-bar__search--results").forEach(f=>{this.addPageChangeEvent(f)})}revealID(n){let r=document.getElementById(n);if(r==null){let o=JSON.parse(sessionStorage.getItem("h2Collection")),l=0;for(let c=0;c<o.length;c++)if(o[c][2].includes(`id="${n}"`)){l=c;break}document.querySelectorAll(".content__selectorbox--item")[l].click()}r=document.getElementById(n);let t=null,i=!1;r.tagName=="H4"?r.dataset.open!=null&&(t=r.dataset.open.split(" "),i=!0):r.parentNode.className.includes("content__")?r.parentNode.firstElementChild.dataset.open!=null&&(t=r.parentNode.firstElementChild.dataset.open.split(" "),i=!0):r.parentNode.parentNode.firstElementChild.dataset.open!=null&&(t=r.parentNode.parentNode.firstElementChild.dataset.open.split(" "),i=!0),i&&t.forEach(o=>{let l=document.getElementsByClassName(o);for(let c of l)c.hidden&&(c.hidden=!1,c.parentNode.className.includes("content__")?c.parentNode.firstElementChild.firstChild.innerHTML=="expand_more"?c.parentNode.firstElementChild.firstChild.innerHTML="expand_less":c.parentNode.firstElementChild.firstChild.innerHTML="remove":c.parentNode.parentNode.firstElementChild.firstChild.innerHTML=="expand_more"?c.parentNode.parentNode.firstElementChild.firstChild.innerHTML="expand_less":c.parentNode.parentNode.firstElementChild.firstChild.innerHTML="remove")})}};var U=class{constructor(n){this.headings=[],this.ticking=!1,this.currentSection=null,this.inTechPage=!1,this.helperObj=n,document.addEventListener("scroll",r=>{!this.ticking&&this.inTechPage&&(window.requestAnimationFrame(()=>{this.highlightTOC(),this.ticking=!1}),this.ticking=!0)})}updateStatus(n,r){this.currentSection=n,this.inTechPage=r}clearHeadings(){this.headings=[]}createTOC(n){let r=document.getElementById("content__tocborder"),t=null,i=null,o=null,l=null,c=null;if(r==null){let f=document.querySelector(".content__selectorbox--item.selected");if(f==null)return;c=f.textContent,i=document.getElementById("content"),r=document.createElement("div"),r.id="content__tocborder",i.appendChild(r),o=document.createElement("div"),o.id="content__toc",l=document.createElement("div"),l.id="content__tocmobile",r.appendChild(o),t=document.getElementById("content__selector"),r.style.top=t.offsetTop+150+"px";let h=document.createElement("span");h.className="material-symbols-rounded",h.id="content__tocicon",h.innerHTML="menu_book",h.style.display="none";let g=document.getElementById("section-container__div");g.appendChild(h);let d=document.createElement("div");d.id="content__tocicon--box",d.style.width="0px",d.tabIndex="0",g.appendChild(d),d.appendChild(l),this.helperObj.dragScrollElement("#"+d.id,1),this.helperObj.dragScrollElement("#"+o.id,1),h.addEventListener("click",v=>{let b=document.getElementById("content__tocicon--box");var A=300;document.body.clientWidth>=480&&(A=360),b.style.width=A+"px",b.className="fadein",d.display="block",b.focus()}),d.addEventListener("focusout",v=>{let b=document.getElementById("content__tocicon--box");v.stopPropagation(),b.className="fadeout"})}else t=document.getElementById("content__selector"),i=document.getElementById("content"),o=document.getElementById("content__toc"),l=document.getElementById("content__tocmobile"),c=document.querySelector(".content__selectorbox--item.selected").textContent;r.style.height="calc(100% - "+(t.offsetTop+150)+"px)";let s="<p>\u2014 On this section:</p><hr>";o.innerHTML=s,l.innerHTML=s,this.headings=[];let a=document.querySelectorAll("#content__currenth2 h3"),u=document.querySelectorAll("#content__currenth2 h4");for(let f of a)this.headings.push({name:f.textContent,clientY:f.getBoundingClientRect().top,isH3:!0,obj:f,id:f.id});for(let f of u)this.headings.push({name:f.textContent,clientY:f.getBoundingClientRect().top,isH3:!1,obj:f,id:f.id});this.headings.sort((f,h)=>f.clientY-h.clientY);let p="";for(let f of this.headings){let h='<div data-document="'+n+'" data-section="'+this.currentSection+'" data-redirect="#'+f.id+'" class="content__toc--search button__redirect" style="display: block;';f.isH3||(h=h.concat("padding-left: 30px; font-size: 13px; padding-top: 6px; padding-bottom: 6px;")),h=h.concat('">'+f.name),h=h.concat("</div>"),p=p.concat(h)}o.innerHTML=s.concat(p),l.innerHTML=s.concat(p)}highlightTOC(){let n=null,r=0,t=0;for(let i=0;i<this.headings.length;i++)if(this.headings[i].isH3)if(this.headings[i].obj.getBoundingClientRect().top<=Math.max(window.innerHeight*.25,150)||n==null)n=this.headings[i],r=i;else{t=i;break}for(let i=0;i<this.headings.length;i++)if(!this.headings[i].isH3)if(i>=r&&i<=t||t==0&&r>0&&i>=r||t==0&&r==0){let o=document.querySelectorAll('.content__toc--search[data-redirect="#'+this.headings[i].id+'"]');for(let l of o)l.style.display="block"}else{let o=document.querySelectorAll('.content__toc--search[data-redirect="#'+this.headings[i].id+'"]');for(let l of o)l.style.display="none"}if(n!=null){let i=document.querySelectorAll(".content__toc--search.active");for(let l of i)l.classList.remove("active");let o=document.querySelectorAll('.content__toc--search[data-redirect="#'+n.id+'"]');for(let l of o)l.classList.add("active")}}clearSectionTOC(){let n=document.getElementById("content__tocicon");n?.remove();let r=document.getElementById("content__tocicon--box");r?.remove()}};var Y=class{constructor(n){this.colorCollection=[["yellow","#f8d959"],["pink","#fe796f"],["teal","#45c9c9"],["green","#58f15b"],["red","#e74a41"],["blue","#205aaa"],["lavender","#c8a2b0"]],this.mouseDown=!1,this.h2Collection=[],this.startX=0,this.scrollLeft=0,this.helperObj=n}returnH2Collection(){return this.h2Collection}collapseHeadings(n){this.h2Collection=[];let r=null,t=null,i=null,o="",l="",c="",s=!0,a=n.innerHTML.split(`
+`),u=document.getElementById("content");for(let h=0;h<a.length;h++)if(a[h].includes('h4 id="')){r&&(l=l+`</div></div>
+`),r=a[h].substring(a[h].indexOf('h4 id="')+7),r=r.substring(0,r.indexOf('"')),l=l+`<div class="content__h4">
+`;let g=a[h].indexOf('h4 id="')+7+r.length+1,d=' data-open="'+r+" "+t+" "+i+'"',v=a[h].substr(0,g)+d+a[h].substr(g);l=l+v+`
+`,l=l+'<div class="'+r+`">
+`}else a[h].includes('h3 id="')?(r&&(l=l+`</div></div>
+`,r=null),t&&(l=l+`<hr class="content__h3--divider" draggable="false" /></div></div>
+`),t=a[h].substring(a[h].indexOf('h3 id="')+7),t=t.substring(0,t.indexOf('"')),l=l+`<div class="content__h3">
+`,l=l+'<button class="content__collapse" data-open="'+t+" "+i+'"><span class="material-symbols-rounded">expand_circle_up</span></button>',l=l+a[h]+`
+`,l=l+'<div class="'+t+`">
+`):a[h].includes('h2 id="')?(s=!1,r&&(l=l+`</div></div>
+`,r=null),t&&(l=l+`</div></div>
+`,t=null),i&&(l=l+`</div></div>
+`,this.h2Collection.push([i,o,l]),l=""),i=a[h].substring(a[h].indexOf('h2 id="')+7),i=i.substring(0,i.indexOf('"')),o=a[h].substring(a[h].indexOf('">')+2),o=o.substring(0,o.indexOf("</h2>")),l=l+'<div class="content__h2" id="'+i+`">
+`,l=l+'<div class="'+i+`">
+`):s?c=c+a[h]+`
+`:l=l+a[h]+`
+`;r&&(l=l+`</div></div>
+`),t&&(l=l+`</div></div>
+`),i&&(l=l+`</div></div>
+`,this.h2Collection.push([i,o,l]),l="");var p='<div id="content__selector"><div id="content__selectorbox">';let f=0;for(;f<this.h2Collection.length;){let h=this.h2Collection[f][0],g=this.h2Collection[f][1],d=this.colorCollection[f%this.colorCollection.length];p=p+'<div class="content__selectorbox--item" data-open="'+h+'" data-highlight="'+d[1]+'">',p=p+g+"</div>",this.h2Collection[f].push(d),f++}sessionStorage.setItem("h2Collection",JSON.stringify(this.h2Collection)),p=p+'</div><hr id="content__selectorhr"></hr></div><div id="content__currenth2"></div>',this.h2Collection.length>0?n.innerHTML=c+p:n.innerHTML=c,this.helperObj.dragScrollElement("#content__selectorbox",0),u.style.visibility="visible"}collapseHeadingStyle(n){let r=n.currentTarget.dataset.open.split(" ");for(let t=0;t<r.length;t++){let i=document.getElementsByClassName(r[t]);for(let o of i)o.hidden?(o.hidden=!1,n.currentTarget.firstChild.innerHTML=="expand_circle_down"?n.currentTarget.firstChild.innerHTML="expand_circle_up":n.currentTarget.firstChild.innerHTML="remove"):t==0&&(o.hidden=!0,n.currentTarget.firstChild.innerHTML=="expand_circle_up"?n.currentTarget.firstChild.innerHTML="expand_circle_down":n.currentTarget.firstChild.innerHTML="add")}}};var Z=class{constructor(n){this.compileDirectives=this.compileDirectives.bind(this),this.sortTables=this.sortTables.bind(this),this.treatSpoilers=this.treatSpoilers.bind(this),this.styleImages=this.styleImages.bind(this),this.compileTags=this.compileTags.bind(this),this.helperObj=n}compileTags(){document.querySelectorAll(".tagging, .tagging-text").forEach(n=>{let r=n.dataset.tags,t=JSON.parse(r.replace(/'/g,'"'));if(t.todo){let i=document.createElement("span");i.className="material-symbols-rounded",i.style.color="goldenrod",i.style.marginRight="15px",i.style.cursor="help",i.textContent="error",n.appendChild(i),i.addEventListener("mouseover",()=>{this.helperObj.loadTooltip(i,"This section needs work, is not confirmed or needs testing.")}),i.addEventListener("mouseleave",()=>{this.helperObj.unloadTooltip(i)})}if(t.versions){let i="Versions: ";i=i.concat(t.versions),i=i.concat(".");let o=document.createElement("span");o.className="material-symbols-rounded",o.style.marginRight="15px",o.style.cursor="help",o.textContent="devices",n.appendChild(o),o.addEventListener("mouseover",()=>{this.helperObj.loadTooltip(o,i)}),o.addEventListener("mouseleave",()=>{this.helperObj.unloadTooltip(o)})}if(t.media){let o=t.media.includes(".webp")||t.media.includes(".png")||t.media.includes(".jpg")?"img":"video",l=(s,a,u,p,f)=>{let h=document.createElement("p");h.style.opacity="1";let g;if(s==="img")g=document.createElement("img");else if(s==="video"){g=document.createElement("video"),g.preload="metadata",g.controls=!0,g.muted=!0,g.loop=!0;let d=document.createElement("source");d.src=a,d.type="video/mp4",g.appendChild(d)}if(g.draggable=!1,g.width=u,g.height=p,g.style.order=2,g.style.width="80%",g.style.height="auto",g.style.maxWidth="640px",g.style.outline="none",g.src=a,h.appendChild(g),f){let d=document.createElement("figcaption");d.textContent="\xA0"+f+"\xA0",h.appendChild(d)}return h},c=(s,a,u)=>{let p=document.createElement("span");p.dataset.media=a,p.className="material-symbols-rounded",p.style.marginRight="15px",p.style.cursor="pointer",p.textContent=s==="img"?"imagesmode":"play_circle",n.appendChild(p),p.addEventListener("click",f=>{let h=s==="img"?"img":"video",g=document.querySelector(`p[data-media~="${f.target.dataset.media}"]`);if(g==null){let d=l(h,f.target.dataset.media,640,480,u);d.dataset.media=f.target.dataset.media,n.parentNode.insertBefore(d,n.nextSibling)}else g.remove()})};if(t.forcedmedia!==!1){let s=l(o,t.media,640,480,t.caption);n.parentNode.insertBefore(s,n.nextSibling)}else c(o,t.media,t.caption)}})}compileDirectives(){this.compileTags(),this.sortTables(),this.treatSpoilers(),this.styleImages()}sortTables(){let n=(t,i)=>t.children[i].innerText||t.children[i].textContent,r=(t,i)=>(o,l)=>((c,s)=>c!==""&&s!==""&&!isNaN(c)&&!isNaN(s)?c-s:c.toString().localeCompare(s))(n(i?o:l,t),n(i?l:o,t));document.querySelectorAll("th").forEach(t=>{t.innerHTML=t.innerHTML+'<span class="material-symbols-rounded">swap_vert</span>',t.addEventListener("click",()=>{let o=t.closest("table").querySelector("tbody");Array.from(o.querySelectorAll("tr")).sort(r(Array.from(t.parentNode.children).indexOf(t),this.asc=!this.asc)).forEach(l=>o.appendChild(l))})})}treatSpoilers(){document.querySelectorAll(".spoiler").forEach(n=>{n.addEventListener("click",r=>{r.target.style.background="transparent"})})}styleImages(){let n=document.getElementsByTagName("img"),r=document.getElementById("content");for(let t=0;t<n.length;t++)r.contains(n[t])&&n[t].classList.add("content__figure")}};var X=class{constructor(){this.mouseDown=!1,this.startX=0,this.scrollLeft=0,this.startY=0,this.scrollTop=0,this.scrollButton=null,this.rootElement=null,this.scrollToTop=this.scrollToTop.bind(this),this.handleScroll=this.handleScroll.bind(this),this.dragScrollElement=this.dragScrollElement.bind(this),this.rotateFlag=!1,this.leaveCounter=0,this.currentVelocity=0}logoInit(){this.titleElement=document.querySelector("#nav-bar__title"),this.titleElement.addEventListener("mouseenter",this.startRotateLogo.bind(this)),this.titleElement.addEventListener("mouseleave",this.stopRotateLogo.bind(this))}addLogoVelocity(){!this.rotateFlag&&this.leaveCounter==0?(this.currentVelocity=4,this.leaveCounter=200,requestAnimationFrame(this.rotateLogo.bind(this))):(this.currentVelocity=4,this.leaveCounter=200)}startRotateLogo(){!this.rotateFlag&&this.leaveCounter==0?(this.rotateFlag=!0,requestAnimationFrame(this.rotateLogo.bind(this))):(this.rotateFlag=!0,this.leaveCounter=200)}stopRotateLogo(){this.rotateFlag=!1}rotateLogo(){let n=document.querySelector("#nav-bar__title__logo"),i=getComputedStyle(n,null).getPropertyValue("transform").split("(")[1].split(")")[0].split(","),o=i[0],l=i[1],c=Math.atan2(l,o)*(180/Math.PI);this.rotateFlag?this.currentVelocity<1.8&&(this.currentVelocity=this.currentVelocity+.06):this.currentVelocity>0&&(this.currentVelocity=this.currentVelocity-.03),this.currentVelocity<0&&(this.currentVelocity=0),c<0&&(c=c+360);let s=this.currentVelocity+c;n.style.transform="rotate("+s+"deg)",this.rotateFlag?(this.leaveCounter=200,requestAnimationFrame(this.rotateLogo.bind(this))):(this.leaveCounter--,this.currentVelocity<=0&&(this.leaveCounter=0),this.leaveCounter>0&&requestAnimationFrame(this.rotateLogo.bind(this)))}scrollInit(){this.scrollButton=document.querySelector("#scroll__button--to-top"),this.rootElement=document.documentElement,this.scrollButton.addEventListener("click",this.scrollToTop),window.addEventListener("scroll",this.handleScroll)}handleScroll(n){n.preventDefault();let r=this.rootElement.scrollHeight-this.rootElement.clientHeight,t=document.getElementById("content__selectorbox"),i=0;if(t!=null){let o=document.body.getBoundingClientRect(),l=t.getBoundingClientRect(),c=l.top-o.top;c>0&&(i=c)}(this.rootElement.scrollTop-i)/r>.05?this.scrollButton.classList.add("scroll__button--to-top--show"):this.scrollButton.classList.remove("scroll__button--to-top--show")}scrollToTop(){let n=document.getElementById("content__selectorbox"),r=0;if(n!=null){let t=document.body.getBoundingClientRect(),i=n.getBoundingClientRect(),o=i.top-t.top;o>0&&(r=o-100)}window.requestAnimationFrame(()=>{let t=navigator.userAgent.includes("Chrome")?"auto":"smooth";this.rootElement.scrollTo({top:r,behavior:t})})}loadTooltip(n,r){let t=document.createElement("tooltip");t.id="nav-bar__tooltip";let i=document.createTextNode(r);if(n!=null){t.innerHTML="",t.appendChild(i),document.getElementsByTagName("tooltip")[0]&&document.getElementsByTagName("tooltip")[0].remove();let o=n.getBoundingClientRect();t.style.top=o.bottom-10+"px",t.style.left=o.right+"px",document.body.appendChild(t)}}unloadTooltip(){document.getElementsByTagName("tooltip")[0]&&document.getElementsByTagName("tooltip")[0].remove()}dragScrollElement(n,r){let t=document.querySelector(n),i=a=>{this.mouseDown=!0,this.startX=a.pageX-t.offsetLeft,isNaN(this.startX)&&(this.startX=a.changedTouches[0].pageX-t.offsetLeft),this.scrollLeft=t.scrollLeft},o=a=>{this.mouseDown=!0,this.startY=a.pageY-t.offsetTop,isNaN(this.startY)&&(this.startY=a.changedTouches[0].pageY-t.offsetTop),this.scrollTop=t.scrollTop},l=a=>{if(a.preventDefault(),a.stopPropagation(),!this.mouseDown)return;let u=a.pageX-t.offsetLeft,p=u-this.startX;if(isNaN(u)){let h=a.changedTouches[0].pageX-t.offsetLeft-this.startX;t.scrollLeft=this.scrollLeft-h}else t.scrollLeft=this.scrollLeft-p},c=a=>{if(a.preventDefault(),a.stopPropagation(),!this.mouseDown)return;let u=a.pageY-t.offsetTop,p=u-this.startY;if(isNaN(u)){let h=a.changedTouches[0].pageY-t.offsetTop-this.startY;t.scrollTop=this.scrollTop-h}else t.scrollTop=this.scrollTop-p},s=a=>{this.mouseDown=!1};r==1?(t.addEventListener("mousemove",c,!1),t.addEventListener("mousedown",o,!1),t.addEventListener("mouseup",s,!1),t.addEventListener("mouseleave",s,!1),t.addEventListener("touchmove",c,!1),t.addEventListener("touchstart",o,!1),t.addEventListener("touchend",s,!1)):(t.addEventListener("mousemove",l,!1),t.addEventListener("mousedown",i,!1),t.addEventListener("mouseup",s,!1),t.addEventListener("mouseleave",s,!1),t.addEventListener("touchmove",l,!1),t.addEventListener("touchstart",i,!1),t.addEventListener("touchend",s,!1))}};var O=null,y=null,L=new X,ce=new Y(L),gr=new Z(L),we=new $,D=new Map,vr=[{document:"./README",section:"README",dim:"N/A"},{document:"./STYLING",section:"Doucment Syling",dim:"N/A"},{document:"./CONTRIBUTING",section:"How to Contribute",dim:"N/A"},{document:"TODPS2",section:"Tales of Destiny",dim:"2D"},{document:"TOL",section:"Tales of Legendia",dim:"2D"},{document:"TOA",section:"Tales of Arise",dim:"3D"},{document:"TOV",section:"Tales of Vesperia",dim:"3D"},{document:"TOTA",section:"Tales of the Abyss",dim:"3D"},{document:"TOX2",section:"Tales of Xillia 2",dim:"3D"},{document:"TOZ",section:"Tales of Zestiria",dim:"3D"}],br=vr.map(e=>e.document.includes("./")?we.parseGFM(e.document).then(n=>{D.set(e,n)}):we.parseGFM("./tech/"+e.document).then(n=>{D.set(e,n)})),q=null;document.readyState==="loading"?document.addEventListener("DOMContentLoaded",Le):Le();function Le(){document.querySelectorAll("div#nav-bar__title, p.footer-container__help--links").forEach(e=>{se(e)}),Promise.all(br).then(()=>{O=new V(D,se),O.initSearchIcons(),y=new U(L),L.scrollInit(),L.logoInit()})}function se(e){e.removeEventListener("click",qe),e.addEventListener("click",qe)}function qe(e){let n=document.getElementById("section-container__text"),r=document.getElementById("content");O.clearFunction(),L.addLogoVelocity(),e.currentTarget.dataset.document=="HOME"?_r(n,r,e):e.currentTarget.dataset.document.includes("./")?xr(n,r,e):kr(n,r,e)}function _r(e,n,r){q=r.currentTarget.dataset.document,y.clearHeadings(),y.updateStatus("HOME",!1),y.clearSectionTOC(),n.innerHTML="UNDER CONSTRUCTION",e.innerHTML="HOME",document.querySelector("#nav-bar").scrollIntoView({behavior:"smooth"})}function xr(e,n,r){q=r.currentTarget.dataset.document,y.clearHeadings(),y.updateStatus(r.currentTarget.dataset.section,!1),y.clearSectionTOC(),e.innerHTML=r.currentTarget.dataset.section;let t=r.currentTarget.dataset.document,i=null;for(let[o,l]of D.entries())if(o.document===t){i=l;break}n.innerHTML=i,Be(),document.querySelector("#nav-bar").scrollIntoView({behavior:"smooth"})}function kr(e,n,r){y.updateStatus(r.currentTarget.dataset.section,!0),e.innerHTML=r.currentTarget.dataset.section;let t=q;q=r.currentTarget.dataset.document;let i=r.currentTarget.dataset;if(q==t)i.redirect!=null?(O.revealID(i.redirect.substring(1)),document.querySelector(i.redirect).scrollIntoView({behavior:"smooth"})):document.querySelector("#nav-bar").scrollIntoView({behavior:"smooth"});else{y.clearHeadings(),y.clearSectionTOC();let o=r.currentTarget.dataset.document,l=null;for(let[c,s]of D.entries())if(c.document===o){l=s;break}n.style.visibility="hidden",n.innerHTML=l,ce.collapseHeadings(n),yr(),i.redirect!=null?(O.revealID(i.redirect.substring(1)),setTimeout(()=>{document.querySelector(i.redirect).scrollIntoView({behavior:"smooth"})},250)):document.querySelector("#nav-bar").scrollIntoView({behavior:"smooth"})}}function Be(){y.createTOC(q),document.querySelectorAll(".content__toc--search").forEach(e=>{se(e)}),gr.compileDirectives()}function yr(){document.querySelectorAll(".content__selectorbox--item").forEach(e=>{e.style.setProperty("--highlight-color",e.dataset.highlight),e.addEventListener("click",n=>{document.querySelectorAll(".content__selectorbox--item").forEach(t=>{t.className="content__selectorbox--item"}),n.currentTarget.className="content__selectorbox--item selected";let r=ce.returnH2Collection().filter(t=>t[0]==n.currentTarget.dataset.open);document.getElementById("content__currenth2").innerHTML=r[0][2],document.querySelectorAll(".content__collapse").forEach(t=>{t.addEventListener("click",ce.collapseHeadingStyle)}),Be(),y.highlightTOC()})}),document.querySelectorAll(".content__selectorbox--item")[0].click()}export{se as addPageChangeEvent,Be as updatePage};
+//# sourceMappingURL=main.js.map
