--- docs/scripts/page.js
@@ -1,59 +1,0 @@
-(()=>{function Ce(){document.querySelectorAll(".nav-bar__tab-bar--links").forEach(e=>{e.addEventListener("mouseover",()=>{j(e,e.dataset.section)}),e.addEventListener("mouseleave",()=>{$()})})}document.readyState!=="loading"?Ce():document.addEventListener("DOMContentLoaded",Ce);function j(e,n){let r=document.createElement("tooltip");r.id="nav-bar__tooltip";let t=document.createTextNode(n);if(e!=null){r.innerHTML="",r.appendChild(t),document.getElementsByTagName("tooltip")[0]&&document.getElementsByTagName("tooltip")[0].remove();let o=e.getBoundingClientRect();r.style.top=o.bottom-10+"px",r.style.left=o.right+"px",document.body.appendChild(r)}}function $(){document.getElementsByTagName("tooltip")[0]&&document.getElementsByTagName("tooltip")[0].remove()}var Ee=[["yellow","#f8d959"],["pink","#fe796f"],["teal","#45c9c9"],["green","#58f15b"],["red","#e74a41"],["blue","#205aaa"],["lavender","#c8a2b0"]];function Ae(){var e=function(r,t){return r.children[t].innerText||r.children[t].textContent},n=function(r,t){return function(o,l){return function(c,i){return c!==""&&i!==""&&!isNaN(c)&&!isNaN(i)?c-i:c.toString().localeCompare(i)}(e(t?o:l,r),e(t?l:o,r))}};document.querySelectorAll("th").forEach(r=>{r.innerHTML=r.innerHTML+'<span class="material-symbols-rounded">swap_vert</span>',r.addEventListener("click",()=>{let o=r.closest("table").querySelector("tbody");Array.from(o.querySelectorAll("tr")).sort(n(Array.from(r.parentNode.children).indexOf(r),this.asc=!this.asc)).forEach(l=>o.appendChild(l))})})}function Te(){document.querySelectorAll(".spoiler").forEach(e=>{e.addEventListener("click",n=>{n.target.style.background="transparent"})})}function Se(){for(var e=document.getElementsByTagName("img"),n=document.getElementById("content"),r,r=0;r<e.length;r++)n.contains(e[r])&&e[r].classList.add("content__figure")}var B=!1,C=[],R,re;function Me(e){C=[];let n=null,r=null,t=null,o="",l="",c=!0,i=e.innerHTML.split(`
-`),a=document.getElementById("content");for(let f=0;f<i.length;f++)if(i[f].includes('h4 id="')){n&&(o=o+`</div></div>
-`),n=i[f].substring(i[f].indexOf('h4 id="')+7),n=n.substring(0,n.indexOf('"')),o=o+`<div class="content__h4">
-`;let d=i[f].indexOf('h4 id="')+7+n.length+1,m=' data-open="'+n+" "+r+" "+t+'"',g=i[f].substr(0,d)+m+i[f].substr(d);o=o+g+`
-`,o=o+'<div class="'+n+`">
-`}else i[f].includes('h3 id="')?(n&&(o=o+`</div></div>
-`,n=null),r&&(o=o+`<hr class="content__h3--divider" draggable="false" /></div></div>
-`),r=i[f].substring(i[f].indexOf('h3 id="')+7),r=r.substring(0,r.indexOf('"')),o=o+`<div class="content__h3">
-`,o=o+'<button class="content__collapse" data-open="'+r+" "+t+'"><span class="material-symbols-rounded">expand_circle_up</span></button>',o=o+i[f]+`
-`,o=o+'<div class="'+r+`">
-`):i[f].includes('h2 id="')?(c=!1,n&&(o=o+`</div></div>
-`,n=null),r&&(o=o+`</div></div>
-`,r=null),t&&(o=o+`</div></div>
-`,C.push([t,currentH2Text,o]),o=""),t=i[f].substring(i[f].indexOf('h2 id="')+7),t=t.substring(0,t.indexOf('"')),currentH2Text=i[f].substring(i[f].indexOf('">')+2),currentH2Text=currentH2Text.substring(0,currentH2Text.indexOf("</h2>")),o=o+'<div class="content__h2" id="'+t+`">
-`,o=o+'<div class="'+t+`">
-`):c?l=l+i[f]+`
-`:o=o+i[f]+`
-`;n&&(o=o+`</div></div>
-`),r&&(o=o+`</div></div>
-`),t&&(o=o+`</div></div>
-`,C.push([t,currentH2Text,o]),o="");var s='<div id="content__selector"><div id="content__selectorbox">';let u=0;for(;u<C.length;)h2Tag=C[u][0],h2Text=C[u][1],h2Color=Ee[u%Ee.length],s=s+'<div class="content__selectorbox--item" data-open="'+h2Tag+'" data-highlight="'+h2Color[1]+'">',s=s+h2Text+"</div>",u++;s=s+'</div><hr id="content__selectorhr"></hr></div><div id="content__currenth2"></div>',C.length>0?e.innerHTML=l+s:e.innerHTML=l,U("#content__selectorbox",0),a.style.visibility="visible"}function U(e,n){let r=document.querySelector(e),t=function(a){B=!0,R=a.pageX-r.offsetLeft,isNaN(R)&&(R=a.changedTouches[0].pageX-r.offsetLeft),re=r.scrollLeft},o=function(a){B=!0,startY=a.pageY-r.offsetTop,isNaN(startY)&&(startY=a.changedTouches[0].pageY-r.offsetTop),scrollTop=r.scrollTop},l=function(a){if(a.preventDefault(),a.stopPropagation(),!B)return;let s=a.pageX-r.offsetLeft,u=s-R;if(isNaN(s)){let d=a.changedTouches[0].pageX-r.offsetLeft-R;r.scrollLeft=re-d}else r.scrollLeft=re-u},c=function(a){if(a.preventDefault(),a.stopPropagation(),!B)return;let s=a.pageY-r.offsetTop,u=s-startY;if(isNaN(s)){let d=a.changedTouches[0].pageY-r.offsetTop-startY;r.scrollTop=scrollTop-d}else r.scrollTop=scrollTop-u},i=function(a){B=!1};n==1?(r.addEventListener("mousemove",c,!1),r.addEventListener("mousedown",o,!1),r.addEventListener("mouseup",i,!1),r.addEventListener("mouseleave",i,!1),r.addEventListener("touchmove",c,!1),r.addEventListener("touchstart",o,!1),r.addEventListener("touchend",i,!1)):(r.addEventListener("mousemove",l,!1),r.addEventListener("mousedown",t,!1),r.addEventListener("mouseup",i,!1),r.addEventListener("mouseleave",i,!1),r.addEventListener("touchmove",l,!1),r.addEventListener("touchstart",t,!1),r.addEventListener("touchend",i,!1))}function we(e){let n=e.currentTarget.dataset.open.split(" ");for(let r=0;r<n.length;r++){targetList=document.getElementsByClassName(n[r]);for(let t of targetList)t.hidden?(t.hidden=!1,e.currentTarget.firstChild.innerHTML=="expand_circle_down"?e.currentTarget.firstChild.innerHTML="expand_circle_up":e.currentTarget.firstChild.innerHTML="remove"):r==0&&(t.hidden=!0,e.currentTarget.firstChild.innerHTML=="expand_circle_up"?e.currentTarget.firstChild.innerHTML="expand_circle_down":e.currentTarget.firstChild.innerHTML="add")}}var V;function Oe(e){return V=V||document.createElement("textarea"),V.innerHTML="&"+e+";",V.value}var en=Object.prototype.hasOwnProperty;function nn(e,n){return e?en.call(e,n):!1}function Pe(e){var n=[].slice.call(arguments,1);return n.forEach(function(r){if(r){if(typeof r!="object")throw new TypeError(r+"must be object");Object.keys(r).forEach(function(t){e[t]=r[t]})}}),e}var rn=/\\([\\!"#$%&'()*+,.\/:;<=>?@[\]^_`{|}~-])/g;function D(e){return e.indexOf("\\")<0?e:e.replace(rn,"$1")}function Fe(e){return!(e>=55296&&e<=57343||e>=64976&&e<=65007||(e&65535)===65535||(e&65535)===65534||e>=0&&e<=8||e===11||e>=14&&e<=31||e>=127&&e<=159||e>1114111)}function ae(e){if(e>65535){e-=65536;var n=55296+(e>>10),r=56320+(e&1023);return String.fromCharCode(n,r)}return String.fromCharCode(e)}var on=/&([a-z#][a-z0-9]{1,31});/gi,tn=/^#((?:x[a-f0-9]{1,8}|[0-9]{1,8}))/i;function ln(e,n){var r=0,t=Oe(n);return n!==t?t:n.charCodeAt(0)===35&&tn.test(n)&&(r=n[1].toLowerCase()==="x"?parseInt(n.slice(2),16):parseInt(n.slice(1),10),Fe(r))?ae(r):e}function M(e){return e.indexOf("&")<0?e:e.replace(on,ln)}var cn=/[&<>"]/,an=/[&<>"]/g,sn={"&":"&amp;","<":"&lt;",">":"&gt;",'"':"&quot;"};function un(e){return sn[e]}function k(e){return cn.test(e)?e.replace(an,un):e}var h={};h.blockquote_open=function(){return`<blockquote>
-`};h.blockquote_close=function(e,n){return"</blockquote>"+w(e,n)};h.code=function(e,n){return e[n].block?"<pre><code>"+k(e[n].content)+"</code></pre>"+w(e,n):"<code>"+k(e[n].content)+"</code>"};h.fence=function(e,n,r,t,o){var l=e[n],c="",i=r.langPrefix,a="",s,u,f;if(l.params){if(s=l.params.split(/\s+/g),u=s.join(" "),nn(o.rules.fence_custom,s[0]))return o.rules.fence_custom[s[0]](e,n,r,t,o);a=k(M(D(u))),c=' class="'+i+a+'"'}return r.highlight?f=r.highlight.apply(r.highlight,[l.content].concat(s))||k(l.content):f=k(l.content),"<pre><code"+c+">"+f+"</code></pre>"+w(e,n)};h.fence_custom={};h.heading_open=function(e,n){return"<h"+e[n].hLevel+">"};h.heading_close=function(e,n){return"</h"+e[n].hLevel+`>
-`};h.hr=function(e,n,r){return(r.xhtmlOut?"<hr />":"<hr>")+w(e,n)};h.bullet_list_open=function(){return`<ul>
-`};h.bullet_list_close=function(e,n){return"</ul>"+w(e,n)};h.list_item_open=function(){return"<li>"};h.list_item_close=function(){return`</li>
-`};h.ordered_list_open=function(e,n){var r=e[n],t=r.order>1?' start="'+r.order+'"':"";return"<ol"+t+`>
-`};h.ordered_list_close=function(e,n){return"</ol>"+w(e,n)};h.paragraph_open=function(e,n){return e[n].tight?"":"<p>"};h.paragraph_close=function(e,n){var r=!(e[n].tight&&n&&e[n-1].type==="inline"&&!e[n-1].content);return(e[n].tight?"":"</p>")+(r?w(e,n):"")};h.link_open=function(e,n,r){var t=e[n].title?' title="'+k(M(e[n].title))+'"':"",o=r.linkTarget?' target="'+r.linkTarget+'"':"";return'<a href="'+k(e[n].href)+'"'+t+o+">"};h.link_close=function(){return"</a>"};h.image=function(e,n,r){var t=' src="'+k(e[n].src)+'"',o=e[n].title?' title="'+k(M(e[n].title))+'"':"",l=' alt="'+(e[n].alt?k(M(D(e[n].alt))):"")+'"',c=r.xhtmlOut?" /":"";return"<img"+t+l+o+c+">"};h.table_open=function(){return`<table>
-`};h.table_close=function(){return`</table>
-`};h.thead_open=function(){return`<thead>
-`};h.thead_close=function(){return`</thead>
-`};h.tbody_open=function(){return`<tbody>
-`};h.tbody_close=function(){return`</tbody>
-`};h.tr_open=function(){return"<tr>"};h.tr_close=function(){return`</tr>
-`};h.th_open=function(e,n){var r=e[n];return"<th"+(r.align?' style="text-align:'+r.align+'"':"")+">"};h.th_close=function(){return"</th>"};h.td_open=function(e,n){var r=e[n];return"<td"+(r.align?' style="text-align:'+r.align+'"':"")+">"};h.td_close=function(){return"</td>"};h.strong_open=function(){return"<strong>"};h.strong_close=function(){return"</strong>"};h.em_open=function(){return"<em>"};h.em_close=function(){return"</em>"};h.del_open=function(){return"<del>"};h.del_close=function(){return"</del>"};h.ins_open=function(){return"<ins>"};h.ins_close=function(){return"</ins>"};h.mark_open=function(){return"<mark>"};h.mark_close=function(){return"</mark>"};h.sub=function(e,n){return"<sub>"+k(e[n].content)+"</sub>"};h.sup=function(e,n){return"<sup>"+k(e[n].content)+"</sup>"};h.hardbreak=function(e,n,r){return r.xhtmlOut?`<br />
-`:`<br>
-`};h.softbreak=function(e,n,r){return r.breaks?r.xhtmlOut?`<br />
-`:`<br>
-`:`
-`};h.text=function(e,n){return k(e[n].content)};h.htmlblock=function(e,n){return e[n].content};h.htmltag=function(e,n){return e[n].content};h.abbr_open=function(e,n){return'<abbr title="'+k(M(e[n].title))+'">'};h.abbr_close=function(){return"</abbr>"};h.footnote_ref=function(e,n){var r=Number(e[n].id+1).toString(),t="fnref"+r;return e[n].subId>0&&(t+=":"+e[n].subId),'<sup class="footnote-ref"><a href="#fn'+r+'" id="'+t+'">['+r+"]</a></sup>"};h.footnote_block_open=function(e,n,r){var t=r.xhtmlOut?`<hr class="footnotes-sep" />
-`:`<hr class="footnotes-sep">
-`;return t+`<section class="footnotes">
-<ol class="footnotes-list">
-`};h.footnote_block_close=function(){return`</ol>
-</section>
-`};h.footnote_open=function(e,n){var r=Number(e[n].id+1).toString();return'<li id="fn'+r+'"  class="footnote-item">'};h.footnote_close=function(){return`</li>
-`};h.footnote_anchor=function(e,n){var r=Number(e[n].id+1).toString(),t="fnref"+r;return e[n].subId>0&&(t+=":"+e[n].subId),' <a href="#'+t+'" class="footnote-backref">\u21A9</a>'};h.dl_open=function(){return`<dl>
-`};h.dt_open=function(){return"<dt>"};h.dd_open=function(){return"<dd>"};h.dl_close=function(){return`</dl>
-`};h.dt_close=function(){return`</dt>
-`};h.dd_close=function(){return`</dd>
-`};function ze(e,n){return++n>=e.length-2?n:e[n].type==="paragraph_open"&&e[n].tight&&e[n+1].type==="inline"&&e[n+1].content.length===0&&e[n+2].type==="paragraph_close"&&e[n+2].tight?ze(e,n+2):n}var w=h.getBreak=function(n,r){return r=ze(n,r),r<n.length&&n[r].type==="list_item_close"?"":`
-`};function ue(){this.rules=Pe({},h),this.getBreak=h.getBreak}ue.prototype.renderInline=function(e,n,r){for(var t=this.rules,o=e.length,l=0,c="";o--;)c+=t[e[l].type](e,l++,n,r,this);return c};ue.prototype.render=function(e,n,r){for(var t=this.rules,o=e.length,l=-1,c="";++l<o;)e[l].type==="inline"?c+=this.renderInline(e[l].children,n,r):c+=t[e[l].type](e,l,n,r,this);return c};function y(){this.__rules__=[],this.__cache__=null}y.prototype.__find__=function(e){for(var n=this.__rules__.length,r=-1;n--;)if(this.__rules__[++r].name===e)return r;return-1};y.prototype.__compile__=function(){var e=this,n=[""];e.__rules__.forEach(function(r){r.enabled&&r.alt.forEach(function(t){n.indexOf(t)<0&&n.push(t)})}),e.__cache__={},n.forEach(function(r){e.__cache__[r]=[],e.__rules__.forEach(function(t){t.enabled&&(r&&t.alt.indexOf(r)<0||e.__cache__[r].push(t.fn))})})};y.prototype.at=function(e,n,r){var t=this.__find__(e),o=r||{};if(t===-1)throw new Error("Parser rule not found: "+e);this.__rules__[t].fn=n,this.__rules__[t].alt=o.alt||[],this.__cache__=null};y.prototype.before=function(e,n,r,t){var o=this.__find__(e),l=t||{};if(o===-1)throw new Error("Parser rule not found: "+e);this.__rules__.splice(o,0,{name:n,enabled:!0,fn:r,alt:l.alt||[]}),this.__cache__=null};y.prototype.after=function(e,n,r,t){var o=this.__find__(e),l=t||{};if(o===-1)throw new Error("Parser rule not found: "+e);this.__rules__.splice(o+1,0,{name:n,enabled:!0,fn:r,alt:l.alt||[]}),this.__cache__=null};y.prototype.push=function(e,n,r){var t=r||{};this.__rules__.push({name:e,enabled:!0,fn:n,alt:t.alt||[]}),this.__cache__=null};y.prototype.enable=function(e,n){e=Array.isArray(e)?e:[e],n&&this.__rules__.forEach(function(r){r.enabled=!1}),e.forEach(function(r){var t=this.__find__(r);if(t<0)throw new Error("Rules manager: invalid rule name "+r);this.__rules__[t].enabled=!0},this),this.__cache__=null};y.prototype.disable=function(e){e=Array.isArray(e)?e:[e],e.forEach(function(n){var r=this.__find__(n);if(r<0)throw new Error("Rules manager: invalid rule name "+n);this.__rules__[r].enabled=!1},this),this.__cache__=null};y.prototype.getRules=function(e){return this.__cache__===null&&this.__compile__(),this.__cache__[e]||[]};function fn(e){e.inlineMode?e.tokens.push({type:"inline",content:e.src.replace(/\n/g," ").trim(),level:0,lines:[0,1],children:[]}):e.block.parse(e.src,e.options,e.env,e.tokens)}function I(e,n,r,t,o){this.src=e,this.env=t,this.options=r,this.parser=n,this.tokens=o,this.pos=0,this.posMax=this.src.length,this.level=0,this.pending="",this.pendingLevel=0,this.cache=[],this.isInLabel=!1,this.linkLevel=0,this.linkContent="",this.labelUnmatchedScopes=0}I.prototype.pushPending=function(){this.tokens.push({type:"text",content:this.pending,level:this.pendingLevel}),this.pending=""};I.prototype.push=function(e){this.pending&&this.pushPending(),this.tokens.push(e),this.pendingLevel=this.level};I.prototype.cacheSet=function(e,n){for(var r=this.cache.length;r<=e;r++)this.cache.push(0);this.cache[e]=n};I.prototype.cacheGet=function(e){return e<this.cache.length?this.cache[e]:0};function O(e,n){var r,t,o,l=-1,c=e.posMax,i=e.pos,a=e.isInLabel;if(e.isInLabel)return-1;if(e.labelUnmatchedScopes)return e.labelUnmatchedScopes--,-1;for(e.pos=n+1,e.isInLabel=!0,r=1;e.pos<c;){if(o=e.src.charCodeAt(e.pos),o===91)r++;else if(o===93&&(r--,r===0)){t=!0;break}e.parser.skipToken(e)}return t?(l=e.pos,e.labelUnmatchedScopes=0):e.labelUnmatchedScopes=r-1,e.pos=i,e.isInLabel=a,l}function pn(e,n,r,t){var o,l,c,i,a,s;if(e.charCodeAt(0)!==42||e.charCodeAt(1)!==91||e.indexOf("]:")===-1||(o=new I(e,n,r,t,[]),l=O(o,1),l<0||e.charCodeAt(l+1)!==58))return-1;for(i=o.posMax,c=l+2;c<i&&o.src.charCodeAt(c)!==10;c++);return a=e.slice(2,l),s=e.slice(l+2,c).trim(),s.length===0?-1:(t.abbreviations||(t.abbreviations={}),typeof t.abbreviations[":"+a]>"u"&&(t.abbreviations[":"+a]=s),c)}function dn(e){var n=e.tokens,r,t,o,l;if(!e.inlineMode){for(r=1,t=n.length-1;r<t;r++)if(n[r-1].type==="paragraph_open"&&n[r].type==="inline"&&n[r+1].type==="paragraph_close"){for(o=n[r].content;o.length&&(l=pn(o,e.inline,e.options,e.env),!(l<0));)o=o.slice(l).trim();n[r].content=o,o.length||(n[r-1].tight=!0,n[r+1].tight=!0)}}}function se(e){var n=M(e);try{n=decodeURI(n)}catch{}return encodeURI(n)}function je(e,n){var r,t,o,l=n,c=e.posMax;if(e.src.charCodeAt(n)===60){for(n++;n<c;){if(r=e.src.charCodeAt(n),r===10)return!1;if(r===62)return o=se(D(e.src.slice(l+1,n))),e.parser.validateLink(o)?(e.pos=n+1,e.linkContent=o,!0):!1;if(r===92&&n+1<c){n+=2;continue}n++}return!1}for(t=0;n<c&&(r=e.src.charCodeAt(n),!(r===32||r<32||r===127));){if(r===92&&n+1<c){n+=2;continue}if(r===40&&(t++,t>1)||r===41&&(t--,t<0))break;n++}return l===n||(o=D(e.src.slice(l,n)),!e.parser.validateLink(o))?!1:(e.linkContent=o,e.pos=n,!0)}function $e(e,n){var r,t=n,o=e.posMax,l=e.src.charCodeAt(n);if(l!==34&&l!==39&&l!==40)return!1;for(n++,l===40&&(l=41);n<o;){if(r=e.src.charCodeAt(n),r===l)return e.pos=n+1,e.linkContent=D(e.src.slice(t+1,n)),!0;if(r===92&&n+1<o){n+=2;continue}n++}return!1}function Ue(e){return e.trim().replace(/\s+/g," ").toUpperCase()}function hn(e,n,r,t){var o,l,c,i,a,s,u,f,d;if(e.charCodeAt(0)!==91||e.indexOf("]:")===-1||(o=new I(e,n,r,t,[]),l=O(o,0),l<0||e.charCodeAt(l+1)!==58))return-1;for(i=o.posMax,c=l+2;c<i&&(a=o.src.charCodeAt(c),!(a!==32&&a!==10));c++);if(!je(o,c))return-1;for(u=o.linkContent,c=o.pos,s=c,c=c+1;c<i&&(a=o.src.charCodeAt(c),!(a!==32&&a!==10));c++);for(c<i&&s!==c&&$e(o,c)?(f=o.linkContent,c=o.pos):(f="",c=s);c<i&&o.src.charCodeAt(c)===32;)c++;return c<i&&o.src.charCodeAt(c)!==10?-1:(d=Ue(e.slice(1,l)),typeof t.references[d]>"u"&&(t.references[d]={title:f,href:u}),c)}function mn(e){var n=e.tokens,r,t,o,l;if(e.env.references=e.env.references||{},!e.inlineMode){for(r=1,t=n.length-1;r<t;r++)if(n[r].type==="inline"&&n[r-1].type==="paragraph_open"&&n[r+1].type==="paragraph_close"){for(o=n[r].content;o.length&&(l=hn(o,e.inline,e.options,e.env),!(l<0));)o=o.slice(l).trim();n[r].content=o,o.length||(n[r-1].tight=!0,n[r+1].tight=!0)}}}function gn(e){var n=e.tokens,r,t,o;for(t=0,o=n.length;t<o;t++)r=n[t],r.type==="inline"&&e.inline.parse(r.content,e.options,e.env,r.children)}function vn(e){var n,r,t,o,l,c,i,a,s,u=0,f=!1,d={};if(e.env.footnotes&&(e.tokens=e.tokens.filter(function(m){return m.type==="footnote_reference_open"?(f=!0,a=[],s=m.label,!1):m.type==="footnote_reference_close"?(f=!1,d[":"+s]=a,!1):(f&&a.push(m),!f)}),!!e.env.footnotes.list)){for(c=e.env.footnotes.list,e.tokens.push({type:"footnote_block_open",level:u++}),n=0,r=c.length;n<r;n++){for(e.tokens.push({type:"footnote_open",id:n,level:u++}),c[n].tokens?(i=[],i.push({type:"paragraph_open",tight:!1,level:u++}),i.push({type:"inline",content:"",level:u,children:c[n].tokens}),i.push({type:"paragraph_close",tight:!1,level:--u})):c[n].label&&(i=d[":"+c[n].label]),e.tokens=e.tokens.concat(i),e.tokens[e.tokens.length-1].type==="paragraph_close"?l=e.tokens.pop():l=null,o=c[n].count>0?c[n].count:1,t=0;t<o;t++)e.tokens.push({type:"footnote_anchor",id:n,subId:t,level:u});l&&e.tokens.push(l),e.tokens.push({type:"footnote_close",level:--u})}e.tokens.push({type:"footnote_block_close",level:--u})}}var Ie=` 
-()[]'".,!?-`;function oe(e){return e.replace(/([-()\[\]{}+?*.$\^|,:#<!\\])/g,"\\$1")}function xn(e){var n,r,t,o,l,c,i,a,s,u,f,d,m=e.tokens;if(e.env.abbreviations){for(e.env.abbrRegExp||(d="(^|["+Ie.split("").map(oe).join("")+"])("+Object.keys(e.env.abbreviations).map(function(g){return g.substr(1)}).sort(function(g,p){return p.length-g.length}).map(oe).join("|")+")($|["+Ie.split("").map(oe).join("")+"])",e.env.abbrRegExp=new RegExp(d,"g")),u=e.env.abbrRegExp,r=0,t=m.length;r<t;r++)if(m[r].type==="inline"){for(o=m[r].children,n=o.length-1;n>=0;n--)if(l=o[n],l.type==="text"){for(a=0,c=l.content,u.lastIndex=0,s=l.level,i=[];f=u.exec(c);)u.lastIndex>a&&i.push({type:"text",content:c.slice(a,f.index+f[1].length),level:s}),i.push({type:"abbr_open",title:e.env.abbreviations[":"+f[2]],level:s++}),i.push({type:"text",content:f[2],level:s}),i.push({type:"abbr_close",level:--s}),a=u.lastIndex-f[3].length;i.length&&(a<c.length&&i.push({type:"text",content:c.slice(a),level:s}),m[r].children=o=[].concat(o.slice(0,n),i,o.slice(n+1)))}}}}var bn=/\+-|\.\.|\?\?\?\?|!!!!|,,|--/,_n=/\((c|tm|r|p)\)/ig,kn={c:"\xA9",r:"\xAE",p:"\xA7",tm:"\u2122"};function yn(e){return e.indexOf("(")<0?e:e.replace(_n,function(n,r){return kn[r.toLowerCase()]})}function Cn(e){var n,r,t,o,l;if(e.options.typographer){for(l=e.tokens.length-1;l>=0;l--)if(e.tokens[l].type==="inline")for(o=e.tokens[l].children,n=o.length-1;n>=0;n--)r=o[n],r.type==="text"&&(t=r.content,t=yn(t),bn.test(t)&&(t=t.replace(/\+-/g,"\xB1").replace(/\.{2,}/g,"\u2026").replace(/([?!])…/g,"$1..").replace(/([?!]){4,}/g,"$1$1$1").replace(/,{2,}/g,",").replace(/(^|[^-])---([^-]|$)/mg,"$1\u2014$2").replace(/(^|\s)--(\s|$)/mg,"$1\u2013$2").replace(/(^|[^-\s])--([^-\s]|$)/mg,"$1\u2013$2")),r.content=t)}}var En=/['"]/,Ne=/['"]/g,An=/[-\s()\[\]]/,qe="\u2019";function Le(e,n){return n<0||n>=e.length?!1:!An.test(e[n])}function L(e,n,r){return e.substr(0,n)+r+e.substr(n+1)}function Tn(e){var n,r,t,o,l,c,i,a,s,u,f,d,m,g,p,v,x;if(e.options.typographer){for(x=[],p=e.tokens.length-1;p>=0;p--)if(e.tokens[p].type==="inline"){for(v=e.tokens[p].children,x.length=0,n=0;n<v.length;n++)if(r=v[n],!(r.type!=="text"||En.test(r.text))){for(i=v[n].level,m=x.length-1;m>=0&&!(x[m].level<=i);m--);x.length=m+1,t=r.content,l=0,c=t.length;e:for(;l<c&&(Ne.lastIndex=l,o=Ne.exec(t),!!o);){if(a=!Le(t,o.index-1),l=o.index+1,g=o[0]==="'",s=!Le(t,l),!s&&!a){g&&(r.content=L(r.content,o.index,qe));continue}if(f=!s,d=!a,d){for(m=x.length-1;m>=0&&(u=x[m],!(x[m].level<i));m--)if(u.single===g&&x[m].level===i){u=x[m],g?(v[u.token].content=L(v[u.token].content,u.pos,e.options.quotes[2]),r.content=L(r.content,o.index,e.options.quotes[3])):(v[u.token].content=L(v[u.token].content,u.pos,e.options.quotes[0]),r.content=L(r.content,o.index,e.options.quotes[1])),x.length=m;continue e}}f?x.push({token:n,pos:o.index,single:g,level:i}):d&&g&&(r.content=L(r.content,o.index,qe))}}}}}var te=[["block",fn],["abbr",dn],["references",mn],["inline",gn],["footnote_tail",vn],["abbr2",xn],["replacements",Cn],["smartquotes",Tn]];function Ve(){this.options={},this.ruler=new y;for(var e=0;e<te.length;e++)this.ruler.push(te[e][0],te[e][1])}Ve.prototype.process=function(e){var n,r,t;for(t=this.ruler.getRules(""),n=0,r=t.length;n<r;n++)t[n](e)};function N(e,n,r,t,o){var l,c,i,a,s,u,f;for(this.src=e,this.parser=n,this.options=r,this.env=t,this.tokens=o,this.bMarks=[],this.eMarks=[],this.tShift=[],this.blkIndent=0,this.line=0,this.lineMax=0,this.tight=!1,this.parentType="root",this.ddIndent=-1,this.level=0,this.result="",c=this.src,u=0,f=!1,i=a=u=0,s=c.length;a<s;a++){if(l=c.charCodeAt(a),!f)if(l===32){u++;continue}else f=!0;(l===10||a===s-1)&&(l!==10&&a++,this.bMarks.push(i),this.eMarks.push(a),this.tShift.push(u),f=!1,u=0,i=a+1)}this.bMarks.push(c.length),this.eMarks.push(c.length),this.tShift.push(0),this.lineMax=this.bMarks.length-1}N.prototype.isEmpty=function(n){return this.bMarks[n]+this.tShift[n]>=this.eMarks[n]};N.prototype.skipEmptyLines=function(n){for(var r=this.lineMax;n<r&&!(this.bMarks[n]+this.tShift[n]<this.eMarks[n]);n++);return n};N.prototype.skipSpaces=function(n){for(var r=this.src.length;n<r&&this.src.charCodeAt(n)===32;n++);return n};N.prototype.skipChars=function(n,r){for(var t=this.src.length;n<t&&this.src.charCodeAt(n)===r;n++);return n};N.prototype.skipCharsBack=function(n,r,t){if(n<=t)return n;for(;n>t;)if(r!==this.src.charCodeAt(--n))return n+1;return n};N.prototype.getLines=function(n,r,t,o){var l,c,i,a,s,u=n;if(n>=r)return"";if(u+1===r)return c=this.bMarks[u]+Math.min(this.tShift[u],t),i=o?this.eMarks[u]+1:this.eMarks[u],this.src.slice(c,i);for(a=new Array(r-n),l=0;u<r;u++,l++)s=this.tShift[u],s>t&&(s=t),s<0&&(s=0),c=this.bMarks[u]+s,u+1<r||o?i=this.eMarks[u]+1:i=this.eMarks[u],a[l]=this.src.slice(c,i);return a.join("")};function Sn(e,n,r){var t,o;if(e.tShift[n]-e.blkIndent<4)return!1;for(o=t=n+1;t<r;){if(e.isEmpty(t)){t++;continue}if(e.tShift[t]-e.blkIndent>=4){t++,o=t;continue}break}return e.line=t,e.tokens.push({type:"code",content:e.getLines(n,o,4+e.blkIndent,!0),block:!0,lines:[n,e.line],level:e.level}),!0}function Mn(e,n,r,t){var o,l,c,i,a,s=!1,u=e.bMarks[n]+e.tShift[n],f=e.eMarks[n];if(u+3>f||(o=e.src.charCodeAt(u),o!==126&&o!==96)||(a=u,u=e.skipChars(u,o),l=u-a,l<3)||(c=e.src.slice(u,f).trim(),c.indexOf("`")>=0))return!1;if(t)return!0;for(i=n;i++,!(i>=r||(u=a=e.bMarks[i]+e.tShift[i],f=e.eMarks[i],u<f&&e.tShift[i]<e.blkIndent));)if(e.src.charCodeAt(u)===o&&!(e.tShift[i]-e.blkIndent>=4)&&(u=e.skipChars(u,o),!(u-a<l)&&(u=e.skipSpaces(u),!(u<f)))){s=!0;break}return l=e.tShift[n],e.line=i+(s?1:0),e.tokens.push({type:"fence",params:c,content:e.getLines(n+1,i,l,!0),lines:[n,e.line],level:e.level}),!0}function wn(e,n,r,t){var o,l,c,i,a,s,u,f,d,m,g,p=e.bMarks[n]+e.tShift[n],v=e.eMarks[n];if(p>v||e.src.charCodeAt(p++)!==62||e.level>=e.options.maxNesting)return!1;if(t)return!0;for(e.src.charCodeAt(p)===32&&p++,a=e.blkIndent,e.blkIndent=0,i=[e.bMarks[n]],e.bMarks[n]=p,p=p<v?e.skipSpaces(p):p,l=p>=v,c=[e.tShift[n]],e.tShift[n]=p-e.bMarks[n],f=e.parser.ruler.getRules("blockquote"),o=n+1;o<r&&(p=e.bMarks[o]+e.tShift[o],v=e.eMarks[o],!(p>=v));o++){if(e.src.charCodeAt(p++)===62){e.src.charCodeAt(p)===32&&p++,i.push(e.bMarks[o]),e.bMarks[o]=p,p=p<v?e.skipSpaces(p):p,l=p>=v,c.push(e.tShift[o]),e.tShift[o]=p-e.bMarks[o];continue}if(l)break;for(g=!1,d=0,m=f.length;d<m;d++)if(f[d](e,o,r,!0)){g=!0;break}if(g)break;i.push(e.bMarks[o]),c.push(e.tShift[o]),e.tShift[o]=-1337}for(s=e.parentType,e.parentType="blockquote",e.tokens.push({type:"blockquote_open",lines:u=[n,0],level:e.level++}),e.parser.tokenize(e,n,o),e.tokens.push({type:"blockquote_close",level:--e.level}),e.parentType=s,u[1]=e.line,d=0;d<c.length;d++)e.bMarks[d+n]=i[d],e.tShift[d+n]=c[d];return e.blkIndent=a,!0}function In(e,n,r,t){var o,l,c,i=e.bMarks[n],a=e.eMarks[n];if(i+=e.tShift[n],i>a||(o=e.src.charCodeAt(i++),o!==42&&o!==45&&o!==95))return!1;for(l=1;i<a;){if(c=e.src.charCodeAt(i++),c!==o&&c!==32)return!1;c===o&&l++}return l<3?!1:(t||(e.line=n+1,e.tokens.push({type:"hr",lines:[n,e.line],level:e.level})),!0)}function He(e,n){var r,t,o;return t=e.bMarks[n]+e.tShift[n],o=e.eMarks[n],t>=o||(r=e.src.charCodeAt(t++),r!==42&&r!==45&&r!==43)||t<o&&e.src.charCodeAt(t)!==32?-1:t}function Be(e,n){var r,t=e.bMarks[n]+e.tShift[n],o=e.eMarks[n];if(t+1>=o||(r=e.src.charCodeAt(t++),r<48||r>57))return-1;for(;;){if(t>=o)return-1;if(r=e.src.charCodeAt(t++),!(r>=48&&r<=57)){if(r===41||r===46)break;return-1}}return t<o&&e.src.charCodeAt(t)!==32?-1:t}function Nn(e,n){var r,t,o=e.level+2;for(r=n+2,t=e.tokens.length-2;r<t;r++)e.tokens[r].level===o&&e.tokens[r].type==="paragraph_open"&&(e.tokens[r+2].tight=!0,e.tokens[r].tight=!0,r+=2)}function qn(e,n,r,t){var o,l,c,i,a,s,u,f,d,m,g,p,v,x,S,b,K,_e,ke=!0,ee,z,ye,ne;if((f=Be(e,n))>=0)v=!0;else if((f=He(e,n))>=0)v=!1;else return!1;if(e.level>=e.options.maxNesting)return!1;if(p=e.src.charCodeAt(f-1),t)return!0;for(S=e.tokens.length,v?(u=e.bMarks[n]+e.tShift[n],g=Number(e.src.substr(u,f-u-1)),e.tokens.push({type:"ordered_list_open",order:g,lines:K=[n,0],level:e.level++})):e.tokens.push({type:"bullet_list_open",lines:K=[n,0],level:e.level++}),o=n,b=!1,ee=e.parser.ruler.getRules("list");o<r&&(x=e.skipSpaces(f),d=e.eMarks[o],x>=d?m=1:m=x-f,m>4&&(m=1),m<1&&(m=1),l=f-e.bMarks[o]+m,e.tokens.push({type:"list_item_open",lines:_e=[n,0],level:e.level++}),i=e.blkIndent,a=e.tight,c=e.tShift[n],s=e.parentType,e.tShift[n]=x-e.bMarks[n],e.blkIndent=l,e.tight=!0,e.parentType="list",e.parser.tokenize(e,n,r,!0),(!e.tight||b)&&(ke=!1),b=e.line-n>1&&e.isEmpty(e.line-1),e.blkIndent=i,e.tShift[n]=c,e.tight=a,e.parentType=s,e.tokens.push({type:"list_item_close",level:--e.level}),o=n=e.line,_e[1]=o,x=e.bMarks[n],!(o>=r||e.isEmpty(o)||e.tShift[o]<e.blkIndent));){for(ne=!1,z=0,ye=ee.length;z<ye;z++)if(ee[z](e,o,r,!0)){ne=!0;break}if(ne)break;if(v){if(f=Be(e,o),f<0)break}else if(f=He(e,o),f<0)break;if(p!==e.src.charCodeAt(f-1))break}return e.tokens.push({type:v?"ordered_list_close":"bullet_list_close",level:--e.level}),K[1]=o,e.line=o,ke&&Nn(e,S),!0}function Ln(e,n,r,t){var o,l,c,i,a,s=e.bMarks[n]+e.tShift[n],u=e.eMarks[n];if(s+4>u||e.src.charCodeAt(s)!==91||e.src.charCodeAt(s+1)!==94||e.level>=e.options.maxNesting)return!1;for(i=s+2;i<u;i++){if(e.src.charCodeAt(i)===32)return!1;if(e.src.charCodeAt(i)===93)break}return i===s+2||i+1>=u||e.src.charCodeAt(++i)!==58?!1:(t||(i++,e.env.footnotes||(e.env.footnotes={}),e.env.footnotes.refs||(e.env.footnotes.refs={}),a=e.src.slice(s+2,i-2),e.env.footnotes.refs[":"+a]=-1,e.tokens.push({type:"footnote_reference_open",label:a,level:e.level++}),o=e.bMarks[n],l=e.tShift[n],c=e.parentType,e.tShift[n]=e.skipSpaces(i)-i,e.bMarks[n]=i,e.blkIndent+=4,e.parentType="footnote",e.tShift[n]<e.blkIndent&&(e.tShift[n]+=e.blkIndent,e.bMarks[n]-=e.blkIndent),e.parser.tokenize(e,n,r,!0),e.parentType=c,e.blkIndent-=4,e.tShift[n]=l,e.bMarks[n]=o,e.tokens.push({type:"footnote_reference_close",level:--e.level})),!0)}function Hn(e,n,r,t){var o,l,c,i=e.bMarks[n]+e.tShift[n],a=e.eMarks[n];if(i>=a||(o=e.src.charCodeAt(i),o!==35||i>=a))return!1;for(l=1,o=e.src.charCodeAt(++i);o===35&&i<a&&l<=6;)l++,o=e.src.charCodeAt(++i);return l>6||i<a&&o!==32?!1:(t||(a=e.skipCharsBack(a,32,i),c=e.skipCharsBack(a,35,i),c>i&&e.src.charCodeAt(c-1)===32&&(a=c),e.line=n+1,e.tokens.push({type:"heading_open",hLevel:l,lines:[n,e.line],level:e.level}),i<a&&e.tokens.push({type:"inline",content:e.src.slice(i,a).trim(),level:e.level+1,lines:[n,e.line],children:[]}),e.tokens.push({type:"heading_close",hLevel:l,level:e.level})),!0)}function Bn(e,n,r){var t,o,l,c=n+1;return c>=r||e.tShift[c]<e.blkIndent||e.tShift[c]-e.blkIndent>3||(o=e.bMarks[c]+e.tShift[c],l=e.eMarks[c],o>=l)||(t=e.src.charCodeAt(o),t!==45&&t!==61)||(o=e.skipChars(o,t),o=e.skipSpaces(o),o<l)?!1:(o=e.bMarks[n]+e.tShift[n],e.line=c+1,e.tokens.push({type:"heading_open",hLevel:t===61?1:2,lines:[n,e.line],level:e.level}),e.tokens.push({type:"inline",content:e.src.slice(o,e.eMarks[n]).trim(),level:e.level+1,lines:[n,e.line-1],children:[]}),e.tokens.push({type:"heading_close",hLevel:t===61?1:2,level:e.level}),!0)}var Ye={};["article","aside","button","blockquote","body","canvas","caption","col","colgroup","dd","div","dl","dt","embed","fieldset","figcaption","figure","footer","form","h1","h2","h3","h4","h5","h6","header","hgroup","hr","iframe","li","map","object","ol","output","p","pre","progress","script","section","style","table","tbody","td","textarea","tfoot","th","tr","thead","ul","video"].forEach(function(e){Ye[e]=!0});var Rn=/^<([a-zA-Z]{1,15})[\s\/>]/,Dn=/^<\/([a-zA-Z]{1,15})[\s>]/;function On(e){var n=e|32;return n>=97&&n<=122}function Pn(e,n,r,t){var o,l,c,i=e.bMarks[n],a=e.eMarks[n],s=e.tShift[n];if(i+=s,!e.options.html||s>3||i+2>=a||e.src.charCodeAt(i)!==60)return!1;if(o=e.src.charCodeAt(i+1),o===33||o===63){if(t)return!0}else if(o===47||On(o)){if(o===47){if(l=e.src.slice(i,a).match(Dn),!l)return!1}else if(l=e.src.slice(i,a).match(Rn),!l)return!1;if(Ye[l[1].toLowerCase()]!==!0)return!1;if(t)return!0}else return!1;for(c=n+1;c<e.lineMax&&!e.isEmpty(c);)c++;return e.line=c,e.tokens.push({type:"htmlblock",level:e.level,lines:[n,e.line],content:e.getLines(n,c,0,!0)}),!0}function le(e,n){var r=e.bMarks[n]+e.blkIndent,t=e.eMarks[n];return e.src.substr(r,t-r)}function Fn(e,n,r,t){var o,l,c,i,a,s,u,f,d,m,g;if(n+2>r||(a=n+1,e.tShift[a]<e.blkIndent)||(c=e.bMarks[a]+e.tShift[a],c>=e.eMarks[a])||(o=e.src.charCodeAt(c),o!==124&&o!==45&&o!==58)||(l=le(e,n+1),!/^[-:| ]+$/.test(l))||(s=l.split("|"),s<=2))return!1;for(f=[],i=0;i<s.length;i++){if(d=s[i].trim(),!d){if(i===0||i===s.length-1)continue;return!1}if(!/^:?-+:?$/.test(d))return!1;d.charCodeAt(d.length-1)===58?f.push(d.charCodeAt(0)===58?"center":"right"):d.charCodeAt(0)===58?f.push("left"):f.push("")}if(l=le(e,n).trim(),l.indexOf("|")===-1||(s=l.replace(/^\||\|$/g,"").split("|"),f.length!==s.length))return!1;if(t)return!0;for(e.tokens.push({type:"table_open",lines:m=[n,0],level:e.level++}),e.tokens.push({type:"thead_open",lines:[n,n+1],level:e.level++}),e.tokens.push({type:"tr_open",lines:[n,n+1],level:e.level++}),i=0;i<s.length;i++)e.tokens.push({type:"th_open",align:f[i],lines:[n,n+1],level:e.level++}),e.tokens.push({type:"inline",content:s[i].trim(),lines:[n,n+1],level:e.level,children:[]}),e.tokens.push({type:"th_close",level:--e.level});for(e.tokens.push({type:"tr_close",level:--e.level}),e.tokens.push({type:"thead_close",level:--e.level}),e.tokens.push({type:"tbody_open",lines:g=[n+2,0],level:e.level++}),a=n+2;a<r&&!(e.tShift[a]<e.blkIndent||(l=le(e,a).trim(),l.indexOf("|")===-1));a++){for(s=l.replace(/^\||\|$/g,"").split("|"),e.tokens.push({type:"tr_open",level:e.level++}),i=0;i<s.length;i++)e.tokens.push({type:"td_open",align:f[i],level:e.level++}),u=s[i].substring(s[i].charCodeAt(0)===124?1:0,s[i].charCodeAt(s[i].length-1)===124?s[i].length-1:s[i].length).trim(),e.tokens.push({type:"inline",content:u,level:e.level,children:[]}),e.tokens.push({type:"td_close",level:--e.level});e.tokens.push({type:"tr_close",level:--e.level})}return e.tokens.push({type:"tbody_close",level:--e.level}),e.tokens.push({type:"table_close",level:--e.level}),m[1]=g[1]=a,e.line=a,!0}function Y(e,n){var r,t,o=e.bMarks[n]+e.tShift[n],l=e.eMarks[n];return o>=l||(t=e.src.charCodeAt(o++),t!==126&&t!==58)||(r=e.skipSpaces(o),o===r)||r>=l?-1:r}function zn(e,n){var r,t,o=e.level+2;for(r=n+2,t=e.tokens.length-2;r<t;r++)e.tokens[r].level===o&&e.tokens[r].type==="paragraph_open"&&(e.tokens[r+2].tight=!0,e.tokens[r].tight=!0,r+=2)}function jn(e,n,r,t){var o,l,c,i,a,s,u,f,d,m,g,p,v,x;if(t)return e.ddIndent<0?!1:Y(e,n)>=0;if(u=n+1,e.isEmpty(u)&&++u>r||e.tShift[u]<e.blkIndent||(o=Y(e,u),o<0)||e.level>=e.options.maxNesting)return!1;s=e.tokens.length,e.tokens.push({type:"dl_open",lines:a=[n,0],level:e.level++}),c=n,l=u;e:for(;;){for(x=!0,v=!1,e.tokens.push({type:"dt_open",lines:[c,c],level:e.level++}),e.tokens.push({type:"inline",content:e.getLines(c,c+1,e.blkIndent,!1).trim(),level:e.level+1,lines:[c,c],children:[]}),e.tokens.push({type:"dt_close",level:--e.level});;){if(e.tokens.push({type:"dd_open",lines:i=[u,0],level:e.level++}),p=e.tight,d=e.ddIndent,f=e.blkIndent,g=e.tShift[l],m=e.parentType,e.blkIndent=e.ddIndent=e.tShift[l]+2,e.tShift[l]=o-e.bMarks[l],e.tight=!0,e.parentType="deflist",e.parser.tokenize(e,l,r,!0),(!e.tight||v)&&(x=!1),v=e.line-l>1&&e.isEmpty(e.line-1),e.tShift[l]=g,e.tight=p,e.parentType=m,e.blkIndent=f,e.ddIndent=d,e.tokens.push({type:"dd_close",level:--e.level}),i[1]=u=e.line,u>=r||e.tShift[u]<e.blkIndent)break e;if(o=Y(e,u),o<0)break;l=u}if(u>=r||(c=u,e.isEmpty(c))||e.tShift[c]<e.blkIndent||(l=c+1,l>=r)||(e.isEmpty(l)&&l++,l>=r)||e.tShift[l]<e.blkIndent||(o=Y(e,l),o<0))break}return e.tokens.push({type:"dl_close",level:--e.level}),a[1]=u,e.line=u,x&&zn(e,s),!0}function $n(e,n){var r,t,o,l,c,i=n+1,a;if(r=e.lineMax,i<r&&!e.isEmpty(i)){for(a=e.parser.ruler.getRules("paragraph");i<r&&!e.isEmpty(i);i++)if(!(e.tShift[i]-e.blkIndent>3)){for(o=!1,l=0,c=a.length;l<c;l++)if(a[l](e,i,r,!0)){o=!0;break}if(o)break}}return t=e.getLines(n,i,e.blkIndent,!1).trim(),e.line=i,t.length&&(e.tokens.push({type:"paragraph_open",tight:!1,lines:[n,e.line],level:e.level}),e.tokens.push({type:"inline",content:t,level:e.level+1,lines:[n,e.line],children:[]}),e.tokens.push({type:"paragraph_close",tight:!1,level:e.level})),!0}var Z=[["code",Sn],["fences",Mn,["paragraph","blockquote","list"]],["blockquote",wn,["paragraph","blockquote","list"]],["hr",In,["paragraph","blockquote","list"]],["list",qn,["paragraph","blockquote"]],["footnote",Ln,["paragraph"]],["heading",Hn,["paragraph","blockquote"]],["lheading",Bn],["htmlblock",Pn,["paragraph","blockquote"]],["table",Fn,["paragraph"]],["deflist",jn,["paragraph"]],["paragraph",$n]];function fe(){this.ruler=new y;for(var e=0;e<Z.length;e++)this.ruler.push(Z[e][0],Z[e][1],{alt:(Z[e][2]||[]).slice()})}fe.prototype.tokenize=function(e,n,r){for(var t=this.ruler.getRules(""),o=t.length,l=n,c=!1,i,a;l<r&&(e.line=l=e.skipEmptyLines(l),!(l>=r||e.tShift[l]<e.blkIndent));){for(a=0;a<o&&(i=t[a](e,l,r,!1),!i);a++);if(e.tight=!c,e.isEmpty(e.line-1)&&(c=!0),l=e.line,l<r&&e.isEmpty(l)){if(c=!0,l++,l<r&&e.parentType==="list"&&e.isEmpty(l))break;e.line=l}}};var Un=/[\n\t]/g,Vn=/\r[\n\u0085]|[\u2424\u2028\u0085]/g,Yn=/\u00a0/g;fe.prototype.parse=function(e,n,r,t){var o,l=0,c=0;if(!e)return[];e=e.replace(Yn," "),e=e.replace(Vn,`
-`),e.indexOf("	")>=0&&(e=e.replace(Un,function(i,a){var s;return e.charCodeAt(a)===10?(l=a+1,c=0,i):(s="    ".slice((a-l-c)%4),c=a-l+1,s)})),o=new N(e,this,n,r,t),this.tokenize(o,o.line,o.lineMax)};function Zn(e){switch(e){case 10:case 92:case 96:case 42:case 95:case 94:case 91:case 93:case 33:case 38:case 60:case 62:case 123:case 125:case 36:case 37:case 64:case 126:case 43:case 61:case 58:return!0;default:return!1}}function Wn(e,n){for(var r=e.pos;r<e.posMax&&!Zn(e.src.charCodeAt(r));)r++;return r===e.pos?!1:(n||(e.pending+=e.src.slice(e.pos,r)),e.pos=r,!0)}function Gn(e,n){var r,t,o=e.pos;if(e.src.charCodeAt(o)!==10)return!1;if(r=e.pending.length-1,t=e.posMax,!n)if(r>=0&&e.pending.charCodeAt(r)===32)if(r>=1&&e.pending.charCodeAt(r-1)===32){for(var l=r-2;l>=0;l--)if(e.pending.charCodeAt(l)!==32){e.pending=e.pending.substring(0,l+1);break}e.push({type:"hardbreak",level:e.level})}else e.pending=e.pending.slice(0,-1),e.push({type:"softbreak",level:e.level});else e.push({type:"softbreak",level:e.level});for(o++;o<t&&e.src.charCodeAt(o)===32;)o++;return e.pos=o,!0}var pe=[];for(ie=0;ie<256;ie++)pe.push(0);var ie;"\\!\"#$%&'()*+,./:;<=>?@[]^_`{|}~-".split("").forEach(function(e){pe[e.charCodeAt(0)]=1});function Xn(e,n){var r,t=e.pos,o=e.posMax;if(e.src.charCodeAt(t)!==92)return!1;if(t++,t<o){if(r=e.src.charCodeAt(t),r<256&&pe[r]!==0)return n||(e.pending+=e.src[t]),e.pos+=2,!0;if(r===10){for(n||e.push({type:"hardbreak",level:e.level}),t++;t<o&&e.src.charCodeAt(t)===32;)t++;return e.pos=t,!0}}return n||(e.pending+="\\"),e.pos++,!0}function Jn(e,n){var r,t,o,l,c,i=e.pos,a=e.src.charCodeAt(i);if(a!==96)return!1;for(r=i,i++,t=e.posMax;i<t&&e.src.charCodeAt(i)===96;)i++;for(o=e.src.slice(r,i),l=c=i;(l=e.src.indexOf("`",c))!==-1;){for(c=l+1;c<t&&e.src.charCodeAt(c)===96;)c++;if(c-l===o.length)return n||e.push({type:"code",content:e.src.slice(i,l).replace(/[ \n]+/g," ").trim(),block:!1,level:e.level}),e.pos=c,!0}return n||(e.pending+=o),e.pos+=o.length,!0}function Qn(e,n){var r,t,o,l=e.posMax,c=e.pos,i,a;if(e.src.charCodeAt(c)!==126||n||c+4>=l||e.src.charCodeAt(c+1)!==126||e.level>=e.options.maxNesting||(i=c>0?e.src.charCodeAt(c-1):-1,a=e.src.charCodeAt(c+2),i===126)||a===126||a===32||a===10)return!1;for(t=c+2;t<l&&e.src.charCodeAt(t)===126;)t++;if(t>c+3)return e.pos+=t-c,n||(e.pending+=e.src.slice(c,t)),!0;for(e.pos=c+2,o=1;e.pos+1<l;){if(e.src.charCodeAt(e.pos)===126&&e.src.charCodeAt(e.pos+1)===126&&(i=e.src.charCodeAt(e.pos-1),a=e.pos+2<l?e.src.charCodeAt(e.pos+2):-1,a!==126&&i!==126&&(i!==32&&i!==10?o--:a!==32&&a!==10&&o++,o<=0))){r=!0;break}e.parser.skipToken(e)}return r?(e.posMax=e.pos,e.pos=c+2,n||(e.push({type:"del_open",level:e.level++}),e.parser.tokenize(e),e.push({type:"del_close",level:--e.level})),e.pos=e.posMax+2,e.posMax=l,!0):(e.pos=c,!1)}function Kn(e,n){var r,t,o,l=e.posMax,c=e.pos,i,a;if(e.src.charCodeAt(c)!==43||n||c+4>=l||e.src.charCodeAt(c+1)!==43||e.level>=e.options.maxNesting||(i=c>0?e.src.charCodeAt(c-1):-1,a=e.src.charCodeAt(c+2),i===43)||a===43||a===32||a===10)return!1;for(t=c+2;t<l&&e.src.charCodeAt(t)===43;)t++;if(t!==c+2)return e.pos+=t-c,n||(e.pending+=e.src.slice(c,t)),!0;for(e.pos=c+2,o=1;e.pos+1<l;){if(e.src.charCodeAt(e.pos)===43&&e.src.charCodeAt(e.pos+1)===43&&(i=e.src.charCodeAt(e.pos-1),a=e.pos+2<l?e.src.charCodeAt(e.pos+2):-1,a!==43&&i!==43&&(i!==32&&i!==10?o--:a!==32&&a!==10&&o++,o<=0))){r=!0;break}e.parser.skipToken(e)}return r?(e.posMax=e.pos,e.pos=c+2,n||(e.push({type:"ins_open",level:e.level++}),e.parser.tokenize(e),e.push({type:"ins_close",level:--e.level})),e.pos=e.posMax+2,e.posMax=l,!0):(e.pos=c,!1)}function er(e,n){var r,t,o,l=e.posMax,c=e.pos,i,a;if(e.src.charCodeAt(c)!==61||n||c+4>=l||e.src.charCodeAt(c+1)!==61||e.level>=e.options.maxNesting||(i=c>0?e.src.charCodeAt(c-1):-1,a=e.src.charCodeAt(c+2),i===61)||a===61||a===32||a===10)return!1;for(t=c+2;t<l&&e.src.charCodeAt(t)===61;)t++;if(t!==c+2)return e.pos+=t-c,n||(e.pending+=e.src.slice(c,t)),!0;for(e.pos=c+2,o=1;e.pos+1<l;){if(e.src.charCodeAt(e.pos)===61&&e.src.charCodeAt(e.pos+1)===61&&(i=e.src.charCodeAt(e.pos-1),a=e.pos+2<l?e.src.charCodeAt(e.pos+2):-1,a!==61&&i!==61&&(i!==32&&i!==10?o--:a!==32&&a!==10&&o++,o<=0))){r=!0;break}e.parser.skipToken(e)}return r?(e.posMax=e.pos,e.pos=c+2,n||(e.push({type:"mark_open",level:e.level++}),e.parser.tokenize(e),e.push({type:"mark_close",level:--e.level})),e.pos=e.posMax+2,e.posMax=l,!0):(e.pos=c,!1)}function Re(e){return e>=48&&e<=57||e>=65&&e<=90||e>=97&&e<=122}function De(e,n){var r=n,t,o,l,c=!0,i=!0,a=e.posMax,s=e.src.charCodeAt(n);for(t=n>0?e.src.charCodeAt(n-1):-1;r<a&&e.src.charCodeAt(r)===s;)r++;return r>=a&&(c=!1),l=r-n,l>=4?c=i=!1:(o=r<a?e.src.charCodeAt(r):-1,(o===32||o===10)&&(c=!1),(t===32||t===10)&&(i=!1),s===95&&(Re(t)&&(c=!1),Re(o)&&(i=!1))),{can_open:c,can_close:i,delims:l}}function nr(e,n){var r,t,o,l,c,i,a,s=e.posMax,u=e.pos,f=e.src.charCodeAt(u);if(f!==95&&f!==42||n)return!1;if(a=De(e,u),r=a.delims,!a.can_open)return e.pos+=r,n||(e.pending+=e.src.slice(u,e.pos)),!0;if(e.level>=e.options.maxNesting)return!1;for(e.pos=u+r,i=[r];e.pos<s;){if(e.src.charCodeAt(e.pos)===f){if(a=De(e,e.pos),t=a.delims,a.can_close){for(l=i.pop(),c=t;l!==c;){if(c<l){i.push(l-c);break}if(c-=l,i.length===0)break;e.pos+=l,l=i.pop()}if(i.length===0){r=l,o=!0;break}e.pos+=t;continue}a.can_open&&i.push(t),e.pos+=t;continue}e.parser.skipToken(e)}return o?(e.posMax=e.pos,e.pos=u+r,n||((r===2||r===3)&&e.push({type:"strong_open",level:e.level++}),(r===1||r===3)&&e.push({type:"em_open",level:e.level++}),e.parser.tokenize(e),(r===1||r===3)&&e.push({type:"em_close",level:--e.level}),(r===2||r===3)&&e.push({type:"strong_close",level:--e.level})),e.pos=e.posMax+r,e.posMax=s,!0):(e.pos=u,!1)}var rr=/\\([ \\!"#$%&'()*+,.\/:;<=>?@[\]^_`{|}~-])/g;function or(e,n){var r,t,o=e.posMax,l=e.pos;if(e.src.charCodeAt(l)!==126||n||l+2>=o||e.level>=e.options.maxNesting)return!1;for(e.pos=l+1;e.pos<o;){if(e.src.charCodeAt(e.pos)===126){r=!0;break}e.parser.skipToken(e)}return!r||l+1===e.pos||(t=e.src.slice(l+1,e.pos),t.match(/(^|[^\\])(\\\\)*\s/))?(e.pos=l,!1):(e.posMax=e.pos,e.pos=l+1,n||e.push({type:"sub",level:e.level,content:t.replace(rr,"$1")}),e.pos=e.posMax+1,e.posMax=o,!0)}var tr=/\\([ \\!"#$%&'()*+,.\/:;<=>?@[\]^_`{|}~-])/g;function lr(e,n){var r,t,o=e.posMax,l=e.pos;if(e.src.charCodeAt(l)!==94||n||l+2>=o||e.level>=e.options.maxNesting)return!1;for(e.pos=l+1;e.pos<o;){if(e.src.charCodeAt(e.pos)===94){r=!0;break}e.parser.skipToken(e)}return!r||l+1===e.pos||(t=e.src.slice(l+1,e.pos),t.match(/(^|[^\\])(\\\\)*\s/))?(e.pos=l,!1):(e.posMax=e.pos,e.pos=l+1,n||e.push({type:"sup",level:e.level,content:t.replace(tr,"$1")}),e.pos=e.posMax+1,e.posMax=o,!0)}function ir(e,n){var r,t,o,l,c,i,a,s,u=!1,f=e.pos,d=e.posMax,m=e.pos,g=e.src.charCodeAt(m);if(g===33&&(u=!0,g=e.src.charCodeAt(++m)),g!==91||e.level>=e.options.maxNesting||(r=m+1,t=O(e,m),t<0))return!1;if(i=t+1,i<d&&e.src.charCodeAt(i)===40){for(i++;i<d&&(s=e.src.charCodeAt(i),!(s!==32&&s!==10));i++);if(i>=d)return!1;for(m=i,je(e,i)?(l=e.linkContent,i=e.pos):l="",m=i;i<d&&(s=e.src.charCodeAt(i),!(s!==32&&s!==10));i++);if(i<d&&m!==i&&$e(e,i))for(c=e.linkContent,i=e.pos;i<d&&(s=e.src.charCodeAt(i),!(s!==32&&s!==10));i++);else c="";if(i>=d||e.src.charCodeAt(i)!==41)return e.pos=f,!1;i++}else{if(e.linkLevel>0)return!1;for(;i<d&&(s=e.src.charCodeAt(i),!(s!==32&&s!==10));i++);if(i<d&&e.src.charCodeAt(i)===91&&(m=i+1,i=O(e,i),i>=0?o=e.src.slice(m,i++):i=m-1),o||(typeof o>"u"&&(i=t+1),o=e.src.slice(r,t)),a=e.env.references[Ue(o)],!a)return e.pos=f,!1;l=a.href,c=a.title}return n||(e.pos=r,e.posMax=t,u?e.push({type:"image",src:l,title:c,alt:e.src.substr(r,t-r),level:e.level}):(e.push({type:"link_open",href:l,title:c,level:e.level++}),e.linkLevel++,e.parser.tokenize(e),e.linkLevel--,e.push({type:"link_close",level:--e.level}))),e.pos=i,e.posMax=d,!0}function cr(e,n){var r,t,o,l,c=e.posMax,i=e.pos;return i+2>=c||e.src.charCodeAt(i)!==94||e.src.charCodeAt(i+1)!==91||e.level>=e.options.maxNesting||(r=i+2,t=O(e,i+1),t<0)?!1:(n||(e.env.footnotes||(e.env.footnotes={}),e.env.footnotes.list||(e.env.footnotes.list=[]),o=e.env.footnotes.list.length,e.pos=r,e.posMax=t,e.push({type:"footnote_ref",id:o,level:e.level}),e.linkLevel++,l=e.tokens.length,e.parser.tokenize(e),e.env.footnotes.list[o]={tokens:e.tokens.splice(l)},e.linkLevel--),e.pos=t+1,e.posMax=c,!0)}function ar(e,n){var r,t,o,l,c=e.posMax,i=e.pos;if(i+3>c||!e.env.footnotes||!e.env.footnotes.refs||e.src.charCodeAt(i)!==91||e.src.charCodeAt(i+1)!==94||e.level>=e.options.maxNesting)return!1;for(t=i+2;t<c;t++){if(e.src.charCodeAt(t)===32||e.src.charCodeAt(t)===10)return!1;if(e.src.charCodeAt(t)===93)break}return t===i+2||t>=c||(t++,r=e.src.slice(i+2,t-1),typeof e.env.footnotes.refs[":"+r]>"u")?!1:(n||(e.env.footnotes.list||(e.env.footnotes.list=[]),e.env.footnotes.refs[":"+r]<0?(o=e.env.footnotes.list.length,e.env.footnotes.list[o]={label:r,count:0},e.env.footnotes.refs[":"+r]=o):o=e.env.footnotes.refs[":"+r],l=e.env.footnotes.list[o].count,e.env.footnotes.list[o].count++,e.push({type:"footnote_ref",id:o,subId:l,level:e.level})),e.pos=t,e.posMax=c,!0)}var sr=["coap","doi","javascript","aaa","aaas","about","acap","cap","cid","crid","data","dav","dict","dns","file","ftp","geo","go","gopher","h323","http","https","iax","icap","im","imap","info","ipp","iris","iris.beep","iris.xpc","iris.xpcs","iris.lwz","ldap","mailto","mid","msrp","msrps","mtqp","mupdate","news","nfs","ni","nih","nntp","opaquelocktoken","pop","pres","rtsp","service","session","shttp","sieve","sip","sips","sms","snmp","soap.beep","soap.beeps","tag","tel","telnet","tftp","thismessage","tn3270","tip","tv","urn","vemmi","ws","wss","xcon","xcon-userid","xmlrpc.beep","xmlrpc.beeps","xmpp","z39.50r","z39.50s","adiumxtra","afp","afs","aim","apt","attachment","aw","beshare","bitcoin","bolo","callto","chrome","chrome-extension","com-eventbrite-attendee","content","cvs","dlna-playsingle","dlna-playcontainer","dtn","dvb","ed2k","facetime","feed","finger","fish","gg","git","gizmoproject","gtalk","hcp","icon","ipn","irc","irc6","ircs","itms","jar","jms","keyparc","lastfm","ldaps","magnet","maps","market","message","mms","ms-help","msnim","mumble","mvn","notes","oid","palm","paparazzi","platform","proxy","psyc","query","res","resource","rmi","rsync","rtmp","secondlife","sftp","sgn","skype","smb","soldat","spotify","ssh","steam","svn","teamspeak","things","udp","unreal","ut2004","ventrilo","view-source","webcal","wtai","wyciwyg","xfire","xri","ymsgr"],ur=/^<([a-zA-Z0-9.!#$%&'*+\/=?^_`{|}~-]+@[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?(?:\.[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?)*)>/,fr=/^<([a-zA-Z.\-]{1,25}):([^<>\x00-\x20]*)>/;function pr(e,n){var r,t,o,l,c,i=e.pos;return e.src.charCodeAt(i)!==60||(r=e.src.slice(i),r.indexOf(">")<0)?!1:(t=r.match(fr),t?sr.indexOf(t[1].toLowerCase())<0||(l=t[0].slice(1,-1),c=se(l),!e.parser.validateLink(l))?!1:(n||(e.push({type:"link_open",href:c,level:e.level}),e.push({type:"text",content:l,level:e.level+1}),e.push({type:"link_close",level:e.level})),e.pos+=t[0].length,!0):(o=r.match(ur),o?(l=o[0].slice(1,-1),c=se("mailto:"+l),e.parser.validateLink(c)?(n||(e.push({type:"link_open",href:c,level:e.level}),e.push({type:"text",content:l,level:e.level+1}),e.push({type:"link_close",level:e.level})),e.pos+=o[0].length,!0):!1):!1))}function W(e,n){return e=e.source,n=n||"",function r(t,o){return t?(o=o.source||o,e=e.replace(t,o),r):new RegExp(e,n)}}var dr=/[a-zA-Z_:][a-zA-Z0-9:._-]*/,hr=/[^"'=<>`\x00-\x20]+/,mr=/'[^']*'/,gr=/"[^"]*"/,vr=W(/(?:unquoted|single_quoted|double_quoted)/)("unquoted",hr)("single_quoted",mr)("double_quoted",gr)(),xr=W(/(?:\s+attr_name(?:\s*=\s*attr_value)?)/)("attr_name",dr)("attr_value",vr)(),br=W(/<[A-Za-z][A-Za-z0-9]*attribute*\s*\/?>/)("attribute",xr)(),_r=/<\/[A-Za-z][A-Za-z0-9]*\s*>/,kr=/<!---->|<!--(?:-?[^>-])(?:-?[^-])*-->/,yr=/<[?].*?[?]>/,Cr=/<![A-Z]+\s+[^>]*>/,Er=/<!\[CDATA\[[\s\S]*?\]\]>/,Ar=W(/^(?:open_tag|close_tag|comment|processing|declaration|cdata)/)("open_tag",br)("close_tag",_r)("comment",kr)("processing",yr)("declaration",Cr)("cdata",Er)();function Tr(e){var n=e|32;return n>=97&&n<=122}function Sr(e,n){var r,t,o,l=e.pos;return!e.options.html||(o=e.posMax,e.src.charCodeAt(l)!==60||l+2>=o)||(r=e.src.charCodeAt(l+1),r!==33&&r!==63&&r!==47&&!Tr(r))||(t=e.src.slice(l).match(Ar),!t)?!1:(n||e.push({type:"htmltag",content:e.src.slice(l,l+t[0].length),level:e.level}),e.pos+=t[0].length,!0)}var Mr=/^&#((?:x[a-f0-9]{1,8}|[0-9]{1,8}));/i,wr=/^&([a-z][a-z0-9]{1,31});/i;function Ir(e,n){var r,t,o,l=e.pos,c=e.posMax;if(e.src.charCodeAt(l)!==38)return!1;if(l+1<c){if(r=e.src.charCodeAt(l+1),r===35){if(o=e.src.slice(l).match(Mr),o)return n||(t=o[1][0].toLowerCase()==="x"?parseInt(o[1].slice(1),16):parseInt(o[1],10),e.pending+=Fe(t)?ae(t):ae(65533)),e.pos+=o[0].length,!0}else if(o=e.src.slice(l).match(wr),o){var i=Oe(o[1]);if(o[1]!==i)return n||(e.pending+=i),e.pos+=o[0].length,!0}}return n||(e.pending+="&"),e.pos++,!0}var ce=[["text",Wn],["newline",Gn],["escape",Xn],["backticks",Jn],["del",Qn],["ins",Kn],["mark",er],["emphasis",nr],["sub",or],["sup",lr],["links",ir],["footnote_inline",cr],["footnote_ref",ar],["autolink",pr],["htmltag",Sr],["entity",Ir]];function G(){this.ruler=new y;for(var e=0;e<ce.length;e++)this.ruler.push(ce[e][0],ce[e][1]);this.validateLink=Nr}G.prototype.skipToken=function(e){var n=this.ruler.getRules(""),r=n.length,t=e.pos,o,l;if((l=e.cacheGet(t))>0){e.pos=l;return}for(o=0;o<r;o++)if(n[o](e,!0)){e.cacheSet(t,e.pos);return}e.pos++,e.cacheSet(t,e.pos)};G.prototype.tokenize=function(e){for(var n=this.ruler.getRules(""),r=n.length,t=e.posMax,o,l;e.pos<t;){for(l=0;l<r&&(o=n[l](e,!1),!o);l++);if(o){if(e.pos>=t)break;continue}e.pending+=e.src[e.pos++]}e.pending&&e.pushPending()};G.prototype.parse=function(e,n,r,t){var o=new I(e,this,n,r,t);this.tokenize(o)};function Nr(e){var n=["vbscript","javascript","file","data"],r=e.trim().toLowerCase();return r=M(r),!(r.indexOf(":")!==-1&&n.indexOf(r.split(":")[0])!==-1)}var qr={options:{html:!1,xhtmlOut:!1,breaks:!1,langPrefix:"language-",linkTarget:"",typographer:!1,quotes:"\u201C\u201D\u2018\u2019",highlight:null,maxNesting:20},components:{core:{rules:["block","inline","references","replacements","smartquotes","references","abbr2","footnote_tail"]},block:{rules:["blockquote","code","fences","footnote","heading","hr","htmlblock","lheading","list","paragraph","table"]},inline:{rules:["autolink","backticks","del","emphasis","entity","escape","footnote_ref","htmltag","links","newline","text"]}}},Lr={options:{html:!1,xhtmlOut:!1,breaks:!1,langPrefix:"language-",linkTarget:"",typographer:!1,quotes:"\u201C\u201D\u2018\u2019",highlight:null,maxNesting:20},components:{core:{},block:{},inline:{}}},Hr={options:{html:!0,xhtmlOut:!0,breaks:!1,langPrefix:"language-",linkTarget:"",typographer:!1,quotes:"\u201C\u201D\u2018\u2019",highlight:null,maxNesting:20},components:{core:{rules:["block","inline","references","abbr2"]},block:{rules:["blockquote","code","fences","heading","hr","htmlblock","lheading","list","paragraph"]},inline:{rules:["autolink","backticks","emphasis","entity","escape","htmltag","links","newline","text"]}}},Br={default:qr,full:Lr,commonmark:Hr};function Ze(e,n,r){this.src=n,this.env=r,this.options=e.options,this.tokens=[],this.inlineMode=!1,this.inline=e.inline,this.block=e.block,this.renderer=e.renderer,this.typographer=e.typographer}function A(e,n){typeof e!="string"&&(n=e,e="default"),n&&n.linkify!=null&&console.warn(`linkify option is removed. Use linkify plugin instead:
-
-import Remarkable from 'remarkable';
-import linkify from 'remarkable/linkify';
-new Remarkable().use(linkify)
-`),this.inline=new G,this.block=new fe,this.core=new Ve,this.renderer=new ue,this.ruler=new y,this.options={},this.configure(Br[e]),this.set(n||{})}A.prototype.set=function(e){Pe(this.options,e)};A.prototype.configure=function(e){var n=this;if(!e)throw new Error("Wrong `remarkable` preset, check name/content");e.options&&n.set(e.options),e.components&&Object.keys(e.components).forEach(function(r){e.components[r].rules&&n[r].ruler.enable(e.components[r].rules,!0)})};A.prototype.use=function(e,n){return e(this,n),this};A.prototype.parse=function(e,n){var r=new Ze(this,e,n);return this.core.process(r),r.tokens};A.prototype.render=function(e,n){return n=n||{},this.renderer.render(this.parse(e,n),this.options,n)};A.prototype.parseInline=function(e,n){var r=new Ze(this,e,n);return r.inlineMode=!0,this.core.process(r),r.tokens};A.prototype.renderInline=function(e,n){return n=n||{},this.renderer.render(this.parseInline(e,n),this.options,n)};var X=null,H=null,J=!1;function We(){document.querySelectorAll("div#nav-bar__title, p.footer-container__help--links, a.nav-bar__search--results").forEach(r=>{Q(r)});let e=document.querySelector("#nav-bar__search--icon");e.addEventListener("click",F),e.targetParam="search";let n=document.querySelector("#nav-bar__games--icon");n.addEventListener("click",F),n.targetParam="games",be()}document.readyState!=="loading"?We():document.addEventListener("DOMContentLoaded",We);function be(){Vr(),Yr(),Dr(),Ae(),Te(),Se()}function Rr(){document.querySelectorAll(".content__selectorbox--item").forEach(e=>{e.style.setProperty("--highlight-color",e.dataset.highlight),e.addEventListener("click",n=>{document.querySelectorAll(".content__selectorbox--item").forEach(t=>{t.className="content__selectorbox--item"}),n.currentTarget.className="content__selectorbox--item selected";let r=C.filter(t=>t[0]==n.currentTarget.dataset.open);document.getElementById("content__currenth2").innerHTML=r[0][2],document.querySelectorAll(".content__collapse").forEach(t=>{t.addEventListener("click",we)}),be(),Ke()})}),document.querySelectorAll(".content__selectorbox--item")[0].click()}function Dr(){document.querySelectorAll(".tagging, .tagging-text").forEach(e=>{let n=e.dataset.tags,r=JSON.parse(n.replace(/'/g,'"'));if(r.todo){let t=document.createElement("span");t.className="material-symbols-rounded",t.style.color="goldenrod",t.style.marginRight="15px",t.style.cursor="help",t.textContent="error",e.appendChild(t),t.addEventListener("mouseover",()=>{j(t,"This section needs work, is not confirmed or needs testing.")}),t.addEventListener("mouseleave",()=>{$(t)})}if(r.versions){let t="Versions: ";t=t.concat(r.versions),t=t.concat(".");let o=document.createElement("span");o.className="material-symbols-rounded",o.style.marginRight="15px",o.style.cursor="help",o.textContent="devices",e.appendChild(o),o.addEventListener("mouseover",()=>{j(o,t)}),o.addEventListener("mouseleave",()=>{$(o)})}if(r.media)if(r.media.includes(".webp")||r.media.includes(".png")||r.media.includes(".jpg"))if(r.forcedmedia!=!1){let t=document.createElement("p");t.style.opacity="1";let o=document.createElement("img");if(o.draggable=!1,o.width="640",o.height="480",o.style.order=2,o.style.width="80%",o.style.height="auto",o.style.maxWidth="640px",o.style.outline="none",o.src=r.media,t.appendChild(o),r.caption){let l=document.createElement("figcaption");l.textContent="\xA0"+r.caption+"\xA0",t.appendChild(l)}e.parentNode.insertBefore(t,e.nextSibling)}else{let t=document.createElement("span");t.dataset.media=r.media,t.className="material-symbols-rounded",t.style.marginRight="15px",t.style.cursor="pointer",t.textContent="imagesmode",e.appendChild(t),t.addEventListener("click",o=>{let l=document.querySelector('p[data-media~="'+o.target.dataset.media+'"]');if(l==null){let c=document.createElement("p");c.style.opacity="1",c.dataset.media=o.target.dataset.media;let i=document.createElement("img");if(i.draggable=!1,i.width="640",i.height="480",i.style.order=2,i.style.width="80%",i.style.height="auto",i.style.maxWidth="640px",i.style.outline="none",i.src=o.target.dataset.media,c.appendChild(i),r.caption){let a=document.createElement("figcaption");a.textContent="\xA0"+r.caption+"\xA0",c.appendChild(a)}e.parentNode.insertBefore(c,e.nextSibling)}else l.remove()})}else if(r.forcedmedia!=!1){let t=document.createElement("p");t.style.opacity="1";let o=document.createElement("video");o.width="640",o.height="480",o.preload="metadata",o.style.order=2,o.style.width="80%",o.style.height="auto",o.style.maxWidth="640px",o.style.outline="none",o.controls=!0,o.muted=!0,o.loop=!0;let l=document.createElement("source");if(l.src=r.media,l.type="video/mp4",o.appendChild(l),t.appendChild(o),r.caption){let c=document.createElement("figcaption");c.textContent="\xA0"+r.caption+"\xA0",t.appendChild(c)}e.parentNode.insertBefore(t,e.nextSibling)}else{let t=document.createElement("span");t.dataset.media=r.media,t.className="material-symbols-rounded",t.style.marginRight="15px",t.style.cursor="pointer",t.textContent="play_circle",e.appendChild(t),t.addEventListener("click",o=>{let l=document.querySelector('p[data-media~="'+o.target.dataset.media+'"]');if(l==null){let c=document.createElement("p");c.style.opacity="1",c.dataset.media=o.target.dataset.media;let i=document.createElement("video");i.width="640",i.height="480",i.preload="metadata",i.style.order=2,i.style.width="80%",i.style.height="auto",i.style.maxWidth="640px",i.style.outline="none",i.controls=!0,i.muted=!0,i.loop=!0;let a=document.createElement("source");if(a.src=o.target.dataset.media,a.type="video/mp4",i.appendChild(a),c.appendChild(i),r.caption){let s=document.createElement("figcaption");s.textContent="\xA0"+r.caption+"\xA0",c.appendChild(s)}e.parentNode.insertBefore(c,e.nextSibling)}else l.remove()})}})}function de(){let e=document.getElementById("content__tocicon");e?.remove();let n=document.getElementById("content__tocicon--box");n?.remove()}function Ge(e){let n,r,t;for(r=document.getElementsByClassName("nav-bar__tab-bar--content"),n=0;n<r.length;n++)r[n].style.display="none";for(t=document.getElementsByClassName("nav-bar__tab-bar--links"),n=0;n<t.length;n++)t[n].className=t[n].className.replace(" active","");if(!(e.currentTarget.dataset.document=="HOME"||e.currentTarget.dataset.document.includes("./")))for(n=0;n<t.length;n++)t[n].dataset.document==e.currentTarget.dataset.document&&(t[n].className+=" active");let o=document.getElementById("section-container__text"),l=document.getElementById("content");if(e.currentTarget.dataset.document=="HOME")_=[],X="HOME",H="HOME",de(),l.innerHTML="UNDER CONSTRUCTION",o.innerHTML="HOME",J=!1,document.querySelector("#nav-bar").scrollIntoView({behavior:"smooth"});else if(e.currentTarget.dataset.document.includes("./"))_=[],de(),J=!1,X=e.currentTarget.dataset.section,H=e.currentTarget.dataset.document,o.innerHTML=e.currentTarget.dataset.section,ve(e.currentTarget.dataset.document).then(c=>{l.innerHTML=c,be(),q(),document.querySelector("#nav-bar").scrollIntoView({behavior:"smooth"})});else{J=!0,o.innerHTML=e.currentTarget.dataset.section;let c=H;X=e.currentTarget.dataset.section,H=e.currentTarget.dataset.document;let i=e.currentTarget.dataset;H==c?(q(),i.redirect!=null?(xe(i.redirect.substring(1)),document.querySelector(i.redirect).scrollIntoView({behavior:"smooth"})):document.querySelector("#nav-bar").scrollIntoView({behavior:"smooth"})):(_=[],de(),ve(("./tech/"+e.currentTarget.dataset.document).toLowerCase()).then(a=>{l.style.visibility="hidden",l.innerHTML=a,Me(l),Rr(),q(),i.redirect!=null?(xe(i.redirect.substring(1)),document.querySelector(i.redirect).scrollIntoView({behavior:"smooth"})):document.querySelector("#nav-bar").scrollIntoView({behavior:"smooth"})}))}}function Q(e){e.removeEventListener("click",Ge),e.addEventListener("click",Ge)}var T=new A("full",{html:!0,typographer:!1}),Or=T.renderer.rules.em_open,Pr=T.renderer.rules.em_close,Xe=T.renderer.rules.heading_close,me=!1,ge=!1,Fr={r:"red",b:"blue",g:"green",y:"yellow",p:"pink",t:"teal","!":"spoiler"};T.renderer.rules.heading_open=function(e,n,r,t){if(e[n+1].content[0]==":"){if(me=!0,e[n+1].content[1]=="{"){let o=Qe(e,n),l="<h"+e[n].hLevel+' id="'+e[n+1].content.replace(/\s+/g,"-").toLowerCase()+'">';return o+l}}else return"<h"+e[n].hLevel+' id="'+e[n+1].content.replace(/\s+/g,"-").toLowerCase()+'">'};T.renderer.rules.em_open=function(e,n,r,t){return zr(e,n,r,t,Or)};T.renderer.rules.heading_close=function(e,n,r,t){return me?(me=!1,Xe(e,n,r,t).concat("</span>").replace(/\n/g," ")+`
-`):Xe(e,n,r,t)};T.renderer.rules.em_close=function(e,n,r,t){return ge?(ge=!1,"</span>"):Pr(e,n,r,t)};function zr(e,n,r,t,o){if(e[n+1].type=="text")if(e[n+1].content[0]==":"){if(ge=!0,e[n+1].content[1]=="{")return Qe(e,n);{let l=e[n+1].content[1];return e[n+1].content=e[n+1].content.slice(2),'<span class="'+Fr[l]+'">'}}else return o(e,n,r,t);else return o(e,n,r,t)}function Qe(e,n){let r=e[n+1].content.indexOf("}"),t=e[n+1].content.substring(1,r+1);e[n+1].content=e[n+1].content.slice(r+2),e[n+1].type=="inline"&&(e[n+1].children[0].content=e[n+1].content);let o="tagging";return e[n].type=="em_open"&&(o="tagging-text"),'<span class="'+o+'" data-tags="'+t+'">'}function jr(e){return new Promise(function(n,r){let t=new XMLHttpRequest;t.open("GET",e,!0),t.onload=function(){this.status>=200&&this.status<300?n(t.response):r({status:this.status,statusText:t.statusText})},t.onerror=function(){r({status:this.status,statusText:t.statusText})},t.send()})}async function ve(e){let n=await jr(e+".md");if(n.length<=1||n==null)return"";let r="";return r=T.render(n),r}var P=!1,E=null;function F(){if(P)return;let e=event.currentTarget.targetParam;if(E=e,e=="search"?Je("#nav-bar__search"):Je("#nav-bar__games"),document.querySelector(".nav-bar__search--results")==null&&Ur(),e=="games"){let t=document.querySelector("#nav-bar__gamesbox").children;for(elem of t)elem.style.display="inline-block";document.querySelector("#nav-bar__games--hr2D").style.display="none",document.querySelector("#nav-bar__games--hr3D").style.display="none"}}function Je(e){if(P)return;document.querySelector(e).removeEventListener("click",F);var r=e+"box",t=e+"--input",o=e+"--icon",l=document.querySelector(r),c=document.querySelector(t);if(l==null){let s=document.createElement("div"),u=document.createElement("div");s.id="nav-bar__searchbox",u.id="nav-bar__gamesbox",s.style.width="0px",u.style.width="0px",u.tabIndex="0";let f=document.createElement("input");f.id="nav-bar__search--input",s.appendChild(f),document.querySelector("#nav-bar__search").appendChild(s),document.querySelector("#nav-bar__games").appendChild(u),f.addEventListener("keyup",$r),f.addEventListener("focusout",g=>{let p=document.querySelector("#nav-bar__search");g.stopPropagation(),p.contains(g.relatedTarget)||q()}),u.addEventListener("focusout",g=>{let p=document.querySelector("#nav-bar__games");g.stopPropagation(),p.contains(g.relatedTarget)||q()}),c=f}l=document.querySelector(r);var i=220;E=="search"?document.body.clientWidth>=480&&(i=360):i=300,l.style.width=i+"px";let a=document.querySelector(o);a.className="material-symbols-rounded logo-fadeout",a.removeEventListener("click",F),a.addEventListener("click",q),E=="search"?(c=document.querySelector(t),c.focus()):l.focus(),l.className="fadein"}function q(){if(E==null)return;var e,n,r=document.querySelector("#nav-bar__searchbox"),t=document.querySelector("#nav-bar__gamesbox");if(r==null||t==null)return;P=!0;var o=null;E=="search"?o=r:o=t;let l="#nav-bar__"+E;if(E=="search"){for(e=o.getElementsByTagName("a"),n=0;n<e.length;n++)e[n].style.display="none";document.querySelector(l+"--hr").style.display="none"}else{for(figs=o.getElementsByTagName("figure"),n=0;n<figs.length;n++)figs[n].style.display="none";document.querySelector(l+"--hr2D").style.display="none",document.querySelector(l+"--hr3D").style.display="none"}var c=document.querySelector(l+"--icon");c.className="material-symbols-rounded logo-fadein",c.removeEventListener("click",q),c.addEventListener("click",F),E=="search"?(c.targetParam="search",setTimeout(()=>{r.className="fadeout",P=!1},"300")):(c.targetParam="games",setTimeout(()=>{t.className="fadeout",P=!1},"300")),E=null}function $r(){var e,n,r,t,o="#nav-bar__search";if(E=="games")return;e=document.querySelector(o+"--input"),n=e.value.toUpperCase().trim(),filterArray=n.split(" "),r=document.querySelector(o+"box").getElementsByTagName("a");let c=0;for(t=0;t<r.length;t++){txtValue=r[t].textContent||r[t].innerText,txtValue=txtValue+" "+r[t].dataset.section+" "+r[t].dataset.document;let i=txtValue.toUpperCase(),a=!0;for(let s of filterArray.values())i.includes(s)&&n.length!==0||(a=!1);a?(c<6||E=="games"?r[t].style.display="block":r[t].style.display="none",c=c+1):r[t].style.display="none"}c>0?document.querySelector(o+"--hr").style.display="block":document.querySelector(o+"--hr").style.display="none"}function Ur(){var e=document.getElementsByClassName("game-instances--links"),n="",r="",t="",o=0,l=document.querySelector("#nav-bar__searchbox"),c=document.querySelector("#nav-bar__gamesbox");for(l.insertAdjacentHTML("beforeend","<hr id='nav-bar__search--hr'  tabindex='0' style='display: none;'>"),c.insertAdjacentHTML("beforeend",'<span style="cursor: default;height: 50px; display: block; float: right; padding-left: 30px;"></span>'),c.insertAdjacentHTML("beforeend","<hr id='nav-bar__games--hr2D'  tabindex='0' style='display: none;'>"),c.insertAdjacentHTML("beforeend","<hr id='nav-bar__games--hr3D'  tabindex='0' style='display: none;'>"),o=0;o<e.length;o++){let f=o;ve("./tech/"+e[o].dataset.document.toLowerCase()).then(d=>{let m=d,g=new DOMParser,p="";p=p.concat('<figure class="nav-bar__games--wrapper" data-document="'),p=p.concat(e[f].dataset.document),p=p.concat('" data-dim="'),p=p.concat(e[f].dataset.dim),p=p.concat('" tabindex="0" data-section="'),p=p.concat(e[f].dataset.section),p=p.concat('"><figcaption>'),p=p.concat(e[f].dataset.document),p=p.concat('</figcaption><img  class="nav-bar__search--results" src="media/'),p=p.concat(e[f].dataset.document.toLowerCase().trim()),p=p.concat("/"),p=p.concat(e[f].dataset.document.trim()),p=p.concat('ICONs.webp"></img></figure>'),e[f].dataset.dim=="2D"?document.querySelector("#nav-bar__games--hr2D").insertAdjacentHTML("beforebegin",p):document.querySelector("#nav-bar__games--hr3D").insertAdjacentHTML("beforebegin",p),p="",document.querySelectorAll("figure.nav-bar__games--wrapper").forEach(b=>{Q(b)});let x=g.parseFromString(m,"text/html").querySelectorAll("h1, h2, h3, h4"),S=null;x.forEach(b=>{b.tagName=="H1"?(S=null,n=n.concat('<a data-document="'),n=n.concat(e[f].dataset.document),n=n.concat('" class = "nav-bar__search--results" tabindex="0" data-section="'),n=n.concat(e[f].dataset.section),n=n.concat('" data-redirect="#'),n=n.concat(b.id),n=n.concat('"><b>'),n=n.concat(e[f].dataset.section),n=n.concat("</b> </a>")):(b.tagName=="H2"&&(S=b.id),(S=="glitches"||S=="techniques")&&b.id!="general-techniques"&&(n=n.concat('<a data-document="'),n=n.concat(e[f].dataset.document),n=n.concat('" class = "nav-bar__search--results" tabindex="0" data-section="'),n=n.concat(e[f].dataset.section),n=n.concat('" data-redirect="#'),n=n.concat(b.id),n=n.concat('"><b>'),n=n.concat(e[f].dataset.document),n=n.concat("</b> <br>&nbsp;&nbsp;&nbsp;"),n=n.concat(b.textContent),n=n.concat("</a>")))}),l.insertAdjacentHTML("beforeend",n),n="",document.querySelectorAll("a.nav-bar__search--results").forEach(b=>{Q(b)})})}}function xe(e){if(targetHeading=document.getElementById(e),targetHeading==null){h2Section=0;for(let t=0;t<C.length;t++)if(C[t][2].includes(`id="${e}"`)){h2Section=t;break}document.querySelectorAll(".content__selectorbox--item")[h2Section].click()}targetHeading=document.getElementById(e);let n=null,r=!1;targetHeading.tagName=="H4"?targetHeading.dataset.open!=null&&(n=targetHeading.dataset.open.split(" "),r=!0):targetHeading.parentNode.className.includes("content__")?targetHeading.parentNode.firstElementChild.dataset.open!=null&&(n=targetHeading.parentNode.firstElementChild.dataset.open.split(" "),r=!0):targetHeading.parentNode.parentNode.firstElementChild.dataset.open!=null&&(n=targetHeading.parentNode.parentNode.firstElementChild.dataset.open.split(" "),r=!0),r&&n.forEach(t=>{let o=document.getElementsByClassName(t);for(let l of o)l.hidden&&(l.hidden=!1,l.parentNode.className.includes("content__")?l.parentNode.firstElementChild.firstChild.innerHTML=="expand_more"?l.parentNode.firstElementChild.firstChild.innerHTML="expand_less":l.parentNode.firstElementChild.firstChild.innerHTML="remove":l.parentNode.parentNode.firstElementChild.firstChild.innerHTML=="expand_more"?l.parentNode.parentNode.firstElementChild.firstChild.innerHTML="expand_less":l.parentNode.parentNode.firstElementChild.firstChild.innerHTML="remove")})}var _=[];function Vr(){let e=document.getElementById("content__tocborder"),n=null,r=null,t=null,o=null;if(e==null){let s=document.querySelector(".content__selectorbox--item.selected");if(s==null)return;h2Title=s.textContent,r=document.getElementById("content"),e=document.createElement("div"),e.id="content__tocborder",r.appendChild(e),t=document.createElement("div"),t.id="content__toc",o=document.createElement("div"),o.id="content__tocmobile",e.appendChild(t),n=document.getElementById("content__selector"),e.style.top=n.offsetTop+150+"px";let u=document.createElement("span");u.className="material-symbols-rounded",u.id="content__tocicon",u.innerHTML="menu_book",u.style.display="none";let f=document.getElementById("section-container__div");f.appendChild(u);let d=document.createElement("div");d.id="content__tocicon--box",d.style.width="0px",d.tabIndex="0",f.appendChild(d),d.appendChild(o),U("#"+d.id,1),U("#"+t.id,1),u.addEventListener("click",m=>{let g=document.getElementById("content__tocicon--box");var p=300;document.body.clientWidth>=480&&(p=360),g.style.width=p+"px",g.className="fadein",d.display="block",g.focus()}),d.addEventListener("focusout",m=>{let g=document.getElementById("content__tocicon--box");m.stopPropagation(),g.className="fadeout"})}else n=document.getElementById("content__selector"),r=document.getElementById("content"),t=document.getElementById("content__toc"),o=document.getElementById("content__tocmobile"),h2Title=document.querySelector(".content__selectorbox--item.selected").textContent;e.style.height="calc(100% - "+(n.offsetTop+150)+"px)";let l="<p>\u2014 On this section:</p><hr>";t.innerHTML=l,o.innerHTML=l,_=[];let c=document.querySelectorAll("#content__currenth2 h3"),i=document.querySelectorAll("#content__currenth2 h4");for(let s of c)_.push({name:s.textContent,clientY:s.getBoundingClientRect().top,isH3:!0,obj:s,id:s.id});for(let s of i)_.push({name:s.textContent,clientY:s.getBoundingClientRect().top,isH3:!1,obj:s,id:s.id});_.sort((s,u)=>s.clientY-u.clientY);let a="";for(let s of _){let u='<a data-document="'+H+'" data-section="'+X+'" data-redirect="#'+s.id+'" class="content__toc--search" style="display: block;';s.isH3||(u=u.concat("padding-left: 30px; font-size: 13px; padding-top: 6px; padding-bottom: 6px;")),u=u.concat('">'+s.name),u=u.concat("</a>"),a=a.concat(u)}t.innerHTML=l.concat(a),o.innerHTML=l.concat(a),document.querySelectorAll(".content__toc--search").forEach(s=>{Q(s)})}var he=!1;function Ke(){let e=null,n=0,r=0;for(let t=0;t<_.length;t++)if(_[t].isH3)if(_[t].obj.getBoundingClientRect().top<=320||e==null)e=_[t],n=t;else{r=t;break}for(let t=0;t<_.length;t++)if(!_[t].isH3)if(t>=n&&t<=r||r==0&&n>0&&t>=n||r==0&&n==0){let o=document.querySelectorAll('.content__toc--search[data-redirect="#'+_[t].id+'"]');for(let l of o)l.style.display="block"}else{let o=document.querySelectorAll('.content__toc--search[data-redirect="#'+_[t].id+'"]');for(let l of o)l.style.display="none"}if(e!=null){let t=document.querySelectorAll(".content__toc--search.active");for(let l of t)l.classList.remove("active");let o=document.querySelectorAll('.content__toc--search[data-redirect="#'+e.id+'"]');for(let l of o)l.classList.add("active")}}document.addEventListener("scroll",e=>{!he&&J&&(window.requestAnimationFrame(()=>{Ke(),he=!1}),he=!0)});function Yr(){document.querySelectorAll('a[href^="#"]').forEach(e=>{e.addEventListener("click",function(n){n.preventDefault(),xe(this.getAttribute("href").substring(1)),document.querySelector(this.getAttribute("href")).scrollIntoView({behavior:"smooth"})})})}})();
-//# sourceMappingURL=page.js.map
