--- docs/scripts/scroll.js
@@ -1,2 +1,0 @@
-(()=>{var c=null,e=null;function s(){c=document.querySelector("#scroll__button--to-top"),e=document.documentElement,c.addEventListener("click",u),window.addEventListener("scroll",r)}document.readyState!=="loading"?s():document.addEventListener("DOMContentLoaded",s);function r(){let o=e.scrollHeight-e.clientHeight,t=document.getElementById("content__selectorbox"),l=0;if(t!=null){let d=document.body.getBoundingClientRect(),n=t.getBoundingClientRect(),i=n.top-d.top;i>0&&(l=i)}(e.scrollTop-l)/o>.05?c.classList.add("scroll__button--to-top--show"):c.classList.remove("scroll__button--to-top--show")}function u(){let o=document.getElementById("content__selectorbox"),t=0;if(o!=null){let l=document.body.getBoundingClientRect(),d=o.getBoundingClientRect(),n=d.top-l.top;n>0&&(t=n-100)}e.scrollTo({top:t,behavior:"smooth"})}})();
-//# sourceMappingURL=scroll.js.map
