--- docs/scripts/style.js
@@ -1,22 +1,0 @@
-(()=>{var T=[["yellow","#f8d959"],["pink","#fe796f"],["teal","#45c9c9"],["green","#58f15b"],["red","#e74a41"],["blue","#205aaa"],["lavender","#c8a2b0"]];function _(){var i=function(e,r){return e.children[r].innerText||e.children[r].textContent},s=function(e,r){return function(t,c){return function(d,n){return d!==""&&n!==""&&!isNaN(d)&&!isNaN(n)?d-n:d.toString().localeCompare(n)}(i(r?t:c,e),i(r?c:t,e))}};document.querySelectorAll("th").forEach(e=>{e.innerHTML=e.innerHTML+'<span class="material-symbols-rounded">swap_vert</span>',e.addEventListener("click",()=>{let t=e.closest("table").querySelector("tbody");Array.from(t.querySelectorAll("tr")).sort(s(Array.from(e.parentNode.children).indexOf(e),this.asc=!this.asc)).forEach(c=>t.appendChild(c))})})}function b(){document.querySelectorAll(".spoiler").forEach(i=>{i.addEventListener("click",s=>{s.target.style.background="transparent"})})}function E(){for(var i=document.getElementsByTagName("img"),s=document.getElementById("content"),e,e=0;e<i.length;e++)s.contains(i[e])&&i[e].classList.add("content__figure")}var h=!1,u=[],p,v;function H(i){u=[];let s=null,e=null,r=null,t="",c="",d=!0,n=i.innerHTML.split(`
-`),o=document.getElementById("content");for(let l=0;l<n.length;l++)if(n[l].includes('h4 id="')){s&&(t=t+`</div></div>
-`),s=n[l].substring(n[l].indexOf('h4 id="')+7),s=s.substring(0,s.indexOf('"')),t=t+`<div class="content__h4">
-`;let g=n[l].indexOf('h4 id="')+7+s.length+1,m=' data-open="'+s+" "+e+" "+r+'"',L=n[l].substr(0,g)+m+n[l].substr(g);t=t+L+`
-`,t=t+'<div class="'+s+`">
-`}else n[l].includes('h3 id="')?(s&&(t=t+`</div></div>
-`,s=null),e&&(t=t+`<hr class="content__h3--divider" draggable="false" /></div></div>
-`),e=n[l].substring(n[l].indexOf('h3 id="')+7),e=e.substring(0,e.indexOf('"')),t=t+`<div class="content__h3">
-`,t=t+'<button class="content__collapse" data-open="'+e+" "+r+'"><span class="material-symbols-rounded">expand_circle_up</span></button>',t=t+n[l]+`
-`,t=t+'<div class="'+e+`">
-`):n[l].includes('h2 id="')?(d=!1,s&&(t=t+`</div></div>
-`,s=null),e&&(t=t+`</div></div>
-`,e=null),r&&(t=t+`</div></div>
-`,u.push([r,currentH2Text,t]),t=""),r=n[l].substring(n[l].indexOf('h2 id="')+7),r=r.substring(0,r.indexOf('"')),currentH2Text=n[l].substring(n[l].indexOf('">')+2),currentH2Text=currentH2Text.substring(0,currentH2Text.indexOf("</h2>")),t=t+'<div class="content__h2" id="'+r+`">
-`,t=t+'<div class="'+r+`">
-`):d?c=c+n[l]+`
-`:t=t+n[l]+`
-`;s&&(t=t+`</div></div>
-`),e&&(t=t+`</div></div>
-`),r&&(t=t+`</div></div>
-`,u.push([r,currentH2Text,t]),t="");var a='<div id="content__selector"><div id="content__selectorbox">';let f=0;for(;f<u.length;)h2Tag=u[f][0],h2Text=u[f][1],h2Color=T[f%T.length],a=a+'<div class="content__selectorbox--item" data-open="'+h2Tag+'" data-highlight="'+h2Color[1]+'">',a=a+h2Text+"</div>",f++;a=a+'</div><hr id="content__selectorhr"></hr></div><div id="content__currenth2"></div>',u.length>0?i.innerHTML=c+a:i.innerHTML=c,x("#content__selectorbox",0),o.style.visibility="visible"}function x(i,s){let e=document.querySelector(i),r=function(o){h=!0,p=o.pageX-e.offsetLeft,isNaN(p)&&(p=o.changedTouches[0].pageX-e.offsetLeft),v=e.scrollLeft},t=function(o){h=!0,startY=o.pageY-e.offsetTop,isNaN(startY)&&(startY=o.changedTouches[0].pageY-e.offsetTop),scrollTop=e.scrollTop},c=function(o){if(o.preventDefault(),o.stopPropagation(),!h)return;let a=o.pageX-e.offsetLeft,f=a-p;if(isNaN(a)){let g=o.changedTouches[0].pageX-e.offsetLeft-p;e.scrollLeft=v-g}else e.scrollLeft=v-f},d=function(o){if(o.preventDefault(),o.stopPropagation(),!h)return;let a=o.pageY-e.offsetTop,f=a-startY;if(isNaN(a)){let g=o.changedTouches[0].pageY-e.offsetTop-startY;e.scrollTop=scrollTop-g}else e.scrollTop=scrollTop-f},n=function(o){h=!1};s==1?(e.addEventListener("mousemove",d,!1),e.addEventListener("mousedown",t,!1),e.addEventListener("mouseup",n,!1),e.addEventListener("mouseleave",n,!1),e.addEventListener("touchmove",d,!1),e.addEventListener("touchstart",t,!1),e.addEventListener("touchend",n,!1)):(e.addEventListener("mousemove",c,!1),e.addEventListener("mousedown",r,!1),e.addEventListener("mouseup",n,!1),e.addEventListener("mouseleave",n,!1),e.addEventListener("touchmove",c,!1),e.addEventListener("touchstart",r,!1),e.addEventListener("touchend",n,!1))}function y(i){let s=i.currentTarget.dataset.open.split(" ");for(let e=0;e<s.length;e++){targetList=document.getElementsByClassName(s[e]);for(let r of targetList)r.hidden?(r.hidden=!1,i.currentTarget.firstChild.innerHTML=="expand_circle_down"?i.currentTarget.firstChild.innerHTML="expand_circle_up":i.currentTarget.firstChild.innerHTML="remove"):e==0&&(r.hidden=!0,i.currentTarget.firstChild.innerHTML=="expand_circle_up"?i.currentTarget.firstChild.innerHTML="expand_circle_down":i.currentTarget.firstChild.innerHTML="add")}}})();
-//# sourceMappingURL=style.js.map
