--- docs/scripts/tooltip.js
@@ -1,2 +1,0 @@
-(()=>{function n(){document.querySelectorAll(".nav-bar__tab-bar--links").forEach(e=>{e.addEventListener("mouseover",()=>{i(e,e.dataset.section)}),e.addEventListener("mouseleave",()=>{a()})})}document.readyState!=="loading"?n():document.addEventListener("DOMContentLoaded",n);function i(e,l){let t=document.createElement("tooltip");t.id="nav-bar__tooltip";let d=document.createTextNode(l);if(e!=null){t.innerHTML="",t.appendChild(d),document.getElementsByTagName("tooltip")[0]&&document.getElementsByTagName("tooltip")[0].remove();let o=e.getBoundingClientRect();t.style.top=o.bottom-10+"px",t.style.left=o.right+"px",document.body.appendChild(t)}}function a(){document.getElementsByTagName("tooltip")[0]&&document.getElementsByTagName("tooltip")[0].remove()}})();
-//# sourceMappingURL=tooltip.js.map
