+++ docs/styles/base.css
@@ -1,1 +1,1 @@
-@font-face{font-family:Material Symbols Rounded;font-style:normal;font-weight:400;font-display:block;src:url(../fonts/gsymbols.woff2) format("woff2")}@font-face{font-family:Montserrat;font-style:normal;font-display:swap;src:url(../fonts/gmontserrat.woff2) format("woff2")}@font-face{font-family:Nunito;font-style:normal;font-display:swap;src:url(../fonts/gnunito.woff2) format("woff2")}.material-symbols-rounded{font-family:Material Symbols Rounded;font-weight:400;font-style:normal;font-size:24px;line-height:1;letter-spacing:normal;text-transform:none;display:inline-block;white-space:nowrap;word-wrap:normal;direction:ltr;-webkit-font-feature-settings:"liga";-moz-font-feature-settings:"liga";font-feature-settings:"liga"}.material-symbols-rounded{border:none!important;border-width:0px;border-color:#161618;color:#fff;opacity:80%;z-index:8;font-family:Material Symbols Rounded!important}*{-moz-osx-font-smoothing:grayscale;-webkit-font-smoothing:antialiased!important;-moz-font-smoothing:antialiased!important;font-kerning:auto}html{overflow-y:scroll;color:#fff}::-webkit-scrollbar{width:9px}::-webkit-scrollbar-track{background:transparent}::-webkit-scrollbar-thumb{background-color:#252529;border-radius:10px;border:transparent}html,body{padding:0;margin:0;background-color:#161618;scrollbar-color:#252529 #161618;scrollbar-width:thin;max-width:100%;overflow-x:clip}video,p img{display:block;margin:25px auto 14px;-webkit-user-drag:none;user-drag:none;-moz-user-select:none;-webkit-user-select:none;-ms-user-select:none;user-select:none}figcaption{font-family:Nunito,sans-serif;text-align:center;font-size:15px;text-decoration:underline;text-underline-offset:6px}hr{border:none;border-top:1px double #ffffff;color:#fff;overflow:visible;text-align:center}h1{font-family:Nunito,sans-serif;font-weight:800;color:#fff;font-weight:300;font-size:40px;letter-spacing:.1em;text-shadow:1px 1px 1px #ffffff;opacity:95%;scroll-margin-top:100px;padding-bottom:.1em;border-bottom:1px solid #252529}h2{font-family:Nunito,sans-serif;color:#fff;font-weight:600;font-size:35px;letter-spacing:.1em;opacity:85%;scroll-margin-top:100px;padding-bottom:.1em;border-bottom:1px solid #252529;margin-right:15px;width:-moz-fit-content;width:fit-content;margin-top:0%;margin-bottom:0%}h3,h4,h5,h6{font-family:Nunito,sans-serif;color:#fff;font-weight:600;font-size:25px;opacity:85%;margin-right:15px;scroll-margin-top:100px;width:-moz-fit-content;width:fit-content;margin-top:0%;margin-bottom:0%}h4,h5,h6{text-decoration:underline}h3{text-indent:6ch;font-size:28px}.content__h3 * h3~.material-symbols-rounded{font-size:30px;padding-top:2px}h4{text-decoration-thickness:2px;font-size:26px}.content__h4 * h4~.material-symbols-rounded{font-size:28px;padding-top:2px}p{font-family:Montserrat,sans-serif;font-weight:400;font-size:18px;max-width:90ch;text-indent:2ch;letter-spacing:.03em;color:#fffc}a:link,a:visited{color:#fffc}a:hover{background:#252529;-webkit-filter:brightness(1.25);filter:brightness(1.25)}a:active{color:#fffc}.spoiler{color:#fff;background-color:#fff;opacity:85%;cursor:pointer;transition-timing-function:ease-out;-webkit-transition:background .35s;-moz-transition:background .35s;-ms-transition:background .35s;-o-transition:background .35s;transition:background .35s}#game-instances{display:none;width:0;height:0}.green{color:#58f15b}.red{color:#e74a41}.blue{color:#205aaa}.yellow{color:#f8d959}.pink{color:#fe796f}.teal{color:#45c9c9}
+@font-face{font-family:Material Symbols Rounded;font-style:normal;font-weight:400;font-display:block;src:url(../fonts/gsymbols.woff2) format("woff2")}@font-face{font-family:Montserrat;font-style:normal;font-display:swap;src:url(../fonts/gmontserrat.woff2) format("woff2")}@font-face{font-family:Nunito;font-style:normal;font-display:swap;src:url(../fonts/gnunito.woff2) format("woff2")}.material-symbols-rounded{font-family:Material Symbols Rounded;font-weight:400;font-style:normal;font-size:24px;line-height:1;letter-spacing:normal;text-transform:none;display:inline-block;white-space:nowrap;word-wrap:normal;direction:ltr;-webkit-font-feature-settings:"liga";-moz-font-feature-settings:"liga";font-feature-settings:"liga"}.material-symbols-rounded{border:none!important;border-width:0px;border-color:#161618;color:#fff;opacity:80%;z-index:8;font-family:Material Symbols Rounded!important}*{-moz-osx-font-smoothing:grayscale;-webkit-font-smoothing:antialiased!important;-moz-font-smoothing:antialiased!important;font-kerning:auto}html{overflow-y:scroll;color:#fff}::-webkit-scrollbar{width:9px}::-webkit-scrollbar-track{background:transparent}::-webkit-scrollbar-thumb{background-color:#252529;border-radius:10px;border:transparent}html,body{padding:0;margin:0;background-color:#161618;scrollbar-color:#252529 #161618;scrollbar-width:thin;max-width:100%;overflow-x:clip}video,p img{display:block;margin:25px auto 14px;-webkit-user-drag:none;user-drag:none;-moz-user-select:none;-webkit-user-select:none;-ms-user-select:none;user-select:none}figcaption{font-family:Nunito,sans-serif;text-align:center;font-size:15px;text-decoration:underline;text-underline-offset:6px}hr{border:none;border-top:1px double #ffffff;color:#fff;overflow:visible;text-align:center}h1{font-family:Nunito,sans-serif;font-weight:800;color:#fff;font-weight:300;font-size:40px;letter-spacing:.1em;text-shadow:1px 1px 1px #ffffff;opacity:95%;scroll-margin-top:100px;padding-bottom:.1em;border-bottom:1px solid #252529}h2{font-family:Nunito,sans-serif;color:#fff;font-weight:600;font-size:35px;letter-spacing:.1em;opacity:85%;scroll-margin-top:100px;padding-bottom:.1em;border-bottom:1px solid #252529;margin-right:15px;width:-moz-fit-content;width:fit-content;margin-top:0%;margin-bottom:0%}h3,h4,h5,h6{font-family:Nunito,sans-serif;color:#fff;font-weight:600;font-size:25px;opacity:85%;margin-right:15px;scroll-margin-top:100px;width:-moz-fit-content;width:fit-content;margin-top:0%;margin-bottom:0%}h4,h5,h6{text-decoration:underline}h3{text-indent:6ch;font-size:28px}.content__h3 * h3~.material-symbols-rounded{font-size:30px;padding-top:2px}h4{text-decoration-thickness:2px;font-size:26px}.content__h4 * h4~.material-symbols-rounded{font-size:28px;padding-top:2px}p{font-family:Montserrat,sans-serif;font-weight:400;font-size:18px;max-width:90ch;text-indent:2ch;letter-spacing:.03em;color:#fffc}a:link,a:visited{color:#fffc}a:hover{background:#252529;-webkit-filter:brightness(1.25);filter:brightness(1.25)}a:active,.button__redirect:link,.button__redirect:visited{color:#fffc}.button__redirect:hover{background:#252529;-webkit-filter:brightness(1.25);filter:brightness(1.25)}.button__redirect:active{color:#fffc}.spoiler{color:#fff;background-color:#fff;opacity:85%;cursor:pointer;transition-timing-function:ease-out;-webkit-transition:background .35s;-moz-transition:background .35s;-ms-transition:background .35s;-o-transition:background .35s;transition:background .35s}.green{color:#58f15b}.red{color:#e74a41}.blue{color:#205aaa}.yellow{color:#f8d959}.pink{color:#fe796f}.teal{color:#45c9c9}
