+++ docs/tech/tota.md
@@ -1,3 +1,5 @@
 # Tales of the Abyss
 
 [TODO]
+
+## Mechanics
