+++ docs/tech/toz.md
@@ -1,3 +1,5 @@
 # Tales of Zestiria
 
 [TODO]
+
+## Mechanics
