+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "tte",
-  "version": "0.15.6",
+  "version": "0.15.8",
   "description": "Tales Tech Encyclopedia",
   "main": "src/documentParser.js",
   "private": true,
@@ -13,7 +13,7 @@
   },
   "scripts": {
     "watch": "nodemon --watch \"src/\" --exec \"npm run build\"",
-    "build": "node esbuild.mjs && minify-html --output docs/index.html --keep-closing-tags --minify-css src/index.html"
+    "build": "rimraf docs/scripts docs/styles && node esbuild.mjs && minify-html --output docs/index.html --keep-closing-tags --minify-css src/index.html"
   },
   "dependencies": {
     "@minify-html/node": "^0.15.0",
@@ -23,6 +23,7 @@
     "katex": "^0.16.11",
     "nodemon": "^3.1.4",
     "remarkable": "^2.0.1",
-    "remarkable-katex": "^1.2.1"
+    "remarkable-katex": "^1.2.1",
+    "rimraf": "^6.0.1"
   }
 }
