+++ src/styles/content.css
@@ -1,4 +1,4 @@
-/* CSS for content items. */
+/* CSS for items within the content div. */
 
 #content {
   position: relative;
